A TLS scan

A data.table from a TLS scan.

data(TLS_scan)

Format

A data.table with five columns, which are:

X

the "X" coordinate

Y

the "Y coordinate

Z

the "Z" coordinate

Target_count

The number of received by the same laser shot

Target_index

The rank of the returned pulse in the target count of received by the same laser shot

A data.table where the rows represent the pulse returns and the three columns represent the *XYZ* coordinates, and the target count and index.

Details

A TLS scan conducted using a Reigel VZ400i with a vertical and horizontal resolution of 0.048 and 0.622 degrees (2082 and 580 lines, respectively). The scanner has frame of scanning between 30 and 130.024 degrees zenith and 0 and 359.90 degrees azimuth. At the moment of the scan the roll, pitch, and yaw of the scanner were 1.026, 0.746, -110.019, respectively. The scanner coordinates in this scan are x = 0, y = 0, z = 0.

Examples

data(TLS_scan)
head(TLS_scan)

#>         X      Y      Z Target_count Target_index
#> 1: -3.392 12.330 -1.200            1            1
#> 2: -3.370 12.248 -1.202            1            1
#> 3: -3.339 12.136 -1.202            1            1
#> 4: -3.722 11.989 -1.169            2            2
#> 5: -3.595 12.039 -1.168            1            1
#> 6: -3.424 11.943 -1.171            1            1