Dimensionality of the Neighboring Points.

Estimate the dimensionality of neighboring points in a cloud.

dimensionality(
  cloud,
  cloud_b = NULL,
  method,
  radius,
  k,
  threads = 1,
  progress = TRUE
)

Arguments

cloud
A data.table with *XYZ* coordinates in the first three columns.

cloud_b
A data.table with *XYZ* coordinates in the first three columns to estimate the neighboring points. If cloud_b is NULL, cloud_b == cloud. NULL as default.

method
A character string specifying the method to estimate the neighbors. It most be one of "sphere" or "knn".

radius
A numeric vector representing the radius of the sphere or spheres to consider. This needs be used if method = "sphere".

k
An integer vector representing the number of neighbors to consider. This needs be used if method = "knn".

threads
An integer specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware.

progress
Show progress. TRUE as default.

Value

A array describing the point of the cloud in rows, the relative eigenvalues in columns, and the radius or k per slide. If method = "sphere", it add in the first column the number of neighbor points.

Details

The function returns the dimensionality of the neighboring points of a given point in code. The dimensionality is represented by the relative values of the eigenvalues derived from a covariance matrix of the neighboring points.

Examples

###Estimate the dimensionality on a sample of 100 points.
#Load data
data("pc_tree")

#Sample data
sample_data <- pc_tree[sample(nrow(pc_tree), 100), ]

#Using neighbors in two spheres of 0.75 and 1
dimensionality(sample_data, pc_tree, method = "sphere", radius = c(0.75, 1))

#> , , radius_0.75
#> 
#>     npoints      eig1      eig2       eig3
#> 1      1783 0.4137448 0.3918825 0.19437269
#> 2      1706 0.3821855 0.3236556 0.29415888
#> 3      2178 0.4840066 0.3104429 0.20555050
#> 4      1858 0.4228015 0.3202460 0.25695247
#> 5      1548 0.4312849 0.3932374 0.17547763
#> 6      2086 0.3699196 0.3314379 0.29864247
#> 7      1870 0.5021278 0.3012456 0.19662662
#> 8      1702 0.5220640 0.3730560 0.10488008
#> 9      1822 0.4183173 0.3251081 0.25657465
#> 10     1394 0.7014396 0.1780345 0.12052591
#> 11     2099 0.4064630 0.3617805 0.23175654
#> 12     2522 0.3994533 0.3367183 0.26382845
#> 13     1953 0.4753233 0.3358715 0.18880518
#> 14      693 0.5930793 0.2434950 0.16342568
#> 15     1043 0.5987329 0.2730370 0.12823012
#> 16     1460 0.5009834 0.3495389 0.14947768
#> 17      965 0.6184630 0.2619207 0.11961621
#> 18      888 0.4189905 0.3200607 0.26094880
#> 19     1943 0.4242559 0.3337638 0.24198030
#> 20     1356 0.4649492 0.3320595 0.20299132
#> 21     1589 0.5238840 0.3369601 0.13915590
#> 22     1721 0.4596026 0.3387698 0.20162761
#> 23     1707 0.4572494 0.3001458 0.24260471
#> 24     2296 0.5500610 0.2879013 0.16203771
#> 25     1724 0.4591769 0.3008549 0.23996827
#> 26      954 0.5019644 0.3462309 0.15180469
#> 27     1766 0.4188614 0.3231966 0.25794197
#> 28     2192 0.4218528 0.3289642 0.24918303
#> 29     1214 0.4968918 0.3110855 0.19202275
#> 30     1697 0.4048173 0.3276517 0.26753107
#> 31     1203 0.4872106 0.3685709 0.14421852
#> 32     1214 0.6279371 0.1990590 0.17300395
#> 33     2247 0.5106855 0.2922286 0.19708588
#> 34     2124 0.4117465 0.3145360 0.27371748
#> 35     1621 0.4647642 0.3105814 0.22465440
#> 36     1497 0.4123949 0.3073524 0.28025275
#> 37     1265 0.5708725 0.2697276 0.15939993
#> 38     2554 0.3816930 0.3367309 0.28157615
#> 39     1594 0.5503930 0.2787581 0.17084894
#> 40     2352 0.3961890 0.3625913 0.24121972
#> 41     1465 0.4293238 0.3396145 0.23106171
#> 42     2142 0.4496978 0.3196008 0.23070138
#> 43     1873 0.4199197 0.3613172 0.21876310
#> 44      660 0.7867813 0.1273825 0.08583613
#> 45     1841 0.4551700 0.3122392 0.23259081
#> 46     1431 0.4223015 0.3586991 0.21899948
#> 47     1417 0.5669316 0.2363614 0.19670700
#> 48     1638 0.5003080 0.2876043 0.21208765
#> 49     1926 0.4245131 0.3154518 0.26003513
#> 50     1963 0.5066446 0.3056277 0.18772769
#> 51      884 0.6717251 0.2346100 0.09366492
#> 52     1629 0.4112755 0.3617370 0.22698756
#> 53      494 0.5805582 0.3222988 0.09714304
#> 54     1685 0.5114444 0.3130228 0.17553279
#> 55     1341 0.6102185 0.2526071 0.13717438
#> 56     1455 0.5622952 0.2534496 0.18425513
#> 57     1945 0.3948805 0.3411095 0.26401004
#> 58     2016 0.3706209 0.3486597 0.28071937
#> 59     2364 0.4211394 0.3208949 0.25796572
#> 60     1870 0.4307045 0.3339143 0.23538119
#> 61     1383 0.4219239 0.3360407 0.24203545
#> 62     1360 0.4612858 0.3811553 0.15755893
#> 63     2471 0.4905652 0.2956432 0.21379159
#> 64     1115 0.6137788 0.2309304 0.15529082
#> 65     1812 0.4347982 0.3229668 0.24223500
#> 66     2388 0.4134588 0.3248961 0.26164503
#> 67     1653 0.5094194 0.2562402 0.23434038
#> 68     1634 0.4490058 0.3257588 0.22523537
#> 69     2095 0.4354230 0.3022564 0.26232069
#> 70     1562 0.4836632 0.3166218 0.19971505
#> 71     2332 0.4329334 0.3304359 0.23663077
#> 72     1988 0.4063563 0.3429324 0.25071129
#> 73     2073 0.4548562 0.2946859 0.25045790
#> 74     2222 0.4240189 0.2957668 0.28021426
#> 75     1905 0.4310061 0.3659733 0.20302057
#> 76     1839 0.4583553 0.3054275 0.23621718
#> 77     1498 0.4847684 0.3027682 0.21246335
#> 78     2158 0.4167635 0.3293630 0.25387349
#> 79     1785 0.4214289 0.3251654 0.25340574
#> 80     2061 0.4400436 0.3750484 0.18490799
#> 81     2038 0.4440340 0.3614537 0.19451226
#> 82     1826 0.4927105 0.3020444 0.20524509
#> 83     2319 0.4042466 0.3291949 0.26655850
#> 84     1606 0.3978242 0.3519531 0.25022274
#> 85     1603 0.3998311 0.3525174 0.24765154
#> 86     1992 0.4350355 0.3242688 0.24069573
#> 87     1849 0.4690357 0.2993497 0.23161455
#> 88     1619 0.3697104 0.3202869 0.31000264
#> 89     1214 0.5112675 0.2947602 0.19397229
#> 90     2374 0.4734988 0.3118493 0.21465183
#> 91     1425 0.6212005 0.2530414 0.12575806
#> 92     2705 0.4444832 0.3375104 0.21800639
#> 93     2000 0.3987296 0.3469611 0.25430931
#> 94     2087 0.4521543 0.3364089 0.21143677
#> 95     2523 0.4005588 0.3168623 0.28257886
#> 96     2468 0.4453893 0.3116582 0.24295249
#> 97     2438 0.4126892 0.3381834 0.24912736
#> 98      709 0.5153733 0.2652329 0.21939385
#> 99     2659 0.4373008 0.3104221 0.25227710
#> 100    2071 0.4960818 0.3135281 0.19039009
#> 
#> , , radius_1
#> 
#>     npoints      eig1      eig2      eig3
#> 1      3682 0.4853512 0.3693391 0.1453096
#> 2      4075 0.3634327 0.3361133 0.3004539
#> 3      3809 0.4630973 0.3000658 0.2368369
#> 4      3982 0.4082838 0.3365149 0.2552013
#> 5      3976 0.4635558 0.3737246 0.1627196
#> 6      4812 0.3629145 0.3419489 0.2951366
#> 7      4241 0.5011149 0.2938191 0.2050660
#> 8      4376 0.4453652 0.4030431 0.1515917
#> 9      3786 0.3823747 0.3405554 0.2770699
#> 10     2628 0.6775672 0.1967634 0.1256693
#> 11     3672 0.4652449 0.2970527 0.2377025
#> 12     6149 0.3844231 0.3613670 0.2542099
#> 13     3546 0.4573504 0.3503041 0.1923455
#> 14     1517 0.5248131 0.3324986 0.1426884
#> 15     2297 0.4897201 0.3468684 0.1634115
#> 16     2207 0.4745114 0.3744524 0.1510362
#> 17     1527 0.6620536 0.2352381 0.1027083
#> 18     2180 0.5233329 0.3222487 0.1544185
#> 19     3707 0.3909364 0.3824657 0.2265979
#> 20     3619 0.4505662 0.3240115 0.2254224
#> 21     2727 0.5334869 0.3362384 0.1302747
#> 22     3622 0.4941936 0.3237175 0.1820889
#> 23     3488 0.4767151 0.3328820 0.1904030
#> 24     4613 0.4809486 0.3201317 0.1989197
#> 25     3975 0.3537089 0.3272881 0.3190030
#> 26     2374 0.5457852 0.2850823 0.1691325
#> 27     4043 0.3619189 0.3532054 0.2848758
#> 28     4834 0.3977249 0.3474953 0.2547798
#> 29     2539 0.4810589 0.3347687 0.1841725
#> 30     4406 0.3783815 0.3284718 0.2931466
#> 31     3212 0.4426547 0.3663127 0.1910326
#> 32     2826 0.5628192 0.2313685 0.2058123
#> 33     4192 0.4436914 0.3170829 0.2392258
#> 34     5147 0.3569063 0.3338297 0.3092640
#> 35     3183 0.4372892 0.3420425 0.2206682
#> 36     3854 0.4324585 0.3194519 0.2480896
#> 37     2706 0.4437227 0.4020797 0.1541975
#> 38     4883 0.4237814 0.3092689 0.2669497
#> 39     3042 0.5224333 0.2981687 0.1793980
#> 40     4174 0.4663676 0.3219457 0.2116867
#> 41     3444 0.4085919 0.3555491 0.2358591
#> 42     3904 0.4406604 0.3164893 0.2428503
#> 43     3604 0.5101453 0.3192895 0.1705652
#> 44     1754 0.6432478 0.1960598 0.1606924
#> 45     3498 0.4605294 0.3190914 0.2203792
#> 46     3720 0.3846478 0.3454998 0.2698523
#> 47     2697 0.5285420 0.3068132 0.1646448
#> 48     3332 0.4582491 0.2720683 0.2696826
#> 49     3369 0.4736420 0.2887103 0.2376477
#> 50     3888 0.4488669 0.3654509 0.1856823
#> 51     1759 0.6845437 0.1998278 0.1156285
#> 52     3203 0.4226299 0.3517075 0.2256626
#> 53     1051 0.4721016 0.3070513 0.2208471
#> 54     3908 0.4656957 0.3219012 0.2124031
#> 55     2335 0.4745368 0.2979552 0.2275080
#> 56     2411 0.5065638 0.2900982 0.2033380
#> 57     4635 0.3985244 0.3375752 0.2639004
#> 58     4856 0.3609009 0.3413281 0.2977710
#> 59     5144 0.4096434 0.3320832 0.2582735
#> 60     3895 0.4961637 0.2849737 0.2188626
#> 61     3735 0.4577137 0.2965486 0.2457377
#> 62     4097 0.4938912 0.3407839 0.1653249
#> 63     4304 0.4668011 0.3036640 0.2295349
#> 64     2202 0.5546149 0.2832884 0.1620967
#> 65     4637 0.4092062 0.3163760 0.2744178
#> 66     5083 0.3697269 0.3576944 0.2725787
#> 67     3172 0.5360124 0.2823106 0.1816770
#> 68     3315 0.4528485 0.3273947 0.2197568
#> 69     4475 0.3929267 0.3620533 0.2450200
#> 70     4104 0.4534348 0.3793205 0.1672447
#> 71     5361 0.4198889 0.3329189 0.2471922
#> 72     4499 0.3919216 0.3659584 0.2421200
#> 73     4969 0.3901698 0.3215904 0.2882398
#> 74     5181 0.4330422 0.3179391 0.2490187
#> 75     4577 0.4358302 0.3678812 0.1962886
#> 76     4114 0.4216079 0.3506172 0.2277749
#> 77     3284 0.4712220 0.3277134 0.2010645
#> 78     4408 0.5046513 0.2759200 0.2194287
#> 79     3429 0.4292803 0.3209809 0.2497388
#> 80     4153 0.4818398 0.3126460 0.2055143
#> 81     3931 0.4688875 0.3489635 0.1821490
#> 82     4188 0.4442919 0.3322430 0.2234651
#> 83     4504 0.3990022 0.3337789 0.2672189
#> 84     3828 0.4337091 0.3487123 0.2175786
#> 85     2731 0.4957256 0.2690434 0.2352310
#> 86     3732 0.4114486 0.3492291 0.2393223
#> 87     3898 0.4977900 0.2878559 0.2143541
#> 88     4438 0.3743628 0.3215193 0.3041179
#> 89     2640 0.5102722 0.3129759 0.1767519
#> 90     4866 0.4401723 0.3010270 0.2588008
#> 91     2481 0.4909229 0.3320839 0.1769932
#> 92     5607 0.4720667 0.2707081 0.2572252
#> 93     5132 0.3632812 0.3416081 0.2951107
#> 94     4108 0.4298048 0.3700405 0.2001547
#> 95     5002 0.3808359 0.3329827 0.2861813
#> 96     5602 0.3874196 0.3344401 0.2781403
#> 97     5269 0.3928179 0.3553121 0.2518700
#> 98     1448 0.4409660 0.3341397 0.2248943
#> 99     5640 0.4189652 0.3501275 0.2309073
#> 100    3326 0.5306092 0.2862474 0.1831435
#> 

#Using two k neighbors of 50 and 100
dimensionality(cloud = sample_data, cloud_b = pc_tree, method = "knn", k = c(50, 100))

#> , , k_50
#> 
#>          eig1       eig2       eig3
#> 1   0.8740180 0.11256242 0.01341961
#> 2   0.7983261 0.13692326 0.06475067
#> 3   0.7350975 0.25070472 0.01419779
#> 4   0.6958831 0.26848791 0.03562896
#> 5   0.5615919 0.29579853 0.14260953
#> 6   0.9112984 0.05667065 0.03203093
#> 7   0.7053318 0.26429339 0.03037485
#> 8   0.4715445 0.35299902 0.17545646
#> 9   0.6779678 0.26705329 0.05497895
#> 10  0.7660054 0.16616764 0.06782694
#> 11  0.7254883 0.16655122 0.10796049
#> 12  0.8750871 0.11181656 0.01309639
#> 13  0.6342231 0.34844168 0.01733518
#> 14  0.8714774 0.08918127 0.03934129
#> 15  0.8275624 0.10765692 0.06478069
#> 16  0.8484588 0.13027946 0.02126173
#> 17  0.6923000 0.25239392 0.05530606
#> 18  0.7378127 0.19637357 0.06581371
#> 19  0.6556888 0.31182092 0.03249024
#> 20  0.6302362 0.25767737 0.11208648
#> 21  0.8034144 0.17335209 0.02323350
#> 22  0.6796124 0.25449544 0.06589220
#> 23  0.8937871 0.07344398 0.03276894
#> 24  0.7666952 0.18332348 0.04998136
#> 25  0.8854354 0.06074679 0.05381785
#> 26  0.7615646 0.22246092 0.01597444
#> 27  0.5757790 0.36236036 0.06186063
#> 28  0.6419157 0.18784685 0.17023748
#> 29  0.8359818 0.12373154 0.04028669
#> 30  0.9033243 0.08631506 0.01036064
#> 31  0.5729842 0.36708414 0.05993170
#> 32  0.8581473 0.12581764 0.01603504
#> 33  0.9192856 0.06598126 0.01473313
#> 34  0.7381911 0.20871979 0.05308911
#> 35  0.6424890 0.34292622 0.01458476
#> 36  0.8113944 0.14241451 0.04619109
#> 37  0.7856111 0.19060223 0.02378663
#> 38  0.6640501 0.28557252 0.05037739
#> 39  0.6758998 0.17650677 0.14759342
#> 40  0.6834291 0.24598705 0.07058386
#> 41  0.6503851 0.27561311 0.07400179
#> 42  0.7375843 0.24819688 0.01421886
#> 43  0.8626556 0.09203315 0.04531123
#> 44  0.6333177 0.34783543 0.01884689
#> 45  0.8133471 0.13788189 0.04877101
#> 46  0.7011056 0.24935174 0.04954264
#> 47  0.6367911 0.32092686 0.04228201
#> 48  0.6082123 0.34563921 0.04614852
#> 49  0.7914162 0.12897376 0.07961002
#> 50  0.8932951 0.09504360 0.01166132
#> 51  0.7301986 0.25873808 0.01106336
#> 52  0.6590998 0.22305449 0.11784570
#> 53  0.6979729 0.20930676 0.09272032
#> 54  0.7439410 0.24246651 0.01359246
#> 55  0.7226400 0.26330293 0.01405709
#> 56  0.6605148 0.30179672 0.03768849
#> 57  0.6827511 0.30334598 0.01390293
#> 58  0.8448656 0.11985806 0.03527639
#> 59  0.7316255 0.25895060 0.00942387
#> 60  0.6809314 0.24380025 0.07526835
#> 61  0.6456979 0.18698701 0.16731512
#> 62  0.6723507 0.19397804 0.13367126
#> 63  0.6427735 0.30288127 0.05434524
#> 64  0.7815088 0.16277275 0.05571846
#> 65  0.6340866 0.35367322 0.01224022
#> 66  0.6755996 0.28324849 0.04115195
#> 67  0.6859737 0.18989655 0.12412980
#> 68  0.7310386 0.24628088 0.02268052
#> 69  0.8717347 0.11007368 0.01819161
#> 70  0.8085294 0.13335411 0.05811653
#> 71  0.6101030 0.22961137 0.16028559
#> 72  0.7200119 0.26609178 0.01389632
#> 73  0.8543168 0.11182789 0.03385530
#> 74  0.6234107 0.35949160 0.01709769
#> 75  0.8910370 0.09789618 0.01106682
#> 76  0.6887483 0.29085039 0.02040132
#> 77  0.8961208 0.09225729 0.01162187
#> 78  0.8739514 0.11503970 0.01100890
#> 79  0.6734362 0.31035872 0.01620503
#> 80  0.6112767 0.22527560 0.16344765
#> 81  0.8345886 0.12583351 0.03957785
#> 82  0.5446080 0.38439480 0.07099720
#> 83  0.6860978 0.27927483 0.03462737
#> 84  0.7326910 0.25297739 0.01433160
#> 85  0.6567097 0.29278546 0.05050486
#> 86  0.7296533 0.25181649 0.01853024
#> 87  0.6969253 0.28369722 0.01937747
#> 88  0.6288687 0.32538445 0.04574687
#> 89  0.5651699 0.24253231 0.19229776
#> 90  0.6558292 0.29846685 0.04570392
#> 91  0.7682786 0.17347772 0.05824366
#> 92  0.8987363 0.05351164 0.04775206
#> 93  0.8824410 0.10372855 0.01383048
#> 94  0.6941194 0.25807406 0.04780653
#> 95  0.7012570 0.26167501 0.03706799
#> 96  0.7771925 0.18228740 0.04052010
#> 97  0.7451067 0.20631319 0.04858016
#> 98  0.7704095 0.14767947 0.08191104
#> 99  0.8357181 0.14818648 0.01609547
#> 100 0.8992234 0.08902185 0.01175479
#> 
#> , , k_100
#> 
#>          eig1       eig2       eig3
#> 1   0.8371168 0.14206985 0.02081335
#> 2   0.7080044 0.23319742 0.05879817
#> 3   0.7436948 0.24133090 0.01497428
#> 4   0.6673337 0.28747193 0.04519433
#> 5   0.6064852 0.24102480 0.15249003
#> 6   0.8769208 0.08948160 0.03359758
#> 7   0.6319549 0.32073227 0.04731287
#> 8   0.4338725 0.36967037 0.19645718
#> 9   0.6383910 0.30710238 0.05450661
#> 10  0.7585889 0.16682109 0.07459003
#> 11  0.6815833 0.18572044 0.13269627
#> 12  0.8938308 0.09408699 0.01208225
#> 13  0.7022100 0.28393171 0.01385826
#> 14  0.8815909 0.07597638 0.04243272
#> 15  0.8176536 0.11737418 0.06497222
#> 16  0.8335008 0.14283503 0.02366414
#> 17  0.6320087 0.32247643 0.04551486
#> 18  0.7901040 0.14531060 0.06458542
#> 19  0.6070058 0.34552317 0.04747104
#> 20  0.6144328 0.25521771 0.13034951
#> 21  0.8243656 0.15329500 0.02233942
#> 22  0.6384485 0.31543218 0.04611932
#> 23  0.8821827 0.09094766 0.02686959
#> 24  0.7303034 0.21300319 0.05669344
#> 25  0.9030792 0.05834134 0.03857944
#> 26  0.7470687 0.23832802 0.01460329
#> 27  0.5381638 0.41466841 0.04716783
#> 28  0.7088037 0.16054870 0.13064762
#> 29  0.8600618 0.10214609 0.03779212
#> 30  0.8799168 0.10535074 0.01473243
#> 31  0.4830412 0.45245269 0.06450608
#> 32  0.8568148 0.12465218 0.01853302
#> 33  0.8847417 0.09926908 0.01598918
#> 34  0.7048490 0.23592332 0.05922769
#> 35  0.7033407 0.28079799 0.01586126
#> 36  0.7919000 0.13381598 0.07428405
#> 37  0.7850999 0.19023597 0.02466411
#> 38  0.7054896 0.24745402 0.04705641
#> 39  0.7063388 0.16548966 0.12817150
#> 40  0.7528923 0.18632812 0.06077958
#> 41  0.6444822 0.28211496 0.07340286
#> 42  0.7455435 0.23752748 0.01692906
#> 43  0.8443049 0.10361243 0.05208271
#> 44  0.7030429 0.27948750 0.01746963
#> 45  0.8268794 0.11970630 0.05341427
#> 46  0.6861837 0.25865293 0.05516334
#> 47  0.5604006 0.39549030 0.04410913
#> 48  0.6341426 0.30907006 0.05678729
#> 49  0.7004975 0.18095373 0.11854879
#> 50  0.8754063 0.10994977 0.01464398
#> 51  0.7412229 0.24545299 0.01332411
#> 52  0.6079940 0.27364924 0.11835673
#> 53  0.6804490 0.16341115 0.15613988
#> 54  0.7873851 0.18360993 0.02900496
#> 55  0.7346940 0.25021315 0.01509287
#> 56  0.5699481 0.38062687 0.04942503
#> 57  0.7606264 0.21220041 0.02717314
#> 58  0.8389387 0.12446706 0.03659419
#> 59  0.7425806 0.24407851 0.01334088
#> 60  0.6243353 0.31181332 0.06385143
#> 61  0.6034545 0.24719059 0.14935493
#> 62  0.6559611 0.20901286 0.13502608
#> 63  0.6288745 0.31271631 0.05840920
#> 64  0.7446004 0.18603953 0.06936003
#> 65  0.7036375 0.28094387 0.01541861
#> 66  0.6442061 0.30815759 0.04763634
#> 67  0.6328508 0.21115860 0.15599064
#> 68  0.7436204 0.23906528 0.01731430
#> 69  0.8390591 0.13977641 0.02116444
#> 70  0.7520623 0.14231187 0.10562580
#> 71  0.6146873 0.21453171 0.17078099
#> 72  0.7456755 0.23763051 0.01669402
#> 73  0.8491093 0.11718157 0.03370913
#> 74  0.7067649 0.27663732 0.01659773
#> 75  0.8975333 0.09246285 0.01000387
#> 76  0.7281935 0.25514204 0.01666450
#> 77  0.8896870 0.09764162 0.01267134
#> 78  0.8507131 0.13729489 0.01199196
#> 79  0.7139555 0.26985127 0.01619324
#> 80  0.6255301 0.20164366 0.17282627
#> 81  0.8907009 0.06399747 0.04530164
#> 82  0.5444219 0.37545016 0.08012796
#> 83  0.5951514 0.35797892 0.04686965
#> 84  0.7924302 0.18046852 0.02710126
#> 85  0.6358393 0.30410164 0.06005902
#> 86  0.7539708 0.23086050 0.01516868
#> 87  0.7310104 0.25332891 0.01566071
#> 88  0.5952914 0.35233517 0.05237343
#> 89  0.5594881 0.22670503 0.21380682
#> 90  0.6673918 0.28288808 0.04972013
#> 91  0.6700242 0.26498014 0.06499571
#> 92  0.8779710 0.09099577 0.03103320
#> 93  0.8822408 0.10344734 0.01431183
#> 94  0.5827564 0.36010790 0.05713567
#> 95  0.6377243 0.31741067 0.04486502
#> 96  0.7592415 0.19637112 0.04438740
#> 97  0.7285179 0.21824520 0.05323688
#> 98  0.7266541 0.17462378 0.09872213
#> 99  0.8293471 0.15113936 0.01951352
#> 100 0.8867504 0.10008780 0.01316177
#>