Filtering of Point Clouds

Filtering of point clouds using different methods

filter(cloud, method, radius, min_n, k, nSigma, threads = 1L, progress = TRUE)

Arguments

cloud
A data.table contain three columns representing the *XYZ* coordinates.

method
A filtering method to use. It most be "SOR" or "min_n".

radius
A numeric vector representing the radius of the sphere to consider. This needs to be used if method = "min_n".

min_n
An integer representing the minimum number of neighbors to keep a given point. This needs to be used if method = "min_n".

k
An integer vector representing the number of neighbors to consider. This needs be used if method = "SOR".

nSigma
A numeric vector representing the standard deviation multiplier. This needs to be used if method = "SOR".

threads
An integer specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware.

progress
Show progress. TRUE as default.

Value

A data.table with the filtered points

Examples

#Load data
data("pc_tree")

#Filter a sample of section of data using SOR
filter(pc_tree[1:5000,], method = "SOR", k = 20, nSigma = 1)

#>             X        Y          Z
#>    1: 9.14850 -1.11025 0.00099992
#>    2: 9.22250 -1.12225 0.00000000
#>    3: 9.12125 -1.09025 0.03899991
#>    4: 9.14725 -1.15775 0.10199999
#>    5: 9.14675 -1.10650 0.08574998
#>   ---                            
#> 4030: 8.52825 -2.08400 2.85474992
#> 4031: 8.57025 -2.05925 2.83424997
#> 4032: 8.57175 -2.07200 2.88599992
#> 4033: 8.63575 -2.06175 2.86675000
#> 4034: 8.56950 -2.00475 2.85299993

#Filter using "min_n" of 1000
filter(pc_tree, method = "min_n", radius = 1.0, min_n = 5000)

#>               X        Y       Z
#>     1:  7.72725 -0.98150 3.00250
#>     2:  7.83075 -0.96225 2.74775
#>     3:  7.79100 -1.01025 2.77925
#>     4:  7.99475 -1.02550 2.72800
#>     5:  7.91475 -0.98625 2.72575
#>    ---                          
#> 11543: 10.85050  1.40150 4.85675
#> 11544: 10.85450  1.54150 4.77000
#> 11545: 10.38700  1.62525 4.64625
#> 11546: 10.54175  1.65700 4.65300
#> 11547: 10.80000  1.61525 4.64600