Geometry features of Neighboring Points.

Estimate geometry features of neighboring points in a cloud.

geometry_features(
  cloud,
  method,
  radius,
  k,
  max_neighbour,
  distance = "euclidean",
  target = FALSE,
  threads = 1L,
  verbose = FALSE,
  progress = TRUE,
  ...
)

Arguments

cloud

A data.table with *XYZ* coordinates in the first three columns.

method

A character string specifying the method to estimate the neighbors. It most be one of "radius_search" or "knn".

radius

A numeric vector representing the radius for search to consider. This needs be used if method = "radius_search".

k

An integer vector representing the number of neighbors to consider. This needs be used if method = "knn".

max_neighbour

An integer specifying the maximum number of points to look around each query point for a given radius. This needs be used if method = "radius_search".

distance

Type of distance to calculate. "euclidean" as default. Look hnsw_knn for more options.

target

Logic. If TRUE, it consider the each target point for the calculations of geometry features.

threads

An integer specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware.

verbose

If TRUE, log messages to the console.

progress

If TRUE, log a progress bar when verbose = TRUE. Tracking progress could cause a small overhead.

...

Arguments passed to hnsw_build and hnsw_search.

Value

A array describing the point of the cloud in rows, the relative eigenvalues in columns, and the radius or k per slide. If method = "radius_search", it add in the first column the number of neighboring points.

Details

The function returns the geometry features of the neighboring points of a given point in cloud. Geometry features are represented by the relative values of the eigenvalues derived from a covariance matrix of the neighboring points. Geometry features are not estimated on target points with less than 3 neighboring points.

Author

J. Antonio Guzmán Q.

Examples

#Create cloud
example <- data.table(X = runif(200, min=0, max=10),
                      Y = runif(200, min=0, max=10),
                      Z = runif(200, min=0, max=10))

#Using knn method with two different k
k_test <- c(5, 10)
geometry_features(example, method = "knn", k = k_test)
#> , , k_5
#> 
#>          eig1       eig2         eig3
#> 1   0.9193864 0.07156593 9.047645e-03
#> 2   0.5929672 0.36294483 4.408801e-02
#> 3   0.6150555 0.32006403 6.488044e-02
#> 4   0.8096369 0.11532375 7.503933e-02
#> 5   0.7248968 0.26686516 8.238007e-03
#> 6   0.6368243 0.32944842 3.372730e-02
#> 7   0.7701714 0.22976558 6.299241e-05
#> 8   0.5211216 0.38432566 9.455278e-02
#> 9   0.7193284 0.19318073 8.749090e-02
#> 10  0.9137502 0.05502018 3.122958e-02
#> 11  0.6713260 0.32572160 2.952399e-03
#> 12  0.7280032 0.24837352 2.362327e-02
#> 13  0.6091849 0.37750420 1.331092e-02
#> 14  0.8186451 0.15855354 2.280137e-02
#> 15  0.8352860 0.13939201 2.532197e-02
#> 16  0.6974960 0.24519586 5.730810e-02
#> 17  0.5211216 0.38432566 9.455278e-02
#> 18  0.6752769 0.20722486 1.174982e-01
#> 19  0.6395648 0.24733570 1.130995e-01
#> 20  0.6462006 0.26583239 8.796706e-02
#> 21  0.7159090 0.22636410 5.772693e-02
#> 22  0.7753630 0.15956176 6.507528e-02
#> 23  0.7651580 0.21875262 1.608941e-02
#> 24  0.6493492 0.30102732 4.962345e-02
#> 25  0.6167483 0.34642529 3.682642e-02
#> 26  0.5488673 0.44734283 3.789885e-03
#> 27  0.6515201 0.29899736 4.948254e-02
#> 28  0.8921870 0.09687459 1.093841e-02
#> 29  0.8049592 0.18552118 9.519573e-03
#> 30  0.6405311 0.30351229 5.595659e-02
#> 31  0.5480241 0.27609634 1.758796e-01
#> 32  0.5815505 0.35081341 6.763609e-02
#> 33  0.7506553 0.21353670 3.580803e-02
#> 34  0.5671573 0.37812404 5.471869e-02
#> 35  0.7085399 0.27723330 1.422681e-02
#> 36  0.7804326 0.13619978 8.336761e-02
#> 37  0.7597826 0.19977776 4.043959e-02
#> 38  0.5211216 0.38432566 9.455278e-02
#> 39  0.4638558 0.40576604 1.303781e-01
#> 40  0.6583781 0.25766268 8.395924e-02
#> 41  0.5947386 0.31696764 8.829377e-02
#> 42  0.8343668 0.16494122 6.919908e-04
#> 43  0.5471280 0.35016246 1.027096e-01
#> 44  0.7878957 0.13880610 7.329820e-02
#> 45  0.5748857 0.41055400 1.456032e-02
#> 46  0.5903301 0.33525750 7.441240e-02
#> 47  0.7144280 0.27067458 1.489742e-02
#> 48  0.5673303 0.37327549 5.939423e-02
#> 49  0.6626489 0.24999541 8.735564e-02
#> 50  0.7519502 0.22742535 2.062450e-02
#> 51  0.5905660 0.38581713 2.361691e-02
#> 52  0.6511326 0.32396268 2.490473e-02
#> 53  0.6630073 0.24231343 9.467926e-02
#> 54  0.5829836 0.37478139 4.223505e-02
#> 55  0.6139680 0.38507157 9.604248e-04
#> 56  0.7291455 0.17414997 9.670456e-02
#> 57  0.7217033 0.16998153 1.083152e-01
#> 58  0.5702352 0.40289042 2.687442e-02
#> 59  0.6713260 0.32572160 2.952399e-03
#> 60  0.6626489 0.24999541 8.735564e-02
#> 61  0.7582929 0.21549700 2.621005e-02
#> 62  0.6849988 0.19740268 1.175985e-01
#> 63  0.8785136 0.11477109 6.715335e-03
#> 64  0.7944593 0.11460186 9.093883e-02
#> 65  0.7291455 0.17414997 9.670456e-02
#> 66  0.6661856 0.22927442 1.045400e-01
#> 67  0.8715296 0.08933933 3.913109e-02
#> 68  0.6432712 0.30564298 5.108579e-02
#> 69  0.6154676 0.24287410 1.416583e-01
#> 70  0.5058677 0.34639621 1.477361e-01
#> 71  0.7637517 0.21924568 1.700264e-02
#> 72  0.7038092 0.23936477 5.682607e-02
#> 73  0.7686422 0.17866534 5.269245e-02
#> 74  0.7928928 0.17809247 2.901472e-02
#> 75  0.8187323 0.16959961 1.166804e-02
#> 76  0.7075391 0.19629812 9.616277e-02
#> 77  0.7344065 0.21871235 4.688112e-02
#> 78  0.6812446 0.31509489 3.660489e-03
#> 79  0.6730169 0.32022430 6.758756e-03
#> 80  0.7028660 0.22988831 6.724571e-02
#> 81  0.6887051 0.29118561 2.010932e-02
#> 82  0.6477768 0.30314562 4.907761e-02
#> 83  0.5913888 0.29358093 1.150303e-01
#> 84  0.7427405 0.25404402 3.215483e-03
#> 85  0.8057772 0.14163970 5.258310e-02
#> 86  0.4818218 0.38138579 1.367924e-01
#> 87  0.7446216 0.24841880 6.959598e-03
#> 88  0.5076585 0.46294958 2.939191e-02
#> 89  0.7217033 0.16998153 1.083152e-01
#> 90  0.5546838 0.29959419 1.457220e-01
#> 91  0.7645298 0.21346261 2.200762e-02
#> 92  0.6334828 0.31006698 5.645026e-02
#> 93  0.5289683 0.44690971 2.412201e-02
#> 94  0.6417573 0.32099187 3.725087e-02
#> 95  0.6708893 0.32296485 6.145860e-03
#> 96  0.6296058 0.34030271 3.009145e-02
#> 97  0.6118688 0.28836162 9.976960e-02
#> 98  0.8227487 0.14895942 2.829188e-02
#> 99  0.7663915 0.22594523 7.663217e-03
#> 100 0.6187302 0.23610040 1.451694e-01
#> 101 0.7217033 0.16998153 1.083152e-01
#> 102 0.9599578 0.03908019 9.620505e-04
#> 103 0.7009872 0.22932281 6.968998e-02
#> 104 0.6233039 0.19769971 1.789964e-01
#> 105 0.5775081 0.34202824 8.046371e-02
#> 106 0.7597826 0.19977776 4.043959e-02
#> 107 0.6023889 0.32573927 7.187186e-02
#> 108 0.6205667 0.36174006 1.769329e-02
#> 109 0.8562509 0.10790882 3.584031e-02
#> 110 0.5427052 0.33872064 1.185741e-01
#> 111 0.7779315 0.21892591 3.142613e-03
#> 112 0.4672339 0.34774265 1.850235e-01
#> 113 0.5448847 0.37508819 8.002715e-02
#> 114 0.9530199 0.04385239 3.127659e-03
#> 115 0.6054080 0.36201846 3.257358e-02
#> 116 0.7597826 0.19977776 4.043959e-02
#> 117 0.5357618 0.32451918 1.397190e-01
#> 118 0.6660671 0.22372813 1.102047e-01
#> 119 0.9219522 0.05358706 2.446073e-02
#> 120 0.6152709 0.32671499 5.801409e-02
#> 121 0.6222818 0.21110398 1.666143e-01
#> 122 0.7158191 0.24136549 4.281541e-02
#> 123 0.6361222 0.26147793 1.023998e-01
#> 124 0.7285617 0.16398902 1.074493e-01
#> 125 0.8875972 0.11128198 1.120790e-03
#> 126 0.8524112 0.11512586 3.246298e-02
#> 127 0.6413841 0.25105298 1.075630e-01
#> 128 0.8263617 0.11454656 5.909170e-02
#> 129 0.8137978 0.14608750 4.011475e-02
#> 130 0.5533878 0.40328578 4.332637e-02
#> 131 0.5520660 0.41469834 3.323570e-02
#> 132 0.6628533 0.22041910 1.167276e-01
#> 133 0.6393948 0.23953209 1.210731e-01
#> 134 0.5513659 0.38453886 6.409521e-02
#> 135 0.6260261 0.27025538 1.037185e-01
#> 136 0.6574598 0.27550729 6.703287e-02
#> 137 0.7005640 0.28604820 1.338781e-02
#> 138 0.5477224 0.31651976 1.357579e-01
#> 139 0.6167483 0.34642529 3.682642e-02
#> 140 0.7046856 0.25748328 3.783113e-02
#> 141 0.7217033 0.16998153 1.083152e-01
#> 142 0.6150555 0.32006403 6.488044e-02
#> 143 0.7150295 0.24185805 4.311249e-02
#> 144 0.6644749 0.30310493 3.242014e-02
#> 145 0.6633493 0.32674203 9.908670e-03
#> 146 0.6008127 0.35653084 4.265648e-02
#> 147 0.6901759 0.27355472 3.626941e-02
#> 148 0.5786450 0.34626975 7.508520e-02
#> 149 0.6395648 0.24733570 1.130995e-01
#> 150 0.5739351 0.41253425 1.353068e-02
#> 151 0.5500095 0.43678692 1.320360e-02
#> 152 0.6369034 0.32357785 3.951879e-02
#> 153 0.7449727 0.21078221 4.424507e-02
#> 154 0.7434689 0.22880514 2.772594e-02
#> 155 0.8196326 0.16164909 1.871832e-02
#> 156 0.6185191 0.27532006 1.061608e-01
#> 157 0.7185973 0.25382039 2.758229e-02
#> 158 0.6834346 0.27538477 4.118062e-02
#> 159 0.5892052 0.40468745 6.107362e-03
#> 160 0.7570305 0.24155953 1.409954e-03
#> 161 0.8187323 0.16959961 1.166804e-02
#> 162 0.8166495 0.14055148 4.279906e-02
#> 163 0.7410067 0.17190170 8.709158e-02
#> 164 0.6154676 0.24287410 1.416583e-01
#> 165 0.7387655 0.17533797 8.589648e-02
#> 166 0.8164413 0.17445027 9.108460e-03
#> 167 0.7127427 0.20016561 8.709173e-02
#> 168 0.6548710 0.28310424 6.202473e-02
#> 169 0.8222865 0.12723542 5.047813e-02
#> 170 0.6728499 0.18682585 1.403243e-01
#> 171 0.5790168 0.27889703 1.420861e-01
#> 172 0.8634226 0.12924664 7.330763e-03
#> 173 0.5832717 0.25282787 1.639005e-01
#> 174 0.5775081 0.34202824 8.046371e-02
#> 175 0.6607852 0.29895755 4.025724e-02
#> 176 0.5858227 0.27876968 1.354077e-01
#> 177 0.6047430 0.33020525 6.505172e-02
#> 178 0.6513019 0.34056120 8.136913e-03
#> 179 0.8559734 0.12194919 2.207737e-02
#> 180 0.7778412 0.19038783 3.177095e-02
#> 181 0.7495250 0.20925764 4.121734e-02
#> 182 0.9219522 0.05358706 2.446073e-02
#> 183 0.5265390 0.30181486 1.716461e-01
#> 184 0.6830616 0.31444246 2.495963e-03
#> 185 0.5753347 0.23012721 1.945381e-01
#> 186 0.8975170 0.08384417 1.863882e-02
#> 187 0.4954550 0.36168830 1.428568e-01
#> 188 0.8638874 0.12719653 8.916068e-03
#> 189 0.7298134 0.19279764 7.738900e-02
#> 190 0.7709902 0.22708622 1.923546e-03
#> 191 0.7775211 0.13253695 8.994190e-02
#> 192 0.6048603 0.36696648 2.817322e-02
#> 193 0.6152709 0.32671499 5.801409e-02
#> 194 0.5753347 0.23012721 1.945381e-01
#> 195 0.5275105 0.39766027 7.482924e-02
#> 196 0.7217033 0.16998153 1.083152e-01
#> 197 0.6870609 0.27779338 3.514569e-02
#> 198 0.5829836 0.37478139 4.223505e-02
#> 199 0.5090279 0.40475973 8.621237e-02
#> 200 0.6630073 0.24231343 9.467926e-02
#> 
#> , , k_10
#> 
#>          eig1      eig2       eig3
#> 1   0.6239650 0.2272108 0.14882426
#> 2   0.5267052 0.3738274 0.09946746
#> 3   0.5208855 0.3935796 0.08553493
#> 4   0.5274770 0.3696826 0.10284040
#> 5   0.6046486 0.2676058 0.12774562
#> 6   0.5188404 0.3098992 0.17126035
#> 7   0.4526317 0.2973989 0.24996939
#> 8   0.5044275 0.3055434 0.19002913
#> 9   0.4954439 0.4181602 0.08639595
#> 10  0.6900794 0.2628014 0.04711917
#> 11  0.5967093 0.2779295 0.12536122
#> 12  0.6213020 0.3314317 0.04726626
#> 13  0.5476929 0.3420974 0.11020964
#> 14  0.6024322 0.3129333 0.08463450
#> 15  0.6216219 0.3028740 0.07550412
#> 16  0.5185650 0.2796210 0.20181394
#> 17  0.4285455 0.3423606 0.22909384
#> 18  0.6086963 0.2864251 0.10487859
#> 19  0.4809524 0.3907168 0.12833087
#> 20  0.5602048 0.3184714 0.12132379
#> 21  0.4529260 0.4016954 0.14537858
#> 22  0.5547331 0.3642326 0.08103427
#> 23  0.5005624 0.3797670 0.11967067
#> 24  0.5266349 0.3556811 0.11768402
#> 25  0.7148805 0.1805997 0.10451980
#> 26  0.5985752 0.2626428 0.13878198
#> 27  0.6952259 0.2140295 0.09074463
#> 28  0.6790215 0.1817633 0.13921518
#> 29  0.4449301 0.3011509 0.25391903
#> 30  0.4048513 0.3432698 0.25187896
#> 31  0.5851833 0.2984490 0.11636766
#> 32  0.5418414 0.2692830 0.18887561
#> 33  0.6170663 0.3531423 0.02979133
#> 34  0.4654535 0.3776213 0.15692528
#> 35  0.7817378 0.1740642 0.04419807
#> 36  0.5329865 0.2847304 0.18228313
#> 37  0.6500611 0.2451193 0.10481956
#> 38  0.4725277 0.4319893 0.09548303
#> 39  0.6484687 0.2184557 0.13307560
#> 40  0.6160848 0.2969538 0.08696138
#> 41  0.6370786 0.3262549 0.03666648
#> 42  0.6803371 0.1748259 0.14483702
#> 43  0.6952259 0.2140295 0.09074463
#> 44  0.4870740 0.4221958 0.09073020
#> 45  0.5265148 0.3276236 0.14586163
#> 46  0.5740455 0.2253038 0.20065073
#> 47  0.5745070 0.3162163 0.10927677
#> 48  0.4834826 0.3552548 0.16126266
#> 49  0.6653280 0.2315337 0.10313830
#> 50  0.6097616 0.3056043 0.08463410
#> 51  0.5867190 0.2570011 0.15627993
#> 52  0.7189209 0.1590535 0.12202562
#> 53  0.4217161 0.3845876 0.19369633
#> 54  0.5137962 0.3255104 0.16069339
#> 55  0.6755627 0.2224036 0.10203370
#> 56  0.5383383 0.3006696 0.16099211
#> 57  0.5528179 0.3075875 0.13959457
#> 58  0.4321482 0.3470670 0.22078478
#> 59  0.5771709 0.3302215 0.09260759
#> 60  0.4975724 0.3013210 0.20110667
#> 61  0.6071200 0.2848328 0.10804726
#> 62  0.7409987 0.1682052 0.09079611
#> 63  0.5819113 0.2801427 0.13794595
#> 64  0.7012725 0.2446285 0.05409904
#> 65  0.5713985 0.3114243 0.11717725
#> 66  0.6550127 0.2291792 0.11580806
#> 67  0.6804644 0.2720107 0.04752485
#> 68  0.5343000 0.3028442 0.16285579
#> 69  0.7768797 0.1565586 0.06656169
#> 70  0.5436090 0.3778449 0.07854608
#> 71  0.6386508 0.2903211 0.07102804
#> 72  0.6153164 0.3062544 0.07842926
#> 73  0.5910943 0.2758219 0.13308382
#> 74  0.5018924 0.3904378 0.10766984
#> 75  0.5434291 0.3137743 0.14279660
#> 76  0.6944874 0.2383451 0.06716752
#> 77  0.5737634 0.2848258 0.14141083
#> 78  0.4475801 0.3985343 0.15388552
#> 79  0.5898792 0.2790793 0.13104146
#> 80  0.6251567 0.2588115 0.11603174
#> 81  0.6257816 0.3297134 0.04450497
#> 82  0.5840158 0.3305014 0.08548280
#> 83  0.7215940 0.2029092 0.07549684
#> 84  0.5202786 0.2563666 0.22335482
#> 85  0.6148529 0.2613862 0.12376088
#> 86  0.4554104 0.3833339 0.16125568
#> 87  0.5684137 0.3320491 0.09953719
#> 88  0.5447360 0.3753983 0.07986574
#> 89  0.5618806 0.2844335 0.15368584
#> 90  0.7084551 0.2092056 0.08233937
#> 91  0.4827273 0.3113118 0.20596096
#> 92  0.5586256 0.3626892 0.07868512
#> 93  0.5591097 0.3068299 0.13406037
#> 94  0.4484870 0.4084774 0.14303566
#> 95  0.4974264 0.4252314 0.07734217
#> 96  0.5305357 0.3046165 0.16484785
#> 97  0.4669159 0.3666832 0.16640085
#> 98  0.6242091 0.2868661 0.08892485
#> 99  0.5881759 0.2367534 0.17507077
#> 100 0.5021599 0.3537536 0.14408644
#> 101 0.5427022 0.3996881 0.05760969
#> 102 0.5318662 0.3359368 0.13219701
#> 103 0.6322556 0.2016536 0.16609084
#> 104 0.6742869 0.2537180 0.07199511
#> 105 0.4897535 0.3966228 0.11362370
#> 106 0.5501892 0.4069141 0.04289669
#> 107 0.6112665 0.3396832 0.04905032
#> 108 0.5659798 0.2929346 0.14108564
#> 109 0.5474112 0.2847313 0.16785751
#> 110 0.4644880 0.3147772 0.22073482
#> 111 0.4559067 0.3121171 0.23197621
#> 112 0.4986461 0.3135582 0.18779577
#> 113 0.4540857 0.3104272 0.23548716
#> 114 0.5912995 0.2995734 0.10912705
#> 115 0.6582008 0.2667166 0.07508262
#> 116 0.6500611 0.2451193 0.10481956
#> 117 0.5290230 0.3829293 0.08804769
#> 118 0.4853253 0.3281637 0.18651103
#> 119 0.5684137 0.3320491 0.09953719
#> 120 0.5453783 0.2495138 0.20510799
#> 121 0.5771709 0.3302215 0.09260759
#> 122 0.6138949 0.2881174 0.09798778
#> 123 0.6623030 0.1990126 0.13868445
#> 124 0.6313294 0.2267371 0.14193348
#> 125 0.6118669 0.2271760 0.16095718
#> 126 0.6360721 0.2719221 0.09200580
#> 127 0.5843426 0.2934088 0.12224858
#> 128 0.7588517 0.1568707 0.08427762
#> 129 0.5638763 0.2971850 0.13893867
#> 130 0.4316071 0.3967569 0.17163602
#> 131 0.6783507 0.2603098 0.06133949
#> 132 0.6495829 0.2090101 0.14140693
#> 133 0.4302874 0.4100820 0.15963065
#> 134 0.6040344 0.3392625 0.05670308
#> 135 0.6972773 0.1677601 0.13496262
#> 136 0.4948318 0.3332213 0.17194686
#> 137 0.6370546 0.2634101 0.09953533
#> 138 0.4807157 0.3723141 0.14697022
#> 139 0.7296147 0.1922208 0.07816449
#> 140 0.5683387 0.3804616 0.05119969
#> 141 0.5353911 0.3526557 0.11195314
#> 142 0.5208855 0.3935796 0.08553493
#> 143 0.5121631 0.2815101 0.20632679
#> 144 0.6965756 0.1804500 0.12297442
#> 145 0.4960773 0.3369218 0.16700099
#> 146 0.5366812 0.3307711 0.13254772
#> 147 0.7688659 0.1292226 0.10191154
#> 148 0.6802050 0.2266576 0.09313742
#> 149 0.4574799 0.3375714 0.20494873
#> 150 0.4620909 0.3483681 0.18954099
#> 151 0.4972202 0.3809540 0.12182580
#> 152 0.5441518 0.3847532 0.07109505
#> 153 0.5350598 0.2757969 0.18914331
#> 154 0.6112665 0.3396832 0.04905032
#> 155 0.6445308 0.2941813 0.06128788
#> 156 0.5001918 0.4234408 0.07636734
#> 157 0.5713113 0.2268834 0.20180529
#> 158 0.5472609 0.2907605 0.16197857
#> 159 0.5078820 0.3543338 0.13778418
#> 160 0.5744921 0.3786706 0.04683734
#> 161 0.5259676 0.3028250 0.17120742
#> 162 0.5907992 0.3185735 0.09062727
#> 163 0.6970094 0.2450075 0.05798309
#> 164 0.4559067 0.3121171 0.23197621
#> 165 0.6448454 0.2540643 0.10109024
#> 166 0.6408079 0.2484354 0.11075674
#> 167 0.8015878 0.1533603 0.04505194
#> 168 0.5079158 0.3211943 0.17088992
#> 169 0.5501657 0.4059407 0.04389356
#> 170 0.6051542 0.2481303 0.14671541
#> 171 0.6135818 0.2682319 0.11818626
#> 172 0.5517420 0.2803212 0.16793674
#> 173 0.4520799 0.3553495 0.19257056
#> 174 0.5751404 0.2814515 0.14340814
#> 175 0.4796124 0.3238410 0.19654659
#> 176 0.5337551 0.3218794 0.14436554
#> 177 0.7347860 0.2100800 0.05513398
#> 178 0.6202512 0.3288789 0.05086994
#> 179 0.4582455 0.3245703 0.21718424
#> 180 0.5534169 0.2495113 0.19707184
#> 181 0.4648260 0.3042224 0.23095162
#> 182 0.5684137 0.3320491 0.09953719
#> 183 0.6612901 0.2517761 0.08693383
#> 184 0.5617350 0.2878849 0.15038002
#> 185 0.5783892 0.2888772 0.13273357
#> 186 0.6071200 0.2848328 0.10804726
#> 187 0.5438045 0.3510728 0.10512270
#> 188 0.6495351 0.3000667 0.05039818
#> 189 0.6031308 0.2259637 0.17090551
#> 190 0.5953049 0.3397394 0.06495566
#> 191 0.6334201 0.2821817 0.08439817
#> 192 0.5141769 0.2510614 0.23476165
#> 193 0.5732267 0.2694545 0.15731875
#> 194 0.5078820 0.3543338 0.13778418
#> 195 0.5502210 0.3594127 0.09036624
#> 196 0.4541904 0.3354263 0.21038333
#> 197 0.5315196 0.2851035 0.18337683
#> 198 0.7257463 0.2096592 0.06459447
#> 199 0.6026688 0.2443298 0.15300138
#> 200 0.5978597 0.2587429 0.14339741
#> 

#Using radius search method with two different radius
radius_test <- c(3, 4)
geometry_features(example, method = "radius_search", radius = radius_test, max_neighbour = 200)
#> , , radius_3
#> 
#>     npoints      eig1      eig2       eig3
#> 1        30 0.4260435 0.3183336 0.25562290
#> 2        14 0.5634200 0.3183706 0.11820941
#> 3        20 0.5160411 0.2806425 0.20331639
#> 4        16 0.4947944 0.3831962 0.12200942
#> 5        21 0.6064150 0.2459597 0.14762529
#> 6        29 0.4111129 0.3416813 0.24720583
#> 7        23 0.5435162 0.2745593 0.18192458
#> 8        23 0.5162294 0.2809024 0.20286827
#> 9        27 0.4103919 0.3243949 0.26521318
#> 10       27 0.4173404 0.3106582 0.27200135
#> 11       24 0.5679106 0.2349166 0.19717275
#> 12       16 0.5365419 0.2784188 0.18503930
#> 13        9 0.7290970 0.2076198 0.06328317
#> 14       12 0.5945810 0.3205492 0.08486980
#> 15       18 0.5890167 0.2582253 0.15275802
#> 16       35 0.4620311 0.3394925 0.19847637
#> 17       25 0.4962473 0.3167083 0.18704439
#> 18       25 0.4662741 0.3410518 0.19267409
#> 19       14 0.4351481 0.3131554 0.25169647
#> 20        8 0.4835479 0.3765822 0.13986987
#> 21       28 0.4025387 0.3104478 0.28701346
#> 22       14 0.4738376 0.3806820 0.14548042
#> 23       29 0.4491389 0.3544408 0.19642029
#> 24       28 0.4826634 0.3375821 0.17975454
#> 25       10 0.7148805 0.1805997 0.10451980
#> 26       35 0.4025561 0.3466171 0.25082676
#> 27       24 0.4591854 0.3745059 0.16630870
#> 28       17 0.5349930 0.2929953 0.17201166
#> 29       23 0.4639021 0.3153318 0.22076605
#> 30       25 0.3726747 0.3451260 0.28219930
#> 31       14 0.6225666 0.2728444 0.10458906
#> 32       28 0.4416176 0.3889341 0.16944831
#> 33       18 0.6537265 0.2775196 0.06875390
#> 34       24 0.4278155 0.4021341 0.17005039
#> 35       10 0.7817378 0.1740642 0.04419807
#> 36       17 0.5110849 0.2692054 0.21970969
#> 37        8 0.6681786 0.2229869 0.10883452
#> 38       19 0.5596030 0.3124207 0.12797629
#> 39        8 0.5644159 0.2986821 0.13690195
#> 40       17 0.4409344 0.4303246 0.12874105
#> 41       20 0.6021087 0.2606144 0.13727696
#> 42       13 0.7995116 0.1143335 0.08615495
#> 43       21 0.4663128 0.4174828 0.11620435
#> 44       13 0.5345193 0.4008266 0.06465409
#> 45        9 0.5908505 0.2278794 0.18127014
#> 46       30 0.5314575 0.2708529 0.19768963
#> 47       15 0.4407418 0.3539969 0.20526123
#> 48        7 0.5871270 0.3341891 0.07868387
#> 49       10 0.6653280 0.2315337 0.10313830
#> 50       12 0.5555974 0.3309249 0.11347765
#> 51       18 0.6735606 0.2018656 0.12457384
#> 52       11 0.7807120 0.1240403 0.09524767
#> 53        7 0.4903031 0.2767193 0.23297760
#> 54       15 0.4996824 0.2863090 0.21400864
#> 55       23 0.4830663 0.3801324 0.13680135
#> 56       27 0.4043245 0.3684770 0.22719855
#> 57       18 0.4701540 0.3571709 0.17267507
#> 58       28 0.5098332 0.2829856 0.20718119
#> 59       19 0.4662987 0.3103672 0.22333410
#> 60       16 0.4797531 0.3553302 0.16491664
#> 61       23 0.4377572 0.3456223 0.21662048
#> 62       13 0.5867418 0.2802123 0.13304588
#> 63       18 0.4586932 0.3148982 0.22640854
#> 64       23 0.5509949 0.2811622 0.16784289
#> 65       33 0.5092682 0.2826257 0.20810604
#> 66       20 0.5362697 0.3386215 0.12510881
#> 67       20 0.5563910 0.2672304 0.17637857
#> 68       11 0.5400417 0.3134073 0.14655098
#> 69       11 0.7884548 0.1411262 0.07041903
#> 70        9 0.6008444 0.3075625 0.09159315
#> 71        8 0.7506638 0.1883832 0.06095304
#> 72       28 0.4116855 0.3408178 0.24749674
#> 73        9 0.5423318 0.3219478 0.13572042
#> 74       30 0.4681003 0.2832035 0.24869616
#> 75        9 0.7420736 0.2050613 0.05286507
#> 76       11 0.6758580 0.2616051 0.06253689
#> 77       15 0.5340813 0.3719530 0.09396573
#> 78       23 0.3991928 0.3314218 0.26938538
#> 79       13 0.4705705 0.3284861 0.20094348
#> 80        9 0.5332144 0.3136108 0.15317479
#> 81       25 0.5856982 0.2484749 0.16582689
#> 82       17 0.4520720 0.3851524 0.16277555
#> 83       10 0.7215940 0.2029092 0.07549684
#> 84       37 0.4548976 0.3039575 0.24114492
#> 85        7 0.5980164 0.3153972 0.08658637
#> 86       16 0.6102800 0.2398389 0.14988108
#> 87       14 0.4868939 0.2696663 0.24343976
#> 88       10 0.5447360 0.3753983 0.07986574
#> 89       23 0.4476176 0.3043056 0.24807685
#> 90       26 0.4763678 0.3090989 0.21453322
#> 91       34 0.4811485 0.3242939 0.19455759
#> 92       30 0.4892223 0.3274277 0.18335005
#> 93       10 0.5591097 0.3068299 0.13406037
#> 94       13 0.4503695 0.4161091 0.13352136
#> 95       17 0.6170061 0.3011703 0.08182362
#> 96       23 0.4305465 0.3579903 0.21146323
#> 97       20 0.5014708 0.3333711 0.16515801
#> 98       13 0.6964148 0.2363250 0.06726022
#> 99        8 0.5466868 0.2800372 0.17327601
#> 100      30 0.4213046 0.3282220 0.25047344
#> 101      19 0.4653667 0.3325817 0.20205165
#> 102       9 0.5092280 0.3301152 0.16065683
#> 103      17 0.4144479 0.4059430 0.17960909
#> 104       7 0.5204341 0.3294886 0.15007724
#> 105       8 0.6250225 0.2308113 0.14416616
#> 106       7 0.6341983 0.3257780 0.04002370
#> 107      15 0.6032689 0.3292871 0.06744394
#> 108      33 0.4289589 0.3467818 0.22425926
#> 109      11 0.4533340 0.3718947 0.17477137
#> 110      22 0.4955014 0.3725761 0.13192248
#> 111      13 0.5196868 0.3125472 0.16776605
#> 112      16 0.5055944 0.3178058 0.17659977
#> 113      33 0.4839983 0.2928849 0.22311679
#> 114      29 0.4452190 0.3436165 0.21116454
#> 115      14 0.5740837 0.2693035 0.15661280
#> 116       6 0.5831308 0.3628493 0.05401994
#> 117       8 0.7410913 0.1581921 0.10071659
#> 118      24 0.5116663 0.3033695 0.18496412
#> 119      10 0.5684137 0.3320491 0.09953719
#> 120      19 0.4699266 0.3157018 0.21437161
#> 121      17 0.4055515 0.3491349 0.24531361
#> 122      12 0.5609535 0.3642410 0.07480550
#> 123       6 0.6200766 0.2154499 0.16447356
#> 124      21 0.5352489 0.3580470 0.10670405
#> 125      15 0.5818232 0.2577706 0.16040611
#> 126      28 0.5020636 0.3069685 0.19096792
#> 127      28 0.4226310 0.3248331 0.25253587
#> 128      15 0.5460828 0.3861147 0.06780251
#> 129      27 0.4176978 0.3790893 0.20321283
#> 130      15 0.5634081 0.2916847 0.14490718
#> 131      15 0.5508045 0.3736660 0.07552950
#> 132      17 0.4804280 0.3731841 0.14638791
#> 133      24 0.4299907 0.3625889 0.20742038
#> 134      12 0.5852123 0.3056489 0.10913877
#> 135      16 0.6203281 0.2430366 0.13663538
#> 136      16 0.5361205 0.3318613 0.13201814
#> 137      20 0.4449935 0.4104046 0.14460192
#> 138      23 0.4190399 0.3100410 0.27091908
#> 139      11 0.5762408 0.2907551 0.13300406
#> 140      16 0.5748325 0.2627459 0.16242163
#> 141      17 0.4657409 0.3720868 0.16217224
#> 142      21 0.5751469 0.2444167 0.18043646
#> 143       8 0.5381856 0.2986686 0.16314575
#> 144      14 0.6165715 0.2624769 0.12095166
#> 145      24 0.5330821 0.2977732 0.16914469
#> 146      12 0.4618035 0.3169249 0.22127163
#> 147      16 0.5409767 0.3341159 0.12490749
#> 148       5 0.5786450 0.3462698 0.07508520
#> 149      14 0.4351481 0.3131554 0.25169647
#> 150      29 0.4341971 0.3606761 0.20512677
#> 151      16 0.4449162 0.3545867 0.20049704
#> 152      27 0.4691624 0.3724338 0.15840383
#> 153      22 0.3971177 0.3464209 0.25646139
#> 154      13 0.6226804 0.2856238 0.09169583
#> 155      19 0.6054733 0.2952592 0.09926748
#> 156      28 0.3792407 0.3715272 0.24923210
#> 157      15 0.4979292 0.2853758 0.21669508
#> 158      18 0.5773846 0.2650248 0.15759055
#> 159      14 0.5730306 0.2878642 0.13910512
#> 160      14 0.5765735 0.3590483 0.06437827
#> 161       9 0.5908505 0.2278794 0.18127014
#> 162      18 0.5032738 0.4029522 0.09377405
#> 163      24 0.5674847 0.2603695 0.17214581
#> 164      13 0.4072283 0.3826483 0.21012348
#> 165       7 0.6627347 0.2271623 0.11010304
#> 166      11 0.5657168 0.3067991 0.12748405
#> 167      13 0.6685258 0.2441498 0.08732442
#> 168      25 0.4885362 0.2956612 0.21580265
#> 169      14 0.5804992 0.3213702 0.09813060
#> 170      13 0.5458636 0.2863592 0.16777719
#> 171      17 0.5931195 0.2066684 0.20021209
#> 172      12 0.5529730 0.2819228 0.16510413
#> 173      20 0.4553297 0.3723100 0.17236038
#> 174      11 0.6670597 0.2225119 0.11042840
#> 175      21 0.4701060 0.3458221 0.18407189
#> 176      13 0.6054138 0.2812116 0.11337457
#> 177      13 0.7506657 0.1801437 0.06919053
#> 178      23 0.3750450 0.3475695 0.27738551
#> 179      12 0.5341737 0.2682035 0.19762287
#> 180      19 0.4655749 0.4107787 0.12364640
#> 181      18 0.5096752 0.2767406 0.21358418
#> 182       8 0.5070871 0.3708132 0.12209968
#> 183      11 0.5798674 0.2643971 0.15573547
#> 184       9 0.5908505 0.2278794 0.18127014
#> 185       9 0.5073595 0.3464347 0.14620580
#> 186      19 0.5023364 0.3268054 0.17085821
#> 187      22 0.4291455 0.3788578 0.19199667
#> 188      22 0.4591667 0.2921515 0.24868179
#> 189      12 0.5533902 0.3114042 0.13520554
#> 190      16 0.5446191 0.2443669 0.21101398
#> 191      13 0.5806725 0.3337176 0.08560991
#> 192       7 0.4377956 0.3303086 0.23189579
#> 193       6 0.6166769 0.2349576 0.14836554
#> 194       8 0.4637570 0.4335238 0.10271923
#> 195      34 0.4830370 0.2727788 0.24418420
#> 196      18 0.4612651 0.3650735 0.17366137
#> 197      16 0.5722341 0.2656128 0.16215308
#> 198      13 0.5127953 0.2836581 0.20354669
#> 199      20 0.5418701 0.2713000 0.18682994
#> 200       8 0.5593988 0.3621907 0.07841051
#> 
#> , , radius_4
#> 
#>     npoints      eig1      eig2       eig3
#> 1        60 0.3732907 0.3475335 0.27917585
#> 2        23 0.4824570 0.3738639 0.14367914
#> 3        49 0.4063751 0.3312171 0.26240779
#> 4        45 0.4721682 0.3897062 0.13812561
#> 5        40 0.5154897 0.2937806 0.19072971
#> 6        61 0.3970181 0.3558437 0.24713824
#> 7        52 0.4342013 0.3338041 0.23199466
#> 8        47 0.4583877 0.3203513 0.22126102
#> 9        55 0.4507979 0.3205079 0.22869420
#> 10       66 0.4031419 0.3766585 0.22019952
#> 11       62 0.4569989 0.2836777 0.25932344
#> 12       31 0.4873526 0.3167098 0.19593754
#> 13       18 0.5158608 0.3495881 0.13455110
#> 14       25 0.5015049 0.3457253 0.15276984
#> 15       32 0.5105320 0.3095655 0.17990254
#> 16       55 0.3912139 0.3465046 0.26228152
#> 17       53 0.4389673 0.3545750 0.20645766
#> 18       63 0.4224305 0.3082983 0.26927123
#> 19       32 0.4361452 0.3324358 0.23141900
#> 20       13 0.5848324 0.2669245 0.14824310
#> 21       51 0.4441745 0.3202604 0.23556514
#> 22       22 0.5171786 0.3699171 0.11290425
#> 23       54 0.3757064 0.3569917 0.26730191
#> 24       43 0.4582321 0.3022910 0.23947691
#> 25       19 0.7246377 0.2335386 0.04182368
#> 26       62 0.4139716 0.3557543 0.23027415
#> 27       42 0.4981499 0.3383700 0.16348013
#> 28       33 0.4985276 0.3089070 0.19256543
#> 29       50 0.3913721 0.3381187 0.27050928
#> 30       42 0.4464474 0.3203031 0.23324949
#> 31       27 0.4551658 0.3683372 0.17649699
#> 32       65 0.4403106 0.3354005 0.22428897
#> 33       25 0.5106681 0.3117488 0.17758310
#> 34       66 0.4541523 0.2925359 0.25331177
#> 35       22 0.5103459 0.3106025 0.17905163
#> 36       36 0.4921431 0.3216695 0.18618742
#> 37       12 0.5168600 0.2911862 0.19195380
#> 38       41 0.4462965 0.3322016 0.22150186
#> 39       21 0.6356228 0.2113022 0.15307498
#> 40       33 0.5024204 0.3979627 0.09961688
#> 41       35 0.5020393 0.2703803 0.22758044
#> 42       29 0.5129345 0.3147418 0.17232368
#> 43       38 0.5485577 0.3324726 0.11896968
#> 44       27 0.4896684 0.3961567 0.11417493
#> 45       19 0.4643858 0.2749367 0.26067750
#> 46       55 0.4212864 0.3143909 0.26432270
#> 47       37 0.5363044 0.2535684 0.21012712
#> 48       10 0.4834826 0.3552548 0.16126266
#> 49       19 0.5701882 0.2728008 0.15701094
#> 50       22 0.4914943 0.2659163 0.24258938
#> 51       34 0.6051049 0.2145296 0.18036550
#> 52       22 0.4752935 0.3558057 0.16890085
#> 53       22 0.5645204 0.2612993 0.17418038
#> 54       22 0.4264834 0.2984534 0.27506319
#> 55       53 0.4309886 0.3646256 0.20438580
#> 56       48 0.5148415 0.2508369 0.23432169
#> 57       28 0.5341781 0.2456152 0.22020676
#> 58       50 0.3972090 0.3270799 0.27571109
#> 59       55 0.4669929 0.3102703 0.22273684
#> 60       25 0.5448075 0.3073694 0.14782311
#> 61       44 0.4145178 0.3605902 0.22489198
#> 62       39 0.5089690 0.3054748 0.18555620
#> 63       46 0.4713873 0.3258918 0.20272089
#> 64       48 0.4330021 0.3225261 0.24447184
#> 65       52 0.4598329 0.2925370 0.24763010
#> 66       41 0.4397620 0.2959391 0.26429889
#> 67       40 0.4856057 0.3336039 0.18079038
#> 68       26 0.4881079 0.3330516 0.17884055
#> 69       18 0.6513198 0.1887706 0.15990963
#> 70       19 0.4987943 0.3912307 0.10997498
#> 71       17 0.6076478 0.2359550 0.15639722
#> 72       57 0.4573954 0.3362011 0.20640348
#> 73       23 0.4759533 0.3844981 0.13954862
#> 74       61 0.4053025 0.3100593 0.28463825
#> 75       13 0.5354512 0.2868354 0.17771332
#> 76       14 0.7649967 0.1847369 0.05026640
#> 77       34 0.5401426 0.3344096 0.12544786
#> 78       58 0.4466555 0.3194700 0.23387448
#> 79       45 0.4286783 0.3675106 0.20381110
#> 80       21 0.5367018 0.3059739 0.15732427
#> 81       45 0.4667918 0.2937565 0.23945176
#> 82       22 0.5174759 0.3045050 0.17801911
#> 83       17 0.6442250 0.2310182 0.12475683
#> 84       62 0.4008486 0.3694571 0.22969433
#> 85       25 0.4774762 0.3383648 0.18415905
#> 86       23 0.5256880 0.3516736 0.12263841
#> 87       19 0.4653667 0.3325817 0.20205165
#> 88       23 0.4948016 0.3794117 0.12578668
#> 89       36 0.5521218 0.2579016 0.18997666
#> 90       64 0.4008661 0.3413323 0.25780155
#> 91       52 0.4772391 0.2672104 0.25555044
#> 92       59 0.3737833 0.3533318 0.27288487
#> 93       30 0.4596036 0.4210599 0.11933655
#> 94       25 0.4968545 0.3531454 0.15000006
#> 95       35 0.5295580 0.3157947 0.15464729
#> 96       51 0.3921679 0.3471230 0.26070914
#> 97       32 0.5195109 0.3366005 0.14388857
#> 98       23 0.5093380 0.3301871 0.16047486
#> 99       14 0.7250372 0.1509354 0.12402741
#> 100      54 0.3861384 0.3199857 0.29387589
#> 101      26 0.5208567 0.2857601 0.19338317
#> 102      17 0.4855815 0.2912075 0.22321104
#> 103      45 0.4644973 0.2965562 0.23894648
#> 104      15 0.6928763 0.2421366 0.06498704
#> 105      21 0.5955789 0.2841924 0.12022871
#> 106      13 0.6598158 0.2635330 0.07665112
#> 107      24 0.5516852 0.3259816 0.12233320
#> 108      67 0.3522678 0.3311148 0.31661734
#> 109      37 0.4638823 0.3201055 0.21601213
#> 110      33 0.5342466 0.2798015 0.18595189
#> 111      34 0.4386578 0.3639263 0.19741596
#> 112      32 0.5156825 0.2890918 0.19522573
#> 113      63 0.4198120 0.3416920 0.23849598
#> 114      64 0.4586661 0.3165786 0.22475528
#> 115      27 0.5603483 0.2570208 0.18263096
#> 116      12 0.5887257 0.3185420 0.09273224
#> 117      14 0.5935086 0.3415099 0.06498147
#> 118      53 0.4844376 0.3043907 0.21117173
#> 119      15 0.4005297 0.3798428 0.21962746
#> 120      35 0.4876608 0.3122178 0.20012147
#> 121      52 0.4625639 0.3041098 0.23332632
#> 122      28 0.6004260 0.3015963 0.09797772
#> 123      12 0.6331021 0.2320760 0.13482196
#> 124      43 0.4550844 0.3777270 0.16718855
#> 125      32 0.5396962 0.2852114 0.17509242
#> 126      56 0.4225839 0.3272548 0.25016135
#> 127      54 0.4748664 0.3164203 0.20871333
#> 128      36 0.5897091 0.3332971 0.07699373
#> 129      50 0.4075732 0.3259065 0.26652028
#> 130      36 0.5841524 0.2914585 0.12438907
#> 131      34 0.4372112 0.3431317 0.21965705
#> 132      33 0.3934501 0.3130333 0.29351663
#> 133      62 0.4421327 0.2949867 0.26288068
#> 134      29 0.5671601 0.2924511 0.14038881
#> 135      40 0.5644586 0.2465524 0.18898894
#> 136      33 0.4318091 0.3103765 0.25781440
#> 137      36 0.6033256 0.2405152 0.15615912
#> 138      44 0.5027164 0.3032167 0.19406691
#> 139      30 0.5021235 0.2947413 0.20313522
#> 140      28 0.6027898 0.2302714 0.16693883
#> 141      25 0.5624476 0.2316417 0.20591067
#> 142      49 0.4063751 0.3312171 0.26240779
#> 143      14 0.5637739 0.2521992 0.18402696
#> 144      29 0.5625010 0.2230365 0.21446243
#> 145      50 0.4239694 0.3186945 0.25733612
#> 146      24 0.4454463 0.3538983 0.20065538
#> 147      32 0.4779235 0.2904731 0.23160339
#> 148      16 0.5587980 0.3216612 0.11954080
#> 149      31 0.4279315 0.3316564 0.24041210
#> 150      52 0.4602162 0.3398659 0.19991790
#> 151      42 0.4221069 0.3094031 0.26848997
#> 152      55 0.4193271 0.3467196 0.23395331
#> 153      48 0.4171299 0.3269204 0.25594974
#> 154      23 0.5639305 0.2713364 0.16473310
#> 155      39 0.4798010 0.3354632 0.18473581
#> 156      48 0.4179119 0.3849965 0.19709154
#> 157      22 0.5387594 0.3035606 0.15768003
#> 158      31 0.4794967 0.3607646 0.15973869
#> 159      22 0.4314176 0.3135619 0.25502051
#> 160      23 0.4624309 0.3584536 0.17911549
#> 161      17 0.5228971 0.3156169 0.16148593
#> 162      30 0.4893073 0.3521878 0.15850496
#> 163      55 0.3893894 0.3398591 0.27075145
#> 164      33 0.4280656 0.3683730 0.20356140
#> 165      15 0.6533012 0.2204851 0.12621371
#> 166      17 0.5047256 0.3210191 0.17425526
#> 167      25 0.4723815 0.3337175 0.19390102
#> 168      54 0.4311465 0.3138412 0.25501230
#> 169      32 0.4719105 0.3679574 0.16013210
#> 170      26 0.4741210 0.3348771 0.19100198
#> 171      30 0.5789278 0.2905587 0.13051344
#> 172      26 0.5473944 0.2539842 0.19862136
#> 173      38 0.4172388 0.3354861 0.24727510
#> 174      20 0.6640654 0.2255897 0.11034487
#> 175      50 0.4249319 0.3318764 0.24319169
#> 176      30 0.5250074 0.3152525 0.15974012
#> 177      26 0.5203028 0.3969683 0.08272888
#> 178      40 0.4786448 0.3087125 0.21264271
#> 179      21 0.5098957 0.2807248 0.20937948
#> 180      37 0.5792596 0.2384759 0.18226455
#> 181      40 0.5363945 0.2438629 0.21974260
#> 182      14 0.4868939 0.2696663 0.24343976
#> 183      19 0.4652890 0.3005483 0.23416265
#> 184      18 0.4438562 0.2929962 0.26314763
#> 185      11 0.5540750 0.2848619 0.16106306
#> 186      41 0.4765200 0.3418087 0.18167137
#> 187      42 0.4026621 0.3759093 0.22142857
#> 188      38 0.4589783 0.3254723 0.21554944
#> 189      27 0.5621689 0.2260172 0.21181386
#> 190      37 0.5427408 0.2438665 0.21339272
#> 191      21 0.5068612 0.3875752 0.10556351
#> 192      13 0.5835078 0.2439842 0.17250800
#> 193      23 0.5190497 0.3556136 0.12533669
#> 194       9 0.5073595 0.3464347 0.14620580
#> 195      60 0.4117233 0.3303738 0.25790287
#> 196      33 0.5061850 0.2853517 0.20846332
#> 197      26 0.5666347 0.2735500 0.15981528
#> 198      20 0.4515835 0.3385606 0.20985589
#> 199      35 0.4884379 0.2978237 0.21373841
#> 200      14 0.6151180 0.2160026 0.16887938
#>