Radius Search of Points

Radius searching of points based on FLANN C++

radius_search(
  query,
  ref,
  radius,
  max_neighbour,
  same = FALSE,
  build = "kdtree",
  threads = 1L,
  checks = 1L
)

Arguments

query
A data.table containing the set of query points where each row represent a point and each column a given coordinate.

ref
A numeric containing the set of reference points where each row represent a point and each column a given coordinate.

radius
A numeric describing the euclidean distance form the each query points.

max_neighbour
An integer specifying the maximun number of points to look around each query point for a given radius.

same
Logic. If TRUE, it delete neighbors with distance of 0, useful when the k search is based on the same query.

build
A character describing the search structure to be used: "kdtree", "kmeans", "linear".

threads
An integer specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware. If 0, then the maximum allowable cores are used.

checks
Number of checks during searching. Higher value gives better search precision but takes longer. See FLANN C++ manual for more details.

Value

A data.table with three columns describing the indices of the query and ref points and the distances.

Details

This function is based on the FLAAN C++ Library (Muja and Lowe 2009) for a fast estimation of neighbors points. It is adapted from rflann (Muja, Lowe, Yee, 2018) to simplify the workflow with TLS data. If you use this function, please consider cite both documents.

References

Muja, M., Lowe, D.G. (2009). Fast approximate nearest neighbors with automatic algorithm configuration. VISAPP (1), 2(331-340), 2.

Muja, M., Lowe, D.G., Yee, J., (2018). rflann: Basic R Interface to the 'FLANN' C++ Library. R package version 1.6.0.

See also

Author

J. Antonio Guzmán Q.

Examples

#Point cloud
data("pc_tree")

#Radius search of 0.1
radius_search(pc_tree, pc_tree, radius = 0.1, 100, TRUE, "kdtree")

#>          query   ref    distance
#>       1:     1     3 0.002586577
#>       2:     1     2 0.005620912
#>       3:     1    12 0.006598868
#>       4:     1     5 0.007199697
#>       5:     1     8 0.010286098
#>      ---                        
#> 7467925: 75848 74488 0.095604995
#> 7467926: 75848 74590 0.095818175
#> 7467927: 75848 70848 0.098668556
#> 7467928: 75848 74496 0.099113594
#> 7467929: 75848 75828 0.099285655