Point Cloud Sphere Covering

Create spheres between two neighboring points in a cloud.

sphere_covering(cloud, radius, kmax = 1, threads = 1, progress = TRUE)

Arguments

cloud
A data.table with *XYZ* coordinates in the first three columns.

radius
A numeric vector representing the maximun radius of the spheres to consider.

kmax
An integer vector equal or greather than 1 representing the number of k nearest neighbors to consider if a points present more than k nearst neighbors with the same distance beteween them.

threads
An integer specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware.

progress
Show progress. TRUE as default.

Value

A data.table describing the cartesian position of the spheres in the cloud and their radius.

Details

The function compute a sphere between to a point and its nearst neighboring point, similar to Wang et al. 2020. If a point present two or more nearst neighboring points with the same distance, kmax could be used to retrieve a maximun of k-nearst points (e.g. usefull in resample point clouds where several points may share the same distance to a target point).

References

Wang, D., Schraik, D., Hovi, A., & Rautiainen, M. (2020). Direct estimation of photon recollision probability using terrestrial laser scanning. Remote Sensing of Environment, 247, 111932.

Author

J. Antonio Guzmán Q.

Examples

###Estimate the dimensionality on a sample of 100 points.
#Load data
data("pc_tree")

#Run
sphere_covering(pc_tree[1:100,], 0.2)

#>            X         Y          Z       radius
#>  1: 9.147625 -1.108375 0.04337495 0.0020691025
#>  2: 9.223125 -1.122375 0.09450000 0.0006375030
#>  3: 9.123250 -1.089375 0.10874992 0.0021828983
#>  4: 9.147875 -1.134000 0.05149996 0.0237582307
#>  5: 9.148375 -1.105500 0.22799998 0.0019076809
#>  6: 9.136625 -1.110250 0.10899997 0.0107972729
#>  7: 9.271375 -1.097000 0.11837494 0.0008385971
#>  8: 9.237750 -1.116875 0.25224996 0.0036933699
#>  9: 9.207250 -1.121625 0.25262499 0.0064963675
#> 10: 9.278625 -1.098500 0.23924994 0.0025154736
#> 11: 9.241875 -1.112000 0.09062492 0.0058218096
#> 12: 9.085375 -1.064000 0.22837495 0.0022532685
#> 13: 9.083125 -1.026000 0.14012492 0.0039466934
#> 14: 9.045125 -0.983250 0.06149995 0.0091146267
#> 15: 9.085875 -1.039125 0.13612497 0.0043480401
#> 16: 9.083250 -1.005000 0.26087492 0.0042788845
#> 17: 9.126250 -1.088750 0.24949992 0.0010307049
#> 18: 9.151375 -1.104500 0.32349997 0.0004505483
#> 19: 9.186625 -1.119250 0.21924996 0.0094114189
#> 20: 9.279500 -1.092750 0.11999994 0.0035089525
#> 21: 9.196000 -1.128250 0.36912495 0.0019526815
#> 22: 9.239750 -1.127625 0.29012496 0.0031252482
#> 23: 9.087500 -1.033375 0.23749995 0.0016441742
#> 24: 9.080875 -0.968625 0.18837492 0.0084722700
#> 25: 9.087375 -1.028500 0.26587492 0.0032715940
#> 26: 9.098125 -1.081625 0.36962494 0.0094753803
#> 27: 9.083750 -0.983375 0.24137496 0.0068795489
#> 28: 9.130625 -1.090750 0.29524994 0.0038017664
#> 29: 9.150500 -1.104375 0.29612500 0.0005155956
#> 30: 9.196125 -1.125500 0.39887494 0.0020349951
#> 31: 9.277625 -1.101375 0.34774995 0.0006373794
#> 32: 9.200000 -1.127750 0.43199992 0.0026101307
#> 33: 9.254250 -1.109500 0.24412495 0.0068233046
#> 34: 9.085625 -0.994625 0.30474996 0.0048638917
#> 35: 9.101000 -0.960625 0.36837494 0.0045016455
#> 36: 9.108375 -1.053875 0.41987494 0.0048252874
#> 37: 9.087625 -1.002250 0.36599993 0.0032523936
#> 38: 9.105750 -0.960500 0.39999998 0.0003534438
#> 39: 9.147625 -1.101000 0.42712498 0.0042295718
#> 40: 9.205500 -1.128250 0.38049996 0.0030101827
#> 41: 9.191250 -1.130125 0.34199994 0.0033096599
#> 42: 9.238000 -1.123500 0.38537497 0.0034821922
#> 43: 9.278500 -1.103125 0.38749996 0.0013520942
#> 44: 9.091625 -1.008000 0.41749996 0.0048235454
#> 45: 9.097125 -0.966125 0.39662492 0.0056595876
#> 46: 9.281375 -0.966375 0.34862494 0.0019120598
#> 47: 9.311000 -1.064250 0.22137493 0.0015206775
#> 48: 9.342875 -1.032250 0.27299994 0.0032715990
#> 49: 9.335000 -0.961375 0.06749993 0.0073024415
#> 50: 9.334500 -0.988250 0.17424994 0.0064032203
#> 51: 9.386750 -0.977750 0.09949993 0.0169059014
#> 52: 9.312250 -1.058125 0.13787496 0.0013053008
#> 53: 9.309625 -1.058250 0.19075000 0.0007603437
#> 54: 9.347250 -1.026875 0.07949995 0.0045291734
#> 55: 9.323875 -0.961125 0.14537495 0.0039253339
#> 56: 9.310250 -1.058750 0.16574996 0.0007906263
#> 57: 9.325125 -1.014625 0.19824994 0.0024558995
#> 58: 9.372875 -0.990750 0.19199996 0.0025769689
#> 59: 9.310625 -0.962000 0.35699994 0.0019565505
#> 60: 9.339250 -1.041500 0.39937496 0.0068234854
#> 61: 9.327875 -0.986250 0.32862496 0.0062161885
#> 62: 9.321000 -0.966250 0.25037497 0.0046097570
#> 63: 9.377000 -1.003750 0.25824993 0.0115245493
#> 64: 9.306750 -1.068625 0.17874998 0.0049261519
#> 65: 9.327000 -1.024875 0.40237499 0.0008005272
#> 66: 9.326750 -0.979625 0.40449995 0.0008002305
#> 67: 9.305500 -1.076500 0.42812496 0.0057065971
#> 68: 8.965750 -0.906250 0.09499997 0.0085036993
#> 69: 9.016125 -0.919750 0.05224996 0.0072510614
#> 70: 9.006125 -0.871250 0.09574998 0.0020498235
#> 71: 9.005000 -0.940000 0.07524997 0.0170291873
#> 72: 9.057250 -0.868750 0.05424999 0.0214970912
#>            X         Y          Z       radius