Stand Counting

Applies the voxels_counting function on a grid base point cloud.

stand_counting(
  cloud,
  xy.res,
  z.res = NULL,
  points.min = NULL,
  min_size,
  edge_sizes = NULL,
  length_out = 10,
  bootstrap = FALSE,
  R = NULL,
  progress = TRUE,
  parallel = FALSE,
  threads = NULL
)

Arguments

cloud
A data.table of a point cloud with xyz coordinates in the first three columns.

xy.res
A positive numeric vector describing the grid resolution of the xy coordinates to perform.

z.res
A positive numeric vector of length 1 describing the vertical resolution. If z.res = NULL vertical profiles are not used.

points.min
A positive numeric vector of length 1 minimum number of points to retain a sub-grid.

min_size
A positive numeric vector of length 1 describing the minimum cube edge length to perform. This is required if edge_sizes = NULL.

edge_sizes
A positive numeric vector describing the edge length of the different cubes to perform within each subgrid when z.res = NULL. If edge_sizes = NULL, it uses the maximum range of values for the xyz coordinates.

length_out
A positive interger of length 1 indicating the number of different edge lengths to use for each subgrid. This is required if edge_sizes = NULL.

bootstrap
Logical. If TRUE, it computes a bootstrap on the H index calculations. FALSE as default.

R
A positive integer of length 1 indicating the number of bootstrap replicates. This need to be used if bootstrap = TRUE.

progress
Logical, if TRUE displays a graphical progress bar. TRUE as default.

parallel
Logical, if TRUE it uses a parallel processing for the voxelization. FALSE as default.

threads
An integer >= 0 describing the number of threads to use. This need to be used if parallel = TRUE.

Value

A data.table with the summary of the voxels per grid created with their features.

See also

Author

J. Antonio Guzmán Q.

Examples

data(pc_tree)

#Applying stand_counting.
if (FALSE) {
stand_counting(pc_tree, xy.res = c(4, 4), min_size = 3)
}

#Applying stand_counting using bootstrap in the H index.
if (FALSE) {
stand_counting(pc_tree,
               xy.res = c(4, 4),
               min_size = 3,
               bootstrap = TRUE,
               R = 10)
}