Canopy Structure

Estimates the canopy structure from a discrete returns scan from different TLS.

canopy_structure(
  TLS.type,
  scan,
  zenith.range,
  zenith.rings,
  azimuth.range,
  vertical.resolution,
  TLS.pulse.counts,
  TLS.resolution = NULL,
  TLS.coordinates = c(0, 0, 0),
  TLS.frame = NULL,
  TLS.angles = NULL,
  threads = 1
)

Arguments

TLS.type
A character describing is the TLS used. It most be one of "single" return, "multiple" return, or "fixed.angle" scanner.

scan
If TLS.type is equal to "single" or "fixed.angle", a data.table with three columns describing *XYZ* coordinates of the discrete return. If TLS.type is equal to "multiple", a data.table with four columns describing *XYZ* coordinates and the target count pulses. Currently, "fixed.angle" present errors, use with discretion.

zenith.range
If TLS.type is equal to "single" or "multiple", a numeric vector of length two describing the min and max range of the zenith angle to use. Theoretically, the max range should be lower than 90 degrees.

zenith.rings
If TLS.type is equal to "single" or "multiple", a numeric vector of length one describing the number of zenith rings to use between zenith.range. This is used to estimate the frequency of laser shots from the scanner and returns in scan. If TLS.type = "fixed.angle", zenith.rings = 1 be default.

azimuth.range
A numeric vector of length two describing the range of the azimuth angle to use. Theoretically, it should be between 0 and 360 degrees.

vertical.resolution
A numeric vector of length one describing the vertical resolution to extract the vertical profiles. Low values lead to more variable profiles. The scale used needs to be in congruence with the scale of scan.

TLS.pulse.counts
If TLS.type is equal to "single" or "multiple", a numeric vector of length two describing the horizontal and vertical pulse counts of the scanner. If TLS.type is equal to "fixed.angle", a numeric vector of length one describing the horizontal pulse counts resolution. Preferred parameter over TLS.resolution to estimate the number of pulses.

TLS.resolution
If TLS.pulse.counts = NULL, the code use the angles resolution to estimate the pulse counts in a given TLS.frame. If TLS.type is equal to "single" or "multiple", a numeric vector of length two describing the horizontal and vertical angle resolution of the scanner. If TLS.type is equal to "fixed.angle", a numeric vector of length one describing the horizontal angle resolution.

TLS.coordinates
A numeric vector of length three describing the scanner coordinates within scan. It assumes that the coordinates are c(X = 0, Y = 0, Z = 0) for default.

TLS.frame
If TLS.type is equal to "single" or "multiple", a numeric vector of length four describing the min and max of the zenith and azimuth angle of the scanner frame. If TLS.type = "fixed.angle", a numeric vector of length three describing the fixed zenith angle and the min and max of the azimuth angle of the scanner frame. If NULL, it assumes that a complete hemisphere (c(zenith.min = 0, zenith.max = 90, azimuth.min = 0, azimuth.max = 360)), or a cone projection (c(zenith = 57.5, azimuth.min = 0, azimuth.max = 360)) depending on TLS.type.

TLS.angles
A numeric vector of length three describing the roll (*X*), pitch (*Y*), and yaw (*Z*) angles of the scanner during the scan. If NULL, it assumes that there is no need to to correction of angles. This needs to be used if TLS.type is equal to "single" or "multiple", since it assumes that "fixed.angle" scanner is previously balanced. NULL as default.

threads
An integer specifying the number of threads to use. Experiment to see what works best for your data on your hardware.

Value

For any TLS.type, it returns a data.table with the height profiles defined by vertical.resolution, the gap probability based on the zenith.range and zenith.rings, and the accumulative L(z) profiles based on the closest zenith ring to 57.5 degrees (hinge angle). If TLS.type is equal to "fixed.angle", it returns f(z) or commonly named PAVD based on on the ratio of the derivative of L(z) and the derivative of height (z). If TLS.type is equal to "single" or "multiple", it returns the normalized average weighting L/LAI, and PAVD: based on the L (hinge angle) at the highest height and the ratio between the derivative of L/LAI average weighted and the derivative of z.

Details

Since scan describes discrete returns measured by the TLS, canopy_structre first simulates the number of pulses emitted based on Danson et al. (2007). The simulated pulses are created based on the TLS properties (TLS.pulse.counts, TLS.resolution, TLS.frame) assuming that the scanner is perfectly balance. Then these pulses are rotated (rotate3D) based on the TLS.angles roll, pitch, and yaw, and move to TLS.coordintates to simulate the positioning of the scanner during the scan. Rotated simulated-pulses of interest and scan returns are then extracted based on the zenith.range and azimuth.range for a given number of zenith.rings, azimuth.rings and vertical profiles. Using the frequency of pulses and returns the probability of gap (Pgap) is estimated. For TLS.type = "multiple", the frequency of returns is estimated using the sum of 1/target count following Lovell et al. (2011).

Using the Pgap estimated per each zenith ring and vertical profile, canopy_structure then estimates the accumulative L(z) profiles based on the closest zenith ring to 57.5 (hinge region) and, if TLS.type = "fixed.angle", the f(z) or commonly named PAVD based on the ratio of the derivative of L(z) and height (z) following Jupp et al. 2009 (Equation 18). If TLS.type is equal to "single" or "multiple", canopy_structure also estimates the normalized average weighted L/LAI, and then PAVD based on the L (hinge angle) at the highest height (LAI) and the ratio between the derivative of L/LAI (average weighted) and the derivative of z (Jupp et al. 2009; Equation 21).

Jupp et al. 2009 excludes the zero zenith or fist ring to conduct the average weighted L/LAI estimations, canopy_structre does not excludes this sections since it depends on the regions of interest of the user. Therefore, user should consider this difference since it may introduce more variability to profile estimations.

References

Danson F.M., Hetherington D., Morsdorf F., Koetz B., Allgower B. 2007. Forest canopy gap fraction from terrestrial laser scanning. IEEE Geosci. Remote Sensing Letters 4:157-160. doi: 10.1109/LGRS.2006.887064.

Lovell J.L., Jupp D.L.B., van Gorsel E., Jimenez-Berni J., Hopkinson C., Chasmer L. 2011. Foliage profiles from ground based waveform and discrete point LiDAR. In: SilviLaser 2011, Hobart, Australia, 16–20 October 2011.

Jupp D.L.B., Culvenor D.S., Lovell J.L., Newnham G.J., Strahler A.H., Woodcock C.E. 2009. Estimating forest LAI profiles and structural parameters using a ground-based laser called “Echidna®”. Tree Physiology 29(2): 171-181. doi: 10.1093/treephys/tpn022.

Examples

#Using a multiple return file

if (FALSE) {
data(TLS_scan)
TLS_scan <- TLS_scan[, 1:4] #Select the four columns required

#This will take a while#
canopy_structure(TLS.type = "multiple", scan = TLS_scan, zenith.range = c(50,70), zenith.rings = 4,
                 azimuth.range = c(0, 360), vertical.resolution = 0.25,
                 TLS.resolution = c(0.04, 0.04), TLS.frame = c(30, 130, 0, 360),
                 TLS.angles =  c(0.293, -0.835, -150.159))

#Using a single return file

data(TLS_scan)
TLS_scan <- TLS_scan[Target_index == 1, 1:3] #Subset to first return observations

#This will take a while#
canopy_structure(TLS.type = "single", scan = TLS_scan, zenith.range = c(50,70), zenith.rings = 4,
                 azimuth.range = c(0, 360), vertical.resolution = 0.25,
                 TLS.resolution = c(0.04, 0.04), TLS.frame = c(30, 130, 0, 360),
                 TLS.angles =  c(0.293, -0.835, -150.159))
}