Geometry features of Neighboring Points.

Estimate geometry features of neighboring points in a cloud.

Usage,
geometry_features(
  cloud,
  method,
  radius,
  k,
  max_neighbour,
  distance = "euclidean",
  target = FALSE,
  threads = 1L,
  verbose = FALSE,
  progress = TRUE,
  ...
)

Arguments

cloud

A data.table with *XYZ* coordinates in the first three columns.

method

A character string specifying the method to estimate the neighbors. It most be one of "radius_search" or "knn".

radius

A numeric vector representing the radius for search to consider. This needs be used if method = "radius_search".

k

An integer vector representing the number of neighbors to consider. This needs be used if method = "knn".

max_neighbour

An integer specifying the maximum number of points to look around each query point for a given radius. This needs be used if method = "radius_search".

distance

Type of distance to calculate. "euclidean" as default. Look hnsw_knn for more options.

target

Logic. If TRUE, it consider the each target point for the calculations of geometry features.

threads

An integer specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware.

verbose

If TRUE, log messages to the console.

progress

If TRUE, log a progress bar when verbose = TRUE. Tracking progress could cause a small overhead.

...

Arguments passed to hnsw_build and hnsw_search.

Value

A array describing the point of the cloud in rows, the relative eigenvalues in columns, and the radius or k per slide. If method = "radius_search", it add in the first column the number of neighboring points.

Details

The function returns the geometry features of the neighboring points of a given point in cloud. Geometry features are represented by the relative values of the eigenvalues derived from a covariance matrix of the neighboring points. Geometry features are not estimated on target points with less than 3 neighboring points.

Author

J. Antonio Guzmán Q.

Examples

#Create cloud
example <- data.table(X = runif(200, min=0, max=10),
                      Y = runif(200, min=0, max=10),
                      Z = runif(200, min=0, max=10))

#Using knn method with two different k
k_test <- c(5, 10)
geometry_features(example, method = "knn", k = k_test)
#> , , k_5
#> 
#>          eig1       eig2         eig3
#> 1   0.7361159 0.22263976 0.0412443670
#> 2   0.8429606 0.15641967 0.0006196987
#> 3   0.8949633 0.08158526 0.0234513967
#> 4   0.7953617 0.12200424 0.0826340741
#> 5   0.5471039 0.32546102 0.1274351073
#> 6   0.6471442 0.18875577 0.1641000684
#> 7   0.5792058 0.33742251 0.0833717396
#> 8   0.5878762 0.36241553 0.0497083009
#> 9   0.7242957 0.21627309 0.0594312202
#> 10  0.6419012 0.31052459 0.0475742115
#> 11  0.6215472 0.34412391 0.0343289274
#> 12  0.7243383 0.25098622 0.0246754682
#> 13  0.5184593 0.40717691 0.0743638196
#> 14  0.7271750 0.23639064 0.0364343367
#> 15  0.5686097 0.38955963 0.0418306298
#> 16  0.7315636 0.17138383 0.0970525861
#> 17  0.7860941 0.16166544 0.0522404339
#> 18  0.8197023 0.12119051 0.0591071970
#> 19  0.8098231 0.17799778 0.0121791589
#> 20  0.7776605 0.12612017 0.0962193579
#> 21  0.6929147 0.27329452 0.0337907358
#> 22  0.7229596 0.26195947 0.0150809821
#> 23  0.5496115 0.31973054 0.1306580008
#> 24  0.7869275 0.20561670 0.0074558394
#> 25  0.8881206 0.06509082 0.0467885809
#> 26  0.6307630 0.30565366 0.0635833737
#> 27  0.6808173 0.17725897 0.1419236832
#> 28  0.5833349 0.26910130 0.1475637895
#> 29  0.8764167 0.11651882 0.0070644298
#> 30  0.7578725 0.23041303 0.0117144872
#> 31  0.5805678 0.33349567 0.0859364953
#> 32  0.5496115 0.31973054 0.1306580008
#> 33  0.7066598 0.27905913 0.0142810727
#> 34  0.7446776 0.23719309 0.0181292633
#> 35  0.8096861 0.16753224 0.0227816207
#> 36  0.6201118 0.34673696 0.0331512019
#> 37  0.5844715 0.41285377 0.0026747706
#> 38  0.6369638 0.29531888 0.0677172792
#> 39  0.6590533 0.22049208 0.1204546574
#> 40  0.6069261 0.36840508 0.0246688004
#> 41  0.5903116 0.38704253 0.0226458730
#> 42  0.7195217 0.27944351 0.0010347618
#> 43  0.9125539 0.08354942 0.0038966608
#> 44  0.5485961 0.40634266 0.0450612094
#> 45  0.7357049 0.15328740 0.1110076779
#> 46  0.6893210 0.18183831 0.1288407275
#> 47  0.5273627 0.37198236 0.1006549071
#> 48  0.8603187 0.13005298 0.0096283428
#> 49  0.7600554 0.18031093 0.0596336458
#> 50  0.8168689 0.14274198 0.0403891240
#> 51  0.6194084 0.31832384 0.0622677459
#> 52  0.7003631 0.27467413 0.0249627657
#> 53  0.6514528 0.28639556 0.0621516741
#> 54  0.6112853 0.37919571 0.0095189600
#> 55  0.6494439 0.29336781 0.0571882913
#> 56  0.7457482 0.24890945 0.0053423680
#> 57  0.6926728 0.28531629 0.0220108876
#> 58  0.6802224 0.31533056 0.0044470055
#> 59  0.8641519 0.10646769 0.0293803822
#> 60  0.7640802 0.22152323 0.0143965998
#> 61  0.7398399 0.17933996 0.0808201156
#> 62  0.5686097 0.38955963 0.0418306298
#> 63  0.5868446 0.33205170 0.0811037355
#> 64  0.8949633 0.08158526 0.0234513967
#> 65  0.8996812 0.08545688 0.0148619244
#> 66  0.4882333 0.31867348 0.1930932639
#> 67  0.6513647 0.22245875 0.1261765038
#> 68  0.7669590 0.23087940 0.0021615690
#> 69  0.5695674 0.40662412 0.0238084411
#> 70  0.7240062 0.18803872 0.0879550669
#> 71  0.6914577 0.29440993 0.0141324022
#> 72  0.6745696 0.29183062 0.0335997798
#> 73  0.6867298 0.28705962 0.0262105846
#> 74  0.7204778 0.23792063 0.0416015716
#> 75  0.7465242 0.22194722 0.0315285839
#> 76  0.8336347 0.09599621 0.0703690433
#> 77  0.8275612 0.11071647 0.0617223242
#> 78  0.7457482 0.24890945 0.0053423680
#> 79  0.6581726 0.33949640 0.0023309787
#> 80  0.5975584 0.40006056 0.0023810436
#> 81  0.8067362 0.13468297 0.0585808573
#> 82  0.6746351 0.25424254 0.0711223766
#> 83  0.5203182 0.39148659 0.0881952394
#> 84  0.8297048 0.14578128 0.0245138974
#> 85  0.6230296 0.31077073 0.0661997047
#> 86  0.5686097 0.38955963 0.0418306298
#> 87  0.5980814 0.37900486 0.0229136996
#> 88  0.4853415 0.29148010 0.2231783986
#> 89  0.4772637 0.35798686 0.1647494568
#> 90  0.7006729 0.24178608 0.0575410169
#> 91  0.6555575 0.27909718 0.0653452710
#> 92  0.7750527 0.16685367 0.0580936072
#> 93  0.8765568 0.11297508 0.0104681534
#> 94  0.8126403 0.11513716 0.0722225875
#> 95  0.7578725 0.23041303 0.0117144872
#> 96  0.7210408 0.24647438 0.0324848055
#> 97  0.5747440 0.34912930 0.0761266647
#> 98  0.8791253 0.11014678 0.0107279525
#> 99  0.4979242 0.29616797 0.2059077865
#> 100 0.7797571 0.19302726 0.0272156839
#> 101 0.6731298 0.23229078 0.0945794482
#> 102 0.6519336 0.21329775 0.1347686062
#> 103 0.7552298 0.24235156 0.0024186813
#> 104 0.5184593 0.40717691 0.0743638196
#> 105 0.5516057 0.26600496 0.1823893453
#> 106 0.6403010 0.32708763 0.0326113444
#> 107 0.8791253 0.11014678 0.0107279525
#> 108 0.7264884 0.23663763 0.0368739242
#> 109 0.8037614 0.18469963 0.0115389622
#> 110 0.7332014 0.25431171 0.0124868813
#> 111 0.8113999 0.16027541 0.0283247175
#> 112 0.7905928 0.19740069 0.0120064694
#> 113 0.5975830 0.40197288 0.0004440809
#> 114 0.6140556 0.20320661 0.1827377521
#> 115 0.8454792 0.10645163 0.0480691284
#> 116 0.7270271 0.16406766 0.1089051883
#> 117 0.7727049 0.20212017 0.0251749225
#> 118 0.9125539 0.08354942 0.0038966608
#> 119 0.6369638 0.29531888 0.0677172792
#> 120 0.7672427 0.16190731 0.0708500226
#> 121 0.5114287 0.46162643 0.0269448214
#> 122 0.8254072 0.15981088 0.0147818916
#> 123 0.5769387 0.29486064 0.1282006825
#> 124 0.7474686 0.17020156 0.0823298075
#> 125 0.7817186 0.21645147 0.0018299105
#> 126 0.6194750 0.26603254 0.1144925057
#> 127 0.5421040 0.43569676 0.0221992340
#> 128 0.8113999 0.16027541 0.0283247175
#> 129 0.6369638 0.29531888 0.0677172792
#> 130 0.8254072 0.15981088 0.0147818916
#> 131 0.9193318 0.06797010 0.0126980680
#> 132 0.5538900 0.44174280 0.0043672441
#> 133 0.6788771 0.29536787 0.0257550061
#> 134 0.6216328 0.27111242 0.1072548236
#> 135 0.5552327 0.36457619 0.0801911259
#> 136 0.8566619 0.14269410 0.0006440018
#> 137 0.7496189 0.24290805 0.0074730471
#> 138 0.7639952 0.17910501 0.0568998265
#> 139 0.7776605 0.12612017 0.0962193579
#> 140 0.8175114 0.11969659 0.0627919990
#> 141 0.6069261 0.36840508 0.0246688004
#> 142 0.5643948 0.39942047 0.0361846909
#> 143 0.8640965 0.08109012 0.0548134191
#> 144 0.7575026 0.24095223 0.0015451775
#> 145 0.5551000 0.38303302 0.0618669351
#> 146 0.8014202 0.19719322 0.0013865484
#> 147 0.8316475 0.16734587 0.0010066027
#> 148 0.5605638 0.30932884 0.1301073407
#> 149 0.8793707 0.11874347 0.0018857934
#> 150 0.6090025 0.33298203 0.0580154260
#> 151 0.6522056 0.32349100 0.0243034361
#> 152 0.5792058 0.33742251 0.0833717396
#> 153 0.6329570 0.31600194 0.0510410851
#> 154 0.6230296 0.31077073 0.0661997047
#> 155 0.5815996 0.40790542 0.0104949817
#> 156 0.7496189 0.24290805 0.0074730471
#> 157 0.6596081 0.29417035 0.0462215233
#> 158 0.6326063 0.29030323 0.0770904662
#> 159 0.8297048 0.14578128 0.0245138974
#> 160 0.7457482 0.24890945 0.0053423680
#> 161 0.8316475 0.16734587 0.0010066027
#> 162 0.6729182 0.18743454 0.1396472084
#> 163 0.6443755 0.33592272 0.0197017946
#> 164 0.6443755 0.33592272 0.0197017946
#> 165 0.6369638 0.29531888 0.0677172792
#> 166 0.7949059 0.19231306 0.0127809993
#> 167 0.6707789 0.22430875 0.1049123019
#> 168 0.7414896 0.21234254 0.0461678630
#> 169 0.6403010 0.32708763 0.0326113444
#> 170 0.8641519 0.10646769 0.0293803822
#> 171 0.5815996 0.40790542 0.0104949817
#> 172 0.6212604 0.36402859 0.0147110278
#> 173 0.7857841 0.13880951 0.0754064119
#> 174 0.6649892 0.25324166 0.0817691860
#> 175 0.5201947 0.43543729 0.0443680363
#> 176 0.6885540 0.22127392 0.0901720627
#> 177 0.6112853 0.37919571 0.0095189600
#> 178 0.5047441 0.33766690 0.1575890192
#> 179 0.8144808 0.18098899 0.0045302335
#> 180 0.7109899 0.27327712 0.0157329366
#> 181 0.8852645 0.10781030 0.0069252083
#> 182 0.5805678 0.33349567 0.0859364953
#> 183 0.7029353 0.25269260 0.0443721456
#> 184 0.5347055 0.38460529 0.0806891950
#> 185 0.9318690 0.04820666 0.0199243238
#> 186 0.6243289 0.29064993 0.0850211684
#> 187 0.6340528 0.29726198 0.0686851718
#> 188 0.7727049 0.20212017 0.0251749225
#> 189 0.6755166 0.30506160 0.0194217906
#> 190 0.6230296 0.31077073 0.0661997047
#> 191 0.7586709 0.21326814 0.0280609758
#> 192 0.6384768 0.22689929 0.1346238822
#> 193 0.5434684 0.38842869 0.0681029576
#> 194 0.6036331 0.26280175 0.1335651120
#> 195 0.7163602 0.21465317 0.0689866223
#> 196 0.8069291 0.19153187 0.0015390656
#> 197 0.8574118 0.12923208 0.0133561691
#> 198 0.6731298 0.23229078 0.0945794482
#> 199 0.7227906 0.19075197 0.0864573826
#> 200 0.8037614 0.18469963 0.0115389622
#> 
#> , , k_10
#> 
#>          eig1      eig2       eig3
#> 1   0.6257309 0.3356461 0.03862300
#> 2   0.6582298 0.3168572 0.02491298
#> 3   0.5579494 0.2922175 0.14983305
#> 4   0.6347433 0.3243766 0.04088016
#> 5   0.5564820 0.2623480 0.18116999
#> 6   0.5453878 0.3396633 0.11494887
#> 7   0.4597216 0.3915039 0.14877454
#> 8   0.5331387 0.3646279 0.10223336
#> 9   0.5912769 0.3563289 0.05239422
#> 10  0.5441997 0.3021157 0.15368452
#> 11  0.4984754 0.2907178 0.21080684
#> 12  0.5408063 0.2544495 0.20474426
#> 13  0.4714882 0.4228305 0.10568126
#> 14  0.5222233 0.3387871 0.13898959
#> 15  0.4599512 0.3726653 0.16738346
#> 16  0.6764861 0.2534586 0.07005529
#> 17  0.4521566 0.3888590 0.15898440
#> 18  0.5231487 0.3413663 0.13548504
#> 19  0.5650862 0.3067722 0.12814152
#> 20  0.6130094 0.2375345 0.14945617
#> 21  0.5346937 0.3721372 0.09316913
#> 22  0.5720904 0.2369059 0.19100372
#> 23  0.4174371 0.3728049 0.20975803
#> 24  0.5232046 0.3191345 0.15766091
#> 25  0.6704864 0.2043569 0.12515674
#> 26  0.5228622 0.3415374 0.13560037
#> 27  0.6453757 0.2514702 0.10315405
#> 28  0.5303287 0.2763454 0.19332593
#> 29  0.5667358 0.3425359 0.09072825
#> 30  0.6176680 0.2908024 0.09152965
#> 31  0.7178719 0.2245161 0.05761203
#> 32  0.5181620 0.3015187 0.18031931
#> 33  0.5669484 0.2803369 0.15271471
#> 34  0.4632947 0.2946073 0.24209804
#> 35  0.7496636 0.1791692 0.07116722
#> 36  0.5054884 0.3791427 0.11536887
#> 37  0.5962929 0.2883061 0.11540095
#> 38  0.5552503 0.3951634 0.04958623
#> 39  0.6167187 0.2035632 0.17971813
#> 40  0.5502434 0.2773262 0.17243033
#> 41  0.4322070 0.3936828 0.17411013
#> 42  0.6573211 0.2175980 0.12508092
#> 43  0.7740764 0.1766680 0.04925555
#> 44  0.5446192 0.3139363 0.14144457
#> 45  0.4347350 0.4151143 0.15015065
#> 46  0.7184666 0.1903650 0.09116840
#> 47  0.7318882 0.2211175 0.04699424
#> 48  0.5027901 0.4156889 0.08152098
#> 49  0.5601094 0.3048540 0.13503663
#> 50  0.6069987 0.2787693 0.11423207
#> 51  0.5696423 0.3157668 0.11459095
#> 52  0.5956877 0.2504476 0.15386466
#> 53  0.5001577 0.3074527 0.19238959
#> 54  0.7091837 0.1875944 0.10322192
#> 55  0.5855810 0.2441537 0.17026529
#> 56  0.5872467 0.3060741 0.10667915
#> 57  0.5237363 0.2905270 0.18573665
#> 58  0.4495090 0.3467651 0.20372587
#> 59  0.4649348 0.2964884 0.23857678
#> 60  0.5034530 0.3759588 0.12058822
#> 61  0.6729277 0.2620562 0.06501609
#> 62  0.4347268 0.3669184 0.19835480
#> 63  0.5744226 0.3001170 0.12546044
#> 64  0.5237562 0.2611711 0.21507273
#> 65  0.6912442 0.2005880 0.10816784
#> 66  0.6650554 0.2229836 0.11196100
#> 67  0.4020779 0.3566599 0.24126220
#> 68  0.5519384 0.3194774 0.12858419
#> 69  0.5260983 0.3396017 0.13430008
#> 70  0.6675553 0.2633396 0.06910510
#> 71  0.5601094 0.3048540 0.13503663
#> 72  0.5564856 0.3592798 0.08423454
#> 73  0.5287788 0.4110783 0.06014282
#> 74  0.6486472 0.2364738 0.11487904
#> 75  0.4678768 0.3593994 0.17272372
#> 76  0.5927905 0.2941099 0.11309966
#> 77  0.5508232 0.4240786 0.02509821
#> 78  0.6133119 0.2619006 0.12478753
#> 79  0.6124147 0.3068618 0.08072355
#> 80  0.5474919 0.3792252 0.07328289
#> 81  0.5705519 0.3622783 0.06716982
#> 82  0.6044398 0.2498959 0.14566426
#> 83  0.6918889 0.2303649 0.07774623
#> 84  0.4783690 0.4140291 0.10760188
#> 85  0.6462155 0.2647412 0.08904323
#> 86  0.5898398 0.2650106 0.14514958
#> 87  0.7098548 0.1768099 0.11333531
#> 88  0.5421816 0.3545883 0.10323011
#> 89  0.6513334 0.2059825 0.14268412
#> 90  0.4883665 0.4057771 0.10585636
#> 91  0.6768758 0.1997884 0.12333580
#> 92  0.6947548 0.1810449 0.12420030
#> 93  0.6982414 0.1628471 0.13891149
#> 94  0.7024853 0.1942352 0.10327954
#> 95  0.6176680 0.2908024 0.09152965
#> 96  0.7005409 0.2250786 0.07438048
#> 97  0.6245288 0.2679843 0.10748691
#> 98  0.5348477 0.3458675 0.11928475
#> 99  0.5750313 0.2921910 0.13277772
#> 100 0.4798282 0.2999794 0.22019236
#> 101 0.5336212 0.2896688 0.17670993
#> 102 0.5615248 0.2948413 0.14363392
#> 103 0.5085485 0.3221887 0.16926276
#> 104 0.4714882 0.4228305 0.10568126
#> 105 0.6167187 0.2035632 0.17971813
#> 106 0.5066099 0.2743673 0.21902280
#> 107 0.4805937 0.3351720 0.18423428
#> 108 0.5237136 0.3799055 0.09638081
#> 109 0.6195708 0.3412218 0.03920743
#> 110 0.5207832 0.2738899 0.20532688
#> 111 0.5921291 0.2372473 0.17062362
#> 112 0.5174046 0.3440286 0.13856681
#> 113 0.6456230 0.3144130 0.03996408
#> 114 0.5070986 0.3329267 0.15997467
#> 115 0.7185082 0.2289065 0.05258528
#> 116 0.5860172 0.3419861 0.07199670
#> 117 0.5526657 0.2898508 0.15748345
#> 118 0.6985638 0.2371414 0.06429487
#> 119 0.4251508 0.3542095 0.22063964
#> 120 0.4868144 0.3191095 0.19407618
#> 121 0.6440608 0.1964217 0.15951750
#> 122 0.5947950 0.2175764 0.18762862
#> 123 0.5759517 0.2649728 0.15907549
#> 124 0.5703532 0.2482191 0.18142764
#> 125 0.5753274 0.3056444 0.11902819
#> 126 0.5743465 0.2702418 0.15541165
#> 127 0.4523607 0.3821156 0.16552369
#> 128 0.5134388 0.2974923 0.18906891
#> 129 0.4781044 0.3058751 0.21602055
#> 130 0.4600666 0.3300544 0.20987895
#> 131 0.5666297 0.3199016 0.11346864
#> 132 0.6982414 0.1628471 0.13891149
#> 133 0.4847123 0.3009154 0.21437236
#> 134 0.6077673 0.2947626 0.09747006
#> 135 0.6205533 0.2246544 0.15479226
#> 136 0.7172449 0.2271667 0.05558835
#> 137 0.5394324 0.3525460 0.10802162
#> 138 0.7665027 0.1588897 0.07460760
#> 139 0.5813098 0.3156904 0.10299988
#> 140 0.4960488 0.3397688 0.16418239
#> 141 0.4728813 0.3471258 0.17999288
#> 142 0.5469119 0.2786978 0.17439028
#> 143 0.4208884 0.3504618 0.22864980
#> 144 0.6670376 0.2151532 0.11780921
#> 145 0.6120844 0.2625457 0.12536997
#> 146 0.4907094 0.3872609 0.12202969
#> 147 0.5091225 0.3233334 0.16754404
#> 148 0.6128457 0.3058803 0.08127400
#> 149 0.5799275 0.3088041 0.11126841
#> 150 0.4960488 0.3397688 0.16418239
#> 151 0.6278764 0.3069785 0.06514509
#> 152 0.6345302 0.2781083 0.08736153
#> 153 0.4400588 0.3110779 0.24886325
#> 154 0.5231069 0.3985939 0.07829915
#> 155 0.5457508 0.3406116 0.11363760
#> 156 0.7889775 0.1512933 0.05972917
#> 157 0.5715191 0.3477634 0.08071756
#> 158 0.6467570 0.2806801 0.07256293
#> 159 0.5758407 0.2440306 0.18012866
#> 160 0.4681145 0.3777510 0.15413449
#> 161 0.5872133 0.3395672 0.07321945
#> 162 0.4960224 0.3947023 0.10927530
#> 163 0.6812105 0.2288217 0.08996777
#> 164 0.6812105 0.2288217 0.08996777
#> 165 0.7657226 0.1672503 0.06702709
#> 166 0.4418495 0.3601760 0.19797450
#> 167 0.6582614 0.2119712 0.12976742
#> 168 0.6731850 0.2688273 0.05798773
#> 169 0.5066099 0.2743673 0.21902280
#> 170 0.4649348 0.2964884 0.23857678
#> 171 0.5457508 0.3406116 0.11363760
#> 172 0.5487988 0.3212180 0.12998316
#> 173 0.5398198 0.3480097 0.11217046
#> 174 0.5464604 0.2938685 0.15967112
#> 175 0.6313442 0.2825557 0.08610010
#> 176 0.7253454 0.2090744 0.06558018
#> 177 0.7889775 0.1512933 0.05972917
#> 178 0.6370715 0.2154058 0.14752262
#> 179 0.4373607 0.3484701 0.21416913
#> 180 0.5193427 0.3136500 0.16700729
#> 181 0.6460176 0.2385519 0.11543051
#> 182 0.6552730 0.2813102 0.06341671
#> 183 0.4907696 0.3219586 0.18727187
#> 184 0.4817649 0.3566997 0.16153540
#> 185 0.5048040 0.3356357 0.15956024
#> 186 0.4647608 0.3329162 0.20232304
#> 187 0.6475576 0.2706109 0.08183157
#> 188 0.4886490 0.3753655 0.13598549
#> 189 0.3917631 0.3635491 0.24468788
#> 190 0.4811958 0.3074748 0.21132938
#> 191 0.6197722 0.2878510 0.09237688
#> 192 0.4886149 0.3444930 0.16689205
#> 193 0.5588226 0.3446128 0.09656461
#> 194 0.6903690 0.1705331 0.13909786
#> 195 0.6370715 0.2154058 0.14752262
#> 196 0.5750991 0.3793001 0.04560080
#> 197 0.5336336 0.3686896 0.09767683
#> 198 0.4808194 0.3616394 0.15754114
#> 199 0.5160677 0.3268263 0.15710601
#> 200 0.6427847 0.3117766 0.04543862
#> 

#Using radius search method with two different radius
radius_test <- c(3, 4)
geometry_features(example, method = "radius_search", radius = radius_test, max_neighbour = 200)
#> , , radius_3
#> 
#>     npoints      eig1      eig2       eig3
#> 1        20 0.5247166 0.3081362 0.16714723
#> 2        10 0.6582298 0.3168572 0.02491298
#> 3        17 0.4205692 0.3447819 0.23464893
#> 4        16 0.5330090 0.3332326 0.13375841
#> 5        21 0.5644997 0.2882047 0.14729565
#> 6        19 0.5025048 0.3174027 0.18009249
#> 7        16 0.4829904 0.3608896 0.15611998
#> 8        19 0.5023141 0.3224023 0.17528359
#> 9        14 0.7286753 0.2081047 0.06321999
#> 10       19 0.5482695 0.3095373 0.14219318
#> 11       20 0.4798028 0.3662029 0.15399431
#> 12       16 0.4496611 0.3953245 0.15501439
#> 13       12 0.6259418 0.2775526 0.09650568
#> 14       10 0.5222233 0.3387871 0.13898959
#> 15        9 0.5169320 0.2885323 0.19453574
#> 16        9 0.6216338 0.3008981 0.07746810
#> 17       16 0.6211729 0.2760108 0.10281632
#> 18       11 0.4491167 0.3615522 0.18933114
#> 19       25 0.4262648 0.3467242 0.22701099
#> 20       19 0.5746102 0.2889160 0.13647381
#> 21       12 0.5388499 0.3607558 0.10039425
#> 22       14 0.4622169 0.3530682 0.18471490
#> 23       22 0.4193739 0.3463410 0.23428514
#> 24       19 0.4891882 0.4059386 0.10487315
#> 25       18 0.5033250 0.3506168 0.14605824
#> 26       18 0.5135460 0.2872182 0.19923575
#> 27       14 0.6439632 0.2346837 0.12135318
#> 28       20 0.4753782 0.3065317 0.21809007
#> 29       19 0.6107866 0.2450519 0.14416152
#> 30       12 0.5842255 0.3520314 0.06374308
#> 31       12 0.7192080 0.1634389 0.11735309
#> 32       20 0.3889690 0.3648491 0.24618191
#> 33       23 0.4685191 0.2789801 0.25250073
#> 34       14 0.4590899 0.2960717 0.24483836
#> 35       21 0.5296998 0.3056586 0.16464159
#> 36       16 0.4650389 0.3793250 0.15563609
#> 37       19 0.4767559 0.3025294 0.22071473
#> 38       13 0.5769724 0.3391276 0.08389994
#> 39       13 0.5075083 0.3782500 0.11424171
#> 40        4 0.7685335 0.1942846 0.03718183
#> 41       21 0.3823375 0.3447445 0.27291794
#> 42       13 0.5993577 0.2263024 0.17433982
#> 43        9 0.7691591 0.1726580 0.05818293
#> 44       14 0.5122823 0.3139843 0.17373342
#> 45       16 0.4972183 0.3632363 0.13954540
#> 46       14 0.6403568 0.2065521 0.15309116
#> 47       27 0.4527831 0.3600671 0.18714985
#> 48       20 0.4275658 0.3442809 0.22815332
#> 49       24 0.4704790 0.2999985 0.22952250
#> 50       16 0.5845021 0.2887137 0.12678420
#> 51       23 0.5208334 0.2782525 0.20091406
#> 52       15 0.4401880 0.3744648 0.18534718
#> 53       23 0.4775519 0.3197049 0.20274324
#> 54       14 0.6862397 0.1823775 0.13138281
#> 55       15 0.6097043 0.2059542 0.18434151
#> 56       18 0.4250859 0.3557824 0.21913166
#> 57       16 0.4761115 0.3314447 0.19244378
#> 58       14 0.4404994 0.3776069 0.18189368
#> 59       26 0.4815344 0.3215460 0.19691954
#> 60       20 0.4366260 0.2943618 0.26901229
#> 61       13 0.7183427 0.2353519 0.04630546
#> 62       12 0.5617846 0.2488099 0.18940547
#> 63        8 0.4852204 0.3545106 0.16026901
#> 64       11 0.5379484 0.2643878 0.19766388
#> 65       17 0.6210862 0.2539434 0.12497044
#> 66       15 0.4768085 0.3772162 0.14597531
#> 67       19 0.4857886 0.3163387 0.19787276
#> 68       11 0.4736501 0.4161651 0.11018475
#> 69       23 0.4025205 0.3107004 0.28677911
#> 70       15 0.5195568 0.3591180 0.12132524
#> 71       23 0.4750140 0.3118366 0.21314935
#> 72       25 0.4836441 0.3144929 0.20186294
#> 73       21 0.4092458 0.3370316 0.25372256
#> 74       15 0.6190070 0.2838286 0.09716434
#> 75       14 0.5450300 0.2711145 0.18385552
#> 76       18 0.5312864 0.2818621 0.18685153
#> 77       16 0.5657918 0.3983774 0.03583075
#> 78       10 0.6133119 0.2619006 0.12478753
#> 79       15 0.5109609 0.3504098 0.13862933
#> 80       10 0.5474919 0.3792252 0.07328289
#> 81       17 0.6205445 0.2747336 0.10472186
#> 82       18 0.5175563 0.3290806 0.15336305
#> 83       14 0.4809603 0.3512110 0.16782870
#> 84       14 0.5557275 0.3000134 0.14425911
#> 85       12 0.5215196 0.2863281 0.19215229
#> 86       10 0.5898398 0.2650106 0.14514958
#> 87       12 0.7458257 0.1503380 0.10383631
#> 88       19 0.4720808 0.3401603 0.18775886
#> 89       14 0.6337146 0.2028275 0.16345784
#> 90       17 0.5419785 0.3266466 0.13137485
#> 91       20 0.5443532 0.3073978 0.14824908
#> 92       17 0.6347781 0.2228795 0.14234242
#> 93       14 0.6330051 0.2489890 0.11800586
#> 94       11 0.6694981 0.2376348 0.09286707
#> 95       10 0.6176680 0.2908024 0.09152965
#> 96       14 0.5119402 0.3150960 0.17296378
#> 97       11 0.6492516 0.2294490 0.12129942
#> 98       20 0.5159886 0.2949136 0.18909778
#> 99       22 0.4621695 0.3508844 0.18694611
#> 100      19 0.4317744 0.3165839 0.25164174
#> 101      13 0.6286546 0.2496844 0.12166100
#> 102      23 0.5252719 0.2591078 0.21562029
#> 103      11 0.5704605 0.2589923 0.17054723
#> 104      14 0.5683274 0.3579525 0.07372019
#> 105      12 0.5958926 0.2904112 0.11369624
#> 106      16 0.4802363 0.3023001 0.21746365
#> 107      21 0.4773397 0.3221093 0.20055091
#> 108      12 0.6460033 0.2664689 0.08752774
#> 109      17 0.6422992 0.2805062 0.07719459
#> 110      21 0.4253871 0.3323768 0.24223610
#> 111      23 0.4023130 0.3282853 0.26940177
#> 112      21 0.4424189 0.3792764 0.17830472
#> 113      16 0.6576650 0.2155160 0.12681901
#> 114      20 0.4513217 0.3608972 0.18778108
#> 115      16 0.5685626 0.2722475 0.15918991
#> 116       6 0.7757119 0.1315594 0.09272867
#> 117       7 0.4635933 0.3040996 0.23230709
#> 118      10 0.6985638 0.2371414 0.06429487
#> 119      18 0.5042642 0.2657137 0.23002210
#> 120      22 0.4786522 0.2972847 0.22406307
#> 121      12 0.5675883 0.2994154 0.13299623
#> 122      15 0.5205697 0.3190637 0.16036656
#> 123      12 0.5737778 0.2815809 0.14464137
#> 124      13 0.4538911 0.3594166 0.18669236
#> 125      13 0.5639963 0.3063311 0.12967264
#> 126      19 0.3919129 0.3169552 0.29113185
#> 127      21 0.3955159 0.3557914 0.24869268
#> 128      24 0.4220795 0.3145801 0.26334048
#> 129      16 0.4962695 0.3423332 0.16139722
#> 130      12 0.5884683 0.2713528 0.14017899
#> 131       8 0.4822051 0.3674283 0.15036659
#> 132      14 0.5920561 0.2728446 0.13509933
#> 133      23 0.4579428 0.3369529 0.20510431
#> 134      10 0.6077673 0.2947626 0.09747006
#> 135      15 0.4977364 0.3180902 0.18417341
#> 136      15 0.7007109 0.2352468 0.06404235
#> 137      14 0.5553794 0.3706172 0.07400344
#> 138      15 0.6087340 0.2906567 0.10060926
#> 139      19 0.5746102 0.2889160 0.13647381
#> 140      16 0.4513476 0.2963112 0.25234117
#> 141       8 0.6135877 0.3402698 0.04614246
#> 142       8 0.7024045 0.1563997 0.14119583
#> 143       7 0.5251267 0.4053457 0.06952759
#> 144      18 0.4359390 0.3419453 0.22211571
#> 145      20 0.3657157 0.3322527 0.30203156
#> 146      13 0.4725871 0.3831712 0.14424166
#> 147      13 0.6324043 0.2221694 0.14542637
#> 148      17 0.4424313 0.3853035 0.17226522
#> 149      15 0.5193387 0.2840436 0.19661774
#> 150      16 0.5472435 0.2693366 0.18341989
#> 151      10 0.6278764 0.3069785 0.06514509
#> 152      11 0.6452461 0.2606069 0.09414698
#> 153      19 0.5080661 0.2827579 0.20917598
#> 154      13 0.5823402 0.3694447 0.04821501
#> 155       8 0.5171513 0.3624928 0.12035590
#> 156      14 0.5430834 0.3161966 0.14072004
#> 157      17 0.5865236 0.2674994 0.14597706
#> 158      16 0.7645939 0.1828591 0.05254700
#> 159      15 0.5469348 0.2855736 0.16749156
#> 160      12 0.6017869 0.2950524 0.10316072
#> 161      18 0.5945152 0.2545879 0.15089690
#> 162      15 0.5967241 0.2559803 0.14729569
#> 163      23 0.5326623 0.2709934 0.19634432
#> 164      20 0.5739305 0.2608125 0.16525698
#> 165      10 0.7657226 0.1672503 0.06702709
#> 166       9 0.4977209 0.3619653 0.14031384
#> 167      14 0.5016295 0.3588482 0.13952229
#> 168      11 0.6432394 0.3074316 0.04932904
#> 169      20 0.4835143 0.2808023 0.23568339
#> 170      25 0.4425639 0.3675096 0.18992650
#> 171       8 0.5171513 0.3624928 0.12035590
#> 172      19 0.5033352 0.3053046 0.19136025
#> 173      13 0.5307193 0.2973861 0.17189458
#> 174      15 0.5870702 0.2895733 0.12335648
#> 175      11 0.5673907 0.2568259 0.17578335
#> 176      15 0.6550985 0.2135886 0.13131288
#> 177      16 0.6638809 0.2227892 0.11332995
#> 178      15 0.6625723 0.1801299 0.15729780
#> 179      21 0.4515402 0.3708454 0.17761441
#> 180      10 0.5193427 0.3136500 0.16700729
#> 181      15 0.6019720 0.2491204 0.14890762
#> 182      17 0.5261896 0.2964256 0.17738475
#> 183      23 0.5067909 0.3118152 0.18139389
#> 184      14 0.5906031 0.2649534 0.14444352
#> 185      28 0.4046045 0.3252003 0.27019518
#> 186      22 0.5669818 0.2297604 0.20325784
#> 187      19 0.4646794 0.3776094 0.15771125
#> 188       9 0.5689429 0.2820993 0.14895781
#> 189      16 0.4283764 0.3418809 0.22974268
#> 190      20 0.5035131 0.3394633 0.15702354
#> 191      28 0.4209820 0.3164315 0.26258648
#> 192      17 0.5092357 0.3091218 0.18164244
#> 193      18 0.5065730 0.3467003 0.14672673
#> 194      12 0.7657532 0.1261755 0.10807129
#> 195      13 0.6494167 0.1929986 0.15758470
#> 196      17 0.4828926 0.4407833 0.07632411
#> 197      22 0.4762211 0.3235620 0.20021691
#> 198      11 0.4656798 0.3677386 0.16658167
#> 199       8 0.4369239 0.4053475 0.15772858
#> 200      19 0.5468124 0.2611919 0.19199573
#> 
#> , , radius_4
#> 
#>     npoints      eig1      eig2       eig3
#> 1        31 0.5735882 0.2484879 0.17792393
#> 2        25 0.4749080 0.3253434 0.19974865
#> 3        36 0.4240135 0.3622002 0.21378626
#> 4        33 0.4835764 0.3814801 0.13494344
#> 5        35 0.4479237 0.3336675 0.21840877
#> 6        35 0.5226307 0.3350172 0.14235207
#> 7        23 0.4816119 0.3675725 0.15081555
#> 8        42 0.4740056 0.3114215 0.21457295
#> 9        34 0.5314924 0.2688741 0.19963347
#> 10       38 0.4123679 0.3046960 0.28293616
#> 11       47 0.4757678 0.3228090 0.20142322
#> 12       31 0.4990337 0.3568241 0.14414212
#> 13       28 0.5461599 0.3059193 0.14792078
#> 14       23 0.5036198 0.3680261 0.12835402
#> 15       20 0.5340663 0.3392250 0.12670872
#> 16       18 0.5645711 0.3049835 0.13044537
#> 17       28 0.4570023 0.3472861 0.19571163
#> 18       25 0.6470020 0.2398681 0.11312993
#> 19       42 0.5010656 0.2645239 0.23441053
#> 20       38 0.4855056 0.3574599 0.15703447
#> 21       27 0.5500312 0.3189085 0.13106023
#> 22       39 0.4942561 0.3392011 0.16654284
#> 23       48 0.4211798 0.2895933 0.28922688
#> 24       35 0.4565633 0.3691648 0.17427195
#> 25       31 0.4491831 0.3956771 0.15513983
#> 26       32 0.4854731 0.3318838 0.18264304
#> 27       43 0.4890041 0.3039759 0.20702002
#> 28       40 0.4392627 0.3608140 0.19992332
#> 29       43 0.4603802 0.3278711 0.21174864
#> 30       30 0.5227417 0.3654127 0.11184559
#> 31       30 0.4945448 0.3778552 0.12760001
#> 32       34 0.4812806 0.3018071 0.21691230
#> 33       42 0.4509545 0.3136576 0.23538790
#> 34       28 0.5643306 0.2497441 0.18592531
#> 35       53 0.4121495 0.3185087 0.26934183
#> 36       43 0.4226976 0.3857926 0.19150985
#> 37       39 0.4293994 0.3881486 0.18245198
#> 38       29 0.4861677 0.2622543 0.25157806
#> 39       21 0.4899901 0.3752010 0.13480894
#> 40       11 0.5493616 0.2987798 0.15185861
#> 41       34 0.4971999 0.3374113 0.16538878
#> 42       29 0.5180603 0.3373093 0.14463040
#> 43       20 0.6428997 0.2795801 0.07752017
#> 44       40 0.4434250 0.3764685 0.18010644
#> 45       34 0.4702753 0.4079018 0.12182285
#> 46       31 0.4730848 0.3818324 0.14508282
#> 47       46 0.4459791 0.2942636 0.25975730
#> 48       48 0.3525658 0.3359634 0.31147079
#> 49       42 0.4787712 0.3094891 0.21173965
#> 50       33 0.4798200 0.3652223 0.15495772
#> 51       51 0.3965420 0.3601605 0.24329750
#> 52       24 0.4925030 0.3687296 0.13876744
#> 53       44 0.4885911 0.3079925 0.20341634
#> 54       27 0.5556903 0.2900664 0.15424324
#> 55       34 0.5304894 0.2901132 0.17939742
#> 56       33 0.5014297 0.3436748 0.15489557
#> 57       38 0.4252323 0.3883207 0.18644697
#> 58       31 0.4965505 0.3837320 0.11971744
#> 59       41 0.4562889 0.3213400 0.22237109
#> 60       52 0.4055075 0.3199285 0.27456398
#> 61       21 0.5580660 0.3002120 0.14172196
#> 62       28 0.4684830 0.3795291 0.15198788
#> 63       14 0.4430283 0.4061204 0.15085130
#> 64       32 0.5087284 0.3518805 0.13939113
#> 65       32 0.4497519 0.3305670 0.21968107
#> 66       20 0.4735074 0.3372769 0.18921564
#> 67       49 0.4906610 0.2810833 0.22825563
#> 68       19 0.4626370 0.3142451 0.22311797
#> 69       39 0.4587390 0.3494234 0.19183763
#> 70       36 0.5676598 0.3057531 0.12658702
#> 71       43 0.4967414 0.2839544 0.21930416
#> 72       35 0.4264266 0.3019108 0.27166264
#> 73       43 0.4847716 0.2907569 0.22447146
#> 74       26 0.5221701 0.3524530 0.12537683
#> 75       28 0.5505921 0.2705967 0.17881122
#> 76       46 0.4237170 0.3415847 0.23469835
#> 77       31 0.4649912 0.4023003 0.13270852
#> 78       30 0.5360913 0.3255328 0.13837592
#> 79       42 0.4173357 0.3193946 0.26326965
#> 80       30 0.4879529 0.3841740 0.12787303
#> 81       33 0.4753491 0.3352904 0.18936049
#> 82       38 0.4726308 0.3870882 0.14028098
#> 83       36 0.4878841 0.3593121 0.15280374
#> 84       26 0.5912428 0.3012385 0.10751875
#> 85       28 0.5587897 0.3025326 0.13867766
#> 86       24 0.4962972 0.3693365 0.13436640
#> 87       28 0.4849370 0.4097155 0.10534746
#> 88       31 0.4505072 0.3649169 0.18457599
#> 89       29 0.4490679 0.3563928 0.19453932
#> 90       46 0.4478021 0.2826778 0.26952011
#> 91       39 0.5439846 0.3259672 0.13004814
#> 92       28 0.5916560 0.2621474 0.14619658
#> 93       25 0.5789672 0.3002630 0.12076981
#> 94       17 0.6260530 0.2072135 0.16673349
#> 95       29 0.5617280 0.3259253 0.11234675
#> 96       32 0.4765404 0.3087592 0.21470042
#> 97       31 0.4467435 0.4155297 0.13772686
#> 98       42 0.4573503 0.3359873 0.20666241
#> 99       43 0.5373641 0.2729186 0.18971727
#> 100      35 0.4439149 0.3226327 0.23345244
#> 101      31 0.6052644 0.3025137 0.09222187
#> 102      44 0.4529817 0.2985935 0.24842475
#> 103      25 0.6442318 0.2378689 0.11789926
#> 104      32 0.5697733 0.2591517 0.17107499
#> 105      27 0.5300870 0.3332232 0.13668985
#> 106      29 0.6069988 0.2329561 0.16004512
#> 107      43 0.4616872 0.3303715 0.20794130
#> 108      18 0.5565217 0.2856719 0.15780638
#> 109      32 0.4187943 0.3317012 0.24950446
#> 110      55 0.4335882 0.3458397 0.22057205
#> 111      48 0.4320847 0.3463114 0.22160384
#> 112      51 0.4089873 0.3687092 0.22230349
#> 113      51 0.3882672 0.3522557 0.25947710
#> 114      47 0.4924421 0.2998852 0.20767276
#> 115      32 0.4299558 0.2931454 0.27689879
#> 116      11 0.5245036 0.2901672 0.18532914
#> 117      16 0.4485320 0.3867385 0.16472942
#> 118      25 0.5538311 0.2854144 0.16075449
#> 119      39 0.4163869 0.3756072 0.20800588
#> 120      37 0.4863423 0.2822525 0.23140529
#> 121      24 0.4945439 0.3438590 0.16159708
#> 122      29 0.4492031 0.2903134 0.26048355
#> 123      21 0.5197932 0.3451441 0.13506265
#> 124      23 0.5358730 0.3483665 0.11576052
#> 125      25 0.5149726 0.3387257 0.14630172
#> 126      47 0.4131416 0.3583090 0.22854944
#> 127      44 0.4114824 0.3483451 0.24017246
#> 128      50 0.4082092 0.3841481 0.20764267
#> 129      37 0.4573035 0.3719554 0.17074108
#> 130      29 0.4474984 0.3000099 0.25249170
#> 131      28 0.5562299 0.3078392 0.13593087
#> 132      31 0.3912090 0.3211526 0.28763841
#> 133      35 0.5137406 0.2537621 0.23249734
#> 134      16 0.5761057 0.2255817 0.19831251
#> 135      30 0.5532063 0.3112783 0.13551533
#> 136      31 0.4160720 0.3692485 0.21467955
#> 137      24 0.5347979 0.2972161 0.16798600
#> 138      28 0.4719111 0.3941680 0.13392094
#> 139      36 0.5393073 0.3100624 0.15063032
#> 140      42 0.4263596 0.3518339 0.22180650
#> 141      14 0.5080238 0.2887222 0.20325399
#> 142      21 0.5191206 0.3429872 0.13789220
#> 143      13 0.4883417 0.3259769 0.18568136
#> 144      37 0.4266768 0.4083518 0.16497143
#> 145      48 0.4007595 0.3453563 0.25388421
#> 146      28 0.6163285 0.2314767 0.15219484
#> 147      33 0.4654675 0.4008308 0.13370177
#> 148      28 0.5132319 0.2903935 0.19637462
#> 149      31 0.4928034 0.4181569 0.08903964
#> 150      41 0.4634620 0.3451075 0.19143052
#> 151      19 0.5227379 0.2640371 0.21322500
#> 152      16 0.4838374 0.3065326 0.20963002
#> 153      51 0.4073871 0.3591557 0.23345718
#> 154      20 0.4744061 0.3485083 0.17708560
#> 155      19 0.5320808 0.3057100 0.16220921
#> 156      26 0.6618532 0.1867926 0.15135426
#> 157      29 0.5580989 0.3212238 0.12067730
#> 158      27 0.5753379 0.2364987 0.18816340
#> 159      35 0.5479943 0.2920519 0.15995373
#> 160      28 0.5306376 0.3135915 0.15577090
#> 161      34 0.5547029 0.2918370 0.15346013
#> 162      30 0.5254536 0.3180108 0.15653565
#> 163      50 0.3938724 0.3441334 0.26199417
#> 164      52 0.3762140 0.3656404 0.25814558
#> 165      32 0.4716193 0.3433990 0.18498167
#> 166      16 0.6815497 0.1911455 0.12730471
#> 167      37 0.4633665 0.4110261 0.12560732
#> 168      27 0.5037178 0.3561494 0.14013281
#> 169      27 0.5979766 0.2449315 0.15709198
#> 170      42 0.4607685 0.3581669 0.18106454
#> 171      25 0.5228327 0.3061270 0.17104033
#> 172      43 0.4536076 0.3400645 0.20632786
#> 173      24 0.5553061 0.2806331 0.16406077
#> 174      31 0.5960339 0.2734814 0.13048471
#> 175      28 0.5132036 0.3616665 0.12512996
#> 176      47 0.4152975 0.3665464 0.21815612
#> 177      30 0.6414190 0.2229003 0.13568072
#> 178      25 0.5092700 0.3616012 0.12912877
#> 179      40 0.4735792 0.3389085 0.18751230
#> 180      23 0.4684870 0.3951914 0.13632158
#> 181      27 0.4773098 0.3045920 0.21809820
#> 182      37 0.4579823 0.3401746 0.20184310
#> 183      44 0.4942195 0.3466558 0.15912471
#> 184      27 0.5355505 0.2680771 0.19637237
#> 185      50 0.3757900 0.3264001 0.29780996
#> 186      54 0.4224036 0.3248555 0.25274090
#> 187      32 0.4324302 0.4146037 0.15296611
#> 188      20 0.5335273 0.3122139 0.15425881
#> 189      36 0.4959588 0.3536718 0.15036934
#> 190      35 0.4619466 0.3098686 0.22818482
#> 191      47 0.4394530 0.3539526 0.20659446
#> 192      28 0.5136481 0.3320631 0.15428882
#> 193      35 0.4644828 0.3395704 0.19594684
#> 194      18 0.5781457 0.2216941 0.20016026
#> 195      24 0.5352899 0.3248692 0.13984098
#> 196      27 0.4858159 0.3379641 0.17621999
#> 197      47 0.4392794 0.3335338 0.22718673
#> 198      30 0.5098048 0.3534611 0.13673406
#> 199      17 0.5447454 0.3133964 0.14185824
#> 200      35 0.4359880 0.3220706 0.24194149
#>