Geometry features of Neighboring Points.

Estimate geometry features of neighboring points in a cloud.

geometry_features(
  cloud,
  method,
  radius,
  k,
  max_neighbour,
  target = FALSE,
  build = "kdtree",
  threads = 1L,
  checks = 1L,
  progress = TRUE
)

Arguments

cloud
A data.table with *XYZ* coordinates in the first three columns.

method
A character string specifying the method to estimate the neighbors. It most be one of "radius_search" or "knn".

radius
A numeric vector representing the radius for search to consider. This needs be used if method = "radius_search".

k
An integer vector representing the number of neighbors to consider. This needs be used if method = "knn".

max_neighbour
An integer specifying the maximum number of points to look around each query point for a given radius. This needs be used if method = "radius_search".

target
Logic. If TRUE, it consider that target point for the calculations of geometry features.

build
A character describing the search structure to be used: "kdtree", "kmeans", "linear".

threads
An integer specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware.

checks
Number of checks during searching. Higher value gives better search precision but takes longer. See FLANN C++ manual for more details.

progress
Show progress. TRUE as default.

Value

A array describing the point of the cloud in rows, the relative eigenvalues in columns, and the radius or k per slide. If method = "radius_search", it add in the first column the number of neighboring points.

Details

The function returns the geometry features of the neighboring points of a given point in cloud. Geometry features are represented by the relative values of the eigenvalues derived from a covariance matrix of the neighboring points. Geometry features are not estimated on target points with less than 3 neighboring points.

Author

J. Antonio Guzmán Q.

Examples

#Create cloud
example <- data.table(X = runif(200, min=0, max=10),
                      Y = runif(200, min=0, max=10),
                      Z = runif(200, min=0, max=10))

#Using knn method with two different k
k_test <- c(5, 10)
geometry_features(example, method = "knn", k = k_test)

#> , , k_5
#> 
#>          eig1       eig2         eig3
#> 1   0.4790217 0.44915002 0.0718282624
#> 2   0.6965881 0.29499400 0.0084178808
#> 3   0.8406839 0.15637601 0.0029400969
#> 4   0.6942528 0.30375218 0.0019950524
#> 5   0.5548494 0.43648688 0.0086637265
#> 6   0.6215036 0.20832798 0.1701684205
#> 7   0.7440442 0.22999141 0.0259644342
#> 8   0.8875413 0.09404938 0.0184093207
#> 9   0.7158584 0.24309297 0.0410486296
#> 10  0.6892934 0.25806895 0.0526376200
#> 11  0.6231786 0.27297686 0.1038445459
#> 12  0.4724720 0.37956331 0.1479646462
#> 13  0.7606817 0.17366993 0.0656484182
#> 14  0.7925793 0.16912050 0.0383002411
#> 15  0.6462667 0.32295316 0.0307801840
#> 16  0.6691792 0.19748200 0.1333388359
#> 17  0.5960616 0.29161258 0.1123258369
#> 18  0.5839795 0.33137894 0.0846415371
#> 19  0.8925082 0.09380789 0.0136838713
#> 20  0.7627834 0.16380583 0.0734107619
#> 21  0.6965881 0.29499400 0.0084178808
#> 22  0.6916896 0.22212526 0.0861851105
#> 23  0.5605969 0.31872125 0.1206818828
#> 24  0.5536289 0.39732454 0.0490465292
#> 25  0.6135753 0.31426182 0.0721628627
#> 26  0.6229094 0.34412334 0.0329672719
#> 27  0.8320269 0.13898250 0.0289905648
#> 28  0.6456891 0.28212945 0.0721814272
#> 29  0.5360791 0.38190466 0.0820162393
#> 30  0.6220592 0.36500504 0.0129357753
#> 31  0.4905281 0.43937425 0.0700976208
#> 32  0.8034356 0.16035026 0.0362140965
#> 33  0.6991280 0.27300150 0.0278704646
#> 34  0.6954595 0.24469237 0.0598481539
#> 35  0.6064448 0.30247659 0.0910786618
#> 36  0.5842394 0.33836836 0.0773922449
#> 37  0.5137735 0.35733929 0.1288871976
#> 38  0.7098419 0.23809702 0.0520610614
#> 39  0.7066957 0.26590346 0.0274008630
#> 40  0.6109232 0.35507435 0.0340024657
#> 41  0.8070409 0.17260390 0.0203551767
#> 42  0.6892934 0.25806895 0.0526376200
#> 43  0.5696637 0.40556608 0.0247702095
#> 44  0.6366413 0.31382518 0.0495334969
#> 45  0.8309702 0.10229035 0.0667394367
#> 46  0.6205739 0.37878448 0.0006416227
#> 47  0.4954891 0.33303436 0.1714765836
#> 48  0.5798636 0.37229317 0.0478432467
#> 49  0.6484641 0.20422592 0.1473099394
#> 50  0.5340196 0.35137000 0.1146103606
#> 51  0.7936528 0.19020911 0.0161380727
#> 52  0.7039913 0.29249699 0.0035116918
#> 53  0.5910241 0.39584102 0.0131348373
#> 54  0.6892934 0.25806895 0.0526376200
#> 55  0.5582235 0.35406513 0.0877113474
#> 56  0.5104378 0.39710300 0.0924591728
#> 57  0.5881973 0.36387858 0.0479241192
#> 58  0.6385769 0.25824168 0.1031813743
#> 59  0.5097063 0.35590952 0.1343841466
#> 60  0.7989863 0.17356493 0.0274487340
#> 61  0.5002052 0.41161567 0.0881790943
#> 62  0.7130236 0.17576478 0.1112115880
#> 63  0.9143279 0.08490108 0.0007710564
#> 64  0.7418607 0.23496310 0.0231761560
#> 65  0.5887016 0.28347598 0.1278224582
#> 66  0.4803899 0.40520264 0.1144074514
#> 67  0.8299015 0.16259232 0.0075062256
#> 68  0.7030027 0.29000640 0.0069909331
#> 69  0.9003475 0.08176126 0.0178912048
#> 70  0.6311003 0.31759042 0.0513092986
#> 71  0.4255586 0.40807400 0.1663674024
#> 72  0.6754898 0.24029391 0.0842163223
#> 73  0.7935190 0.14477246 0.0617085428
#> 74  0.7071271 0.24755722 0.0453156336
#> 75  0.5346000 0.38188516 0.0835148792
#> 76  0.5656727 0.39879813 0.0355291556
#> 77  0.5323414 0.36873461 0.0989239855
#> 78  0.6291945 0.32824054 0.0425649606
#> 79  0.5578501 0.30122175 0.1409281938
#> 80  0.8348532 0.15329016 0.0118566611
#> 81  0.6017863 0.37085214 0.0273615414
#> 82  0.8335132 0.14953482 0.0169520042
#> 83  0.7619088 0.21554064 0.0225506104
#> 84  0.7642671 0.20549809 0.0302347912
#> 85  0.5492024 0.31965040 0.1311472378
#> 86  0.6829494 0.31414807 0.0029025737
#> 87  0.7925793 0.16912050 0.0383002411
#> 88  0.6175282 0.21037013 0.1721016686
#> 89  0.7426020 0.21342891 0.0439691354
#> 90  0.6039344 0.29321197 0.1028535919
#> 91  0.5572734 0.27312638 0.1696002610
#> 92  0.8003880 0.18640073 0.0132112268
#> 93  0.6857772 0.28002340 0.0341994049
#> 94  0.6954595 0.24469237 0.0598481539
#> 95  0.7451530 0.22586058 0.0289864663
#> 96  0.6698757 0.30501178 0.0251125145
#> 97  0.7935190 0.14477246 0.0617085428
#> 98  0.6408310 0.27830228 0.0808667076
#> 99  0.5096704 0.37722873 0.1131008833
#> 100 0.8778416 0.08766632 0.0344920984
#> 101 0.5287811 0.42122524 0.0499936602
#> 102 0.5224557 0.29437722 0.1831670424
#> 103 0.7284117 0.21378267 0.0578056620
#> 104 0.8160740 0.18163987 0.0022861335
#> 105 0.8160740 0.18163987 0.0022861335
#> 106 0.6892934 0.25806895 0.0526376200
#> 107 0.6138562 0.26460139 0.1215424271
#> 108 0.6578895 0.27407102 0.0680394773
#> 109 0.7273583 0.23558178 0.0370598847
#> 110 0.7642671 0.20549809 0.0302347912
#> 111 0.7567924 0.23649622 0.0067113529
#> 112 0.7273583 0.23558178 0.0370598847
#> 113 0.7138192 0.28520531 0.0009754373
#> 114 0.6813828 0.24671245 0.0719047693
#> 115 0.6537274 0.31842164 0.0278509135
#> 116 0.7720518 0.16611008 0.0618380853
#> 117 0.7247039 0.23747518 0.0378209029
#> 118 0.5353062 0.40830318 0.0563905830
#> 119 0.5773456 0.28277339 0.1398809967
#> 120 0.5831642 0.29952823 0.1173075616
#> 121 0.5559179 0.28770370 0.1563783990
#> 122 0.8701864 0.12859316 0.0012204304
#> 123 0.7272217 0.27125641 0.0015218635
#> 124 0.5360791 0.38190466 0.0820162393
#> 125 0.5462514 0.29292748 0.1608211457
#> 126 0.5725745 0.33878176 0.0886437630
#> 127 0.6138562 0.26460139 0.1215424271
#> 128 0.5033831 0.35418885 0.1424280739
#> 129 0.5115316 0.46734554 0.0211228867
#> 130 0.5902923 0.33990031 0.0698073538
#> 131 0.7199105 0.25727167 0.0228178311
#> 132 0.6175282 0.21037013 0.1721016686
#> 133 0.6340034 0.33290481 0.0330918210
#> 134 0.7844127 0.20789079 0.0076965329
#> 135 0.5598287 0.31980306 0.1203682789
#> 136 0.6252287 0.33352471 0.0412466364
#> 137 0.5798636 0.37229317 0.0478432467
#> 138 0.7925172 0.11109063 0.0963921683
#> 139 0.5225841 0.31015831 0.1672575976
#> 140 0.6825717 0.17990548 0.1375227881
#> 141 0.6000693 0.38805314 0.0118775511
#> 142 0.7635668 0.17931828 0.0571148792
#> 143 0.8333971 0.15102955 0.0155733292
#> 144 0.8778416 0.08766632 0.0344920984
#> 145 0.5979369 0.28467925 0.1173838519
#> 146 0.4790452 0.29758541 0.2233693432
#> 147 0.6478679 0.27448428 0.0776477830
#> 148 0.6768505 0.29112494 0.0320245960
#> 149 0.8804208 0.08961830 0.0299609503
#> 150 0.6215036 0.20832798 0.1701684205
#> 151 0.6348475 0.34753590 0.0176165704
#> 152 0.5336650 0.46208385 0.0042511445
#> 153 0.7560014 0.16201604 0.0819826041
#> 154 0.5187471 0.45213014 0.0291227674
#> 155 0.5323414 0.36873461 0.0989239855
#> 156 0.8033986 0.11299729 0.0836041194
#> 157 0.6897168 0.24939276 0.0608904102
#> 158 0.8335132 0.14953482 0.0169520042
#> 159 0.6693287 0.32815865 0.0025126658
#> 160 0.7312443 0.25169640 0.0170592800
#> 161 0.8377001 0.12563699 0.0366629246
#> 162 0.7544519 0.19183317 0.0537149129
#> 163 0.7933408 0.15003832 0.0566209211
#> 164 0.7099913 0.25955888 0.0304498363
#> 165 0.7626842 0.17633439 0.0609813876
#> 166 0.6671434 0.25737691 0.0754797104
#> 167 0.6236773 0.20788511 0.1684376214
#> 168 0.8422584 0.11066023 0.0470813348
#> 169 0.6046616 0.29175159 0.1035868443
#> 170 0.5991550 0.32833360 0.0725113692
#> 171 0.6585020 0.23142063 0.1100773633
#> 172 0.8238702 0.15244159 0.0236882483
#> 173 0.5475491 0.34910004 0.1033508212
#> 174 0.7441320 0.21017334 0.0456946272
#> 175 0.6091905 0.22779297 0.1630165163
#> 176 0.6691792 0.19748200 0.1333388359
#> 177 0.5265974 0.45439262 0.0190099381
#> 178 0.7137294 0.23883352 0.0474371296
#> 179 0.8777789 0.11097233 0.0112487654
#> 180 0.5193503 0.37923932 0.1014103271
#> 181 0.5440994 0.45155413 0.0043464328
#> 182 0.6892934 0.25806895 0.0526376200
#> 183 0.6310356 0.29797500 0.0709893807
#> 184 0.8104341 0.13921339 0.0503524609
#> 185 0.5685589 0.39335695 0.0380841707
#> 186 0.7059588 0.28913339 0.0049078410
#> 187 0.6948333 0.27950328 0.0256634487
#> 188 0.8437839 0.10321591 0.0530002097
#> 189 0.8146010 0.14962096 0.0357780853
#> 190 0.8238702 0.15244159 0.0236882483
#> 191 0.7418607 0.23496310 0.0231761560
#> 192 0.6790985 0.31215855 0.0087429823
#> 193 0.7087050 0.23626152 0.0550335261
#> 194 0.6533202 0.34033298 0.0063468633
#> 195 0.5930888 0.34909085 0.0578202970
#> 196 0.6929984 0.29933818 0.0076634006
#> 197 0.8181084 0.17921695 0.0026747021
#> 198 0.5297569 0.45058185 0.0196612214
#> 199 0.8508534 0.12223679 0.0269098377
#> 200 0.6355382 0.36113583 0.0033260183
#> 
#> , , k_10
#> 
#>          eig1      eig2       eig3
#> 1   0.6754917 0.2247007 0.09980759
#> 2   0.4824976 0.3426604 0.17484204
#> 3   0.6313999 0.2291962 0.13940389
#> 4   0.5754943 0.3372818 0.08722391
#> 5   0.4960176 0.3190439 0.18493845
#> 6   0.7110546 0.2194718 0.06947363
#> 7   0.5355925 0.3034048 0.16100271
#> 8   0.4488793 0.3964825 0.15463823
#> 9   0.6161120 0.2267056 0.15718240
#> 10  0.5515603 0.3058851 0.14255453
#> 11  0.5849990 0.2339110 0.18109000
#> 12  0.6434711 0.2462494 0.11027955
#> 13  0.4739797 0.3973332 0.12868713
#> 14  0.6795547 0.2206685 0.09977687
#> 15  0.5321043 0.3838080 0.08408770
#> 16  0.5107192 0.3031839 0.18609694
#> 17  0.5104121 0.3615168 0.12807107
#> 18  0.5938081 0.3015222 0.10466975
#> 19  0.6092503 0.3049060 0.08584362
#> 20  0.5789677 0.3018264 0.11920587
#> 21  0.6510890 0.2836559 0.06525502
#> 22  0.7101249 0.1754351 0.11444007
#> 23  0.5825599 0.2927841 0.12465595
#> 24  0.5479596 0.3910963 0.06094407
#> 25  0.4815189 0.2810016 0.23747945
#> 26  0.5997581 0.3297156 0.07052627
#> 27  0.6960091 0.2583440 0.04564687
#> 28  0.5955846 0.3162283 0.08818715
#> 29  0.4376329 0.3800214 0.18234573
#> 30  0.5371111 0.3421462 0.12074279
#> 31  0.5827592 0.2353270 0.18191381
#> 32  0.5158917 0.3286959 0.15541249
#> 33  0.6799450 0.2731780 0.04687695
#> 34  0.4630811 0.4448373 0.09208164
#> 35  0.6551696 0.2482799 0.09655056
#> 36  0.6617396 0.2443689 0.09389143
#> 37  0.6827897 0.2027898 0.11442054
#> 38  0.6495196 0.2141935 0.13628697
#> 39  0.4943602 0.3524798 0.15316002
#> 40  0.7763290 0.1610831 0.06258791
#> 41  0.6176516 0.3248803 0.05746807
#> 42  0.5519353 0.2887453 0.15931939
#> 43  0.4848132 0.3413447 0.17384208
#> 44  0.6095327 0.3646316 0.02583571
#> 45  0.4417784 0.3398613 0.21836029
#> 46  0.6674035 0.2008629 0.13173360
#> 47  0.7235694 0.1630496 0.11338103
#> 48  0.5125645 0.3498791 0.13755641
#> 49  0.4924644 0.4022843 0.10525130
#> 50  0.4781626 0.3744425 0.14739490
#> 51  0.6778808 0.2395971 0.08252217
#> 52  0.5686240 0.3596192 0.07175685
#> 53  0.5401234 0.2489674 0.21090922
#> 54  0.5774787 0.3306747 0.09184665
#> 55  0.6515792 0.2195426 0.12887820
#> 56  0.5560559 0.3159468 0.12799733
#> 57  0.6289291 0.2573856 0.11368529
#> 58  0.5682364 0.2927681 0.13899556
#> 59  0.5825912 0.3113943 0.10601449
#> 60  0.7576108 0.1940655 0.04832370
#> 61  0.5180854 0.3263570 0.15555760
#> 62  0.6147567 0.2751946 0.11004873
#> 63  0.5124451 0.3719382 0.11561668
#> 64  0.6057438 0.2452616 0.14899465
#> 65  0.6025292 0.2385424 0.15892840
#> 66  0.5282943 0.2704436 0.20126215
#> 67  0.6949872 0.1700694 0.13494340
#> 68  0.4908724 0.3947523 0.11437528
#> 69  0.5108163 0.4046577 0.08452609
#> 70  0.5084062 0.3725692 0.11902463
#> 71  0.5129294 0.3356005 0.15147006
#> 72  0.5086343 0.4003214 0.09104433
#> 73  0.5640680 0.2885250 0.14740698
#> 74  0.5060891 0.3859565 0.10795435
#> 75  0.5572929 0.2846379 0.15806920
#> 76  0.4483130 0.3226047 0.22908228
#> 77  0.6188322 0.2590418 0.12212604
#> 78  0.5632916 0.2413093 0.19539911
#> 79  0.5808646 0.2739456 0.14518976
#> 80  0.4987899 0.3199094 0.18130077
#> 81  0.4094722 0.3619062 0.22862152
#> 82  0.5824026 0.3370308 0.08056667
#> 83  0.4691992 0.3805294 0.15027144
#> 84  0.5595704 0.2709654 0.16946420
#> 85  0.5179501 0.2980395 0.18401042
#> 86  0.6680484 0.2360056 0.09594607
#> 87  0.6795547 0.2206685 0.09977687
#> 88  0.5275713 0.2829152 0.18951346
#> 89  0.5988682 0.2554396 0.14569217
#> 90  0.6767506 0.1803151 0.14293429
#> 91  0.5528511 0.3169492 0.13019969
#> 92  0.6152544 0.2882724 0.09647321
#> 93  0.6183340 0.2547121 0.12695387
#> 94  0.4630811 0.4448373 0.09208164
#> 95  0.5621377 0.2777750 0.16008728
#> 96  0.6066928 0.2648691 0.12843808
#> 97  0.6081058 0.2728965 0.11899763
#> 98  0.5796708 0.3198818 0.10044741
#> 99  0.5395998 0.3450760 0.11532421
#> 100 0.5573949 0.3265964 0.11600872
#> 101 0.6136900 0.2761199 0.11019004
#> 102 0.6031810 0.2567108 0.14010822
#> 103 0.7190183 0.1985009 0.08248075
#> 104 0.4145598 0.3184990 0.26694116
#> 105 0.4551083 0.3101560 0.23473569
#> 106 0.6611319 0.2307064 0.10816166
#> 107 0.5096626 0.3660173 0.12432011
#> 108 0.5374365 0.2922453 0.17031821
#> 109 0.6001810 0.2924135 0.10740554
#> 110 0.5595704 0.2709654 0.16946420
#> 111 0.5152236 0.4598464 0.02493006
#> 112 0.6001810 0.2924135 0.10740554
#> 113 0.7432601 0.1879434 0.06879650
#> 114 0.5959628 0.2785139 0.12552326
#> 115 0.6925118 0.2239142 0.08357404
#> 116 0.5373280 0.3945593 0.06811266
#> 117 0.4977752 0.3197295 0.18249529
#> 118 0.4526533 0.3374300 0.20991671
#> 119 0.5732026 0.3189406 0.10785680
#> 120 0.6097230 0.3134463 0.07683063
#> 121 0.5599529 0.3011278 0.13891927
#> 122 0.4096617 0.3213454 0.26899291
#> 123 0.4729573 0.3844849 0.14255783
#> 124 0.6021792 0.2756485 0.12217229
#> 125 0.3930709 0.3572521 0.24967697
#> 126 0.4286704 0.3998675 0.17146214
#> 127 0.6538820 0.2105729 0.13554503
#> 128 0.7262507 0.1819049 0.09184440
#> 129 0.4434360 0.3524681 0.20409596
#> 130 0.5345547 0.3626982 0.10274712
#> 131 0.4495701 0.3018455 0.24858438
#> 132 0.6865002 0.2456275 0.06787222
#> 133 0.6997696 0.2191107 0.08111966
#> 134 0.5388993 0.4107218 0.05037889
#> 135 0.5966458 0.2740430 0.12931116
#> 136 0.6617259 0.2192331 0.11904091
#> 137 0.5515346 0.3301326 0.11833280
#> 138 0.4812196 0.3720127 0.14676771
#> 139 0.5305266 0.3861596 0.08331376
#> 140 0.5963509 0.2991417 0.10450740
#> 141 0.5437282 0.3131895 0.14308227
#> 142 0.4769196 0.3377261 0.18535434
#> 143 0.5673713 0.3259193 0.10670935
#> 144 0.5906591 0.3338393 0.07550164
#> 145 0.4431818 0.4082020 0.14861627
#> 146 0.4632452 0.3651263 0.17162854
#> 147 0.5932195 0.2861613 0.12061916
#> 148 0.6376631 0.2682262 0.09411064
#> 149 0.8059454 0.1151484 0.07890617
#> 150 0.6055196 0.3538277 0.04065275
#> 151 0.4306519 0.3681840 0.20116412
#> 152 0.4531234 0.3495654 0.19731113
#> 153 0.5566640 0.3165391 0.12679696
#> 154 0.4767715 0.3852247 0.13800379
#> 155 0.5452746 0.2932118 0.16151360
#> 156 0.5205697 0.3073923 0.17203800
#> 157 0.5849241 0.2795027 0.13557315
#> 158 0.5010102 0.3581092 0.14088066
#> 159 0.5034739 0.3502297 0.14629636
#> 160 0.4314948 0.3476972 0.22080795
#> 161 0.4708232 0.3608753 0.16830148
#> 162 0.4951022 0.2727034 0.23219438
#> 163 0.5874633 0.2831973 0.12933941
#> 164 0.5456650 0.2990692 0.15526575
#> 165 0.5368327 0.2555018 0.20766556
#> 166 0.6318228 0.2299426 0.13823454
#> 167 0.5941038 0.2700169 0.13587931
#> 168 0.6411988 0.2622820 0.09651914
#> 169 0.5650417 0.2946983 0.14026004
#> 170 0.4286704 0.3998675 0.17146214
#> 171 0.5283122 0.3013948 0.17029304
#> 172 0.5004700 0.3330660 0.16646402
#> 173 0.5921832 0.2309847 0.17683213
#> 174 0.6308632 0.2770626 0.09207423
#> 175 0.7383188 0.1794483 0.08223298
#> 176 0.4630560 0.3447681 0.19217587
#> 177 0.6049748 0.2064743 0.18855087
#> 178 0.5544159 0.3895807 0.05600332
#> 179 0.4879362 0.3846437 0.12742011
#> 180 0.4804270 0.2949109 0.22466203
#> 181 0.4996686 0.3663349 0.13399648
#> 182 0.5519353 0.2887453 0.15931939
#> 183 0.5921832 0.2309847 0.17683213
#> 184 0.6536009 0.2022587 0.14414037
#> 185 0.6100999 0.3225216 0.06737852
#> 186 0.6854478 0.2060059 0.10854637
#> 187 0.6326896 0.2176462 0.14966421
#> 188 0.4797067 0.3768800 0.14341328
#> 189 0.6954855 0.1899051 0.11460948
#> 190 0.4991174 0.3062092 0.19467331
#> 191 0.4699590 0.3440402 0.18600080
#> 192 0.6400193 0.2846361 0.07534463
#> 193 0.6330273 0.2147541 0.15221864
#> 194 0.5147039 0.3160721 0.16922403
#> 195 0.5794805 0.2701296 0.15038992
#> 196 0.5512751 0.2318916 0.21683336
#> 197 0.5352023 0.3252354 0.13956230
#> 198 0.5477870 0.3347659 0.11744715
#> 199 0.4449785 0.4000622 0.15495936
#> 200 0.5204887 0.3194541 0.16005716
#> 

#Using radius search method with two different radius
radius_test <- c(3, 4)
geometry_features(example, method = "radius_search", radius = radius_test, max_neighbour = 200)

#> , , radius_3
#> 
#>     npoints      eig1       eig2         eig3
#> 1         3       NaN        NaN          NaN
#> 2         3       NaN        NaN          NaN
#> 3         1       NaN        NaN          NaN
#> 4         5 0.6942528 0.30375218 0.0019950524
#> 5         7 0.6508165 0.30563016 0.0435533326
#> 6         4 0.5817195 0.31003977 0.1082407770
#> 7         1       NaN        NaN          NaN
#> 8         6 0.7350695 0.24748664 0.0174438886
#> 9         6 0.5046528 0.45910746 0.0362397544
#> 10        6 0.6878363 0.28537170 0.0267919793
#> 11        7 0.6543955 0.22599592 0.1196085599
#> 12        2       NaN        NaN          NaN
#> 13        3       NaN        NaN          NaN
#> 14        5 0.7925793 0.16912050 0.0383002411
#> 15        3       NaN        NaN          NaN
#> 16        6 0.6917138 0.25182488 0.0564613647
#> 17        3       NaN        NaN          NaN
#> 18        5 0.5839795 0.33137894 0.0846415371
#> 19        3       NaN        NaN          NaN
#> 20        6 0.6380604 0.29198374 0.0699558967
#> 21        5 0.6965881 0.29499400 0.0084178808
#> 22        2       NaN        NaN          NaN
#> 23        3       NaN        NaN          NaN
#> 24        2       NaN        NaN          NaN
#> 25        5 0.6135753 0.31426182 0.0721628627
#> 26        3       NaN        NaN          NaN
#> 27        6 0.5821421 0.39148676 0.0263711230
#> 28        2       NaN        NaN          NaN
#> 29        5 0.5360791 0.38190466 0.0820162393
#> 30        7 0.6314744 0.32627812 0.0422474430
#> 31        4 0.6279810 0.34413994 0.0278790565
#> 32        2       NaN        NaN          NaN
#> 33        2       NaN        NaN          NaN
#> 34        5 0.6954595 0.24469237 0.0598481539
#> 35        4 0.5431999 0.31126948 0.1455306457
#> 36        1       NaN        NaN          NaN
#> 37        3       NaN        NaN          NaN
#> 38        6 0.5046528 0.45910746 0.0362397544
#> 39        4 0.8764926 0.10812209 0.0153852777
#> 40        3       NaN        NaN          NaN
#> 41        5 0.8070409 0.17260390 0.0203551767
#> 42        6 0.6527159 0.26342426 0.0838598500
#> 43        7 0.5221186 0.38336991 0.0945114519
#> 44        9 0.6684813 0.29637967 0.0351390598
#> 45        6 0.6706138 0.26361447 0.0657716954
#> 46        3       NaN        NaN          NaN
#> 47        5 0.4954891 0.33303436 0.1714765836
#> 48        8 0.6147812 0.24891370 0.1363050804
#> 49        6 0.6166546 0.27787238 0.1054730704
#> 50        2       NaN        NaN          NaN
#> 51        5 0.7936528 0.19020911 0.0161380727
#> 52        3       NaN        NaN          NaN
#> 53        7 0.5741137 0.27427032 0.1516159500
#> 54        5 0.6892934 0.25806895 0.0526376200
#> 55        7 0.6647927 0.29014729 0.0450600551
#> 56        5 0.5104378 0.39710300 0.0924591728
#> 57        1       NaN        NaN          NaN
#> 58        3       NaN        NaN          NaN
#> 59        4 0.6626556 0.33055720 0.0067871901
#> 60        3       NaN        NaN          NaN
#> 61        3       NaN        NaN          NaN
#> 62        2       NaN        NaN          NaN
#> 63        6 0.7127303 0.25189340 0.0353763236
#> 64        6 0.6274503 0.23767094 0.1348787894
#> 65        4 0.7402555 0.25795032 0.0017941992
#> 66        4 0.7050148 0.28323504 0.0117501629
#> 67        8 0.5113946 0.27533704 0.2132683304
#> 68        5 0.7030027 0.29000640 0.0069909331
#> 69        6 0.7406715 0.24307047 0.0162580298
#> 70        7 0.5398713 0.32209798 0.1380306805
#> 71        3       NaN        NaN          NaN
#> 72        5 0.6754898 0.24029391 0.0842163223
#> 73        3       NaN        NaN          NaN
#> 74        4 0.6229617 0.34079045 0.0362478696
#> 75        6 0.4103368 0.36082649 0.2288367511
#> 76        5 0.5656727 0.39879813 0.0355291556
#> 77        4 0.6392793 0.25726930 0.1034514431
#> 78        2       NaN        NaN          NaN
#> 79        3       NaN        NaN          NaN
#> 80        3       NaN        NaN          NaN
#> 81        5 0.6017863 0.37085214 0.0273615414
#> 82        2       NaN        NaN          NaN
#> 83        9 0.5918837 0.32940200 0.0787142735
#> 84        4 0.7310781 0.25490815 0.0140137699
#> 85        7 0.4493961 0.32917990 0.2214239862
#> 86        6 0.7860543 0.19470706 0.0192386163
#> 87        3       NaN        NaN          NaN
#> 88        5 0.6175282 0.21037013 0.1721016686
#> 89        3       NaN        NaN          NaN
#> 90        4 0.7606951 0.17033320 0.0689716530
#> 91        1       NaN        NaN          NaN
#> 92        6 0.7932249 0.16524467 0.0415304077
#> 93        6 0.5306334 0.43061153 0.0387550780
#> 94        3       NaN        NaN          NaN
#> 95        7 0.5767691 0.33552906 0.0877018842
#> 96        7 0.7707917 0.18112305 0.0480852159
#> 97        3       NaN        NaN          NaN
#> 98        4 0.6401803 0.35851249 0.0013071957
#> 99        4 0.7115248 0.27417392 0.0143012869
#> 100       2       NaN        NaN          NaN
#> 101       5 0.5287811 0.42122524 0.0499936602
#> 102       2       NaN        NaN          NaN
#> 103       2       NaN        NaN          NaN
#> 104       5 0.8160740 0.18163987 0.0022861335
#> 105       5 0.8160740 0.18163987 0.0022861335
#> 106       6 0.7687876 0.21117996 0.0200324136
#> 107       9 0.6847639 0.21800226 0.0972338712
#> 108       5 0.6578895 0.27407102 0.0680394773
#> 109       6 0.5943478 0.27804484 0.1276073162
#> 110       3       NaN        NaN          NaN
#> 111       8 0.5670722 0.41476825 0.0181595103
#> 112       7 0.4945689 0.33821931 0.1672117765
#> 113       2       NaN        NaN          NaN
#> 114       8 0.4803103 0.35771433 0.1619753646
#> 115       6 0.5317410 0.25067443 0.2175845995
#> 116       1       NaN        NaN          NaN
#> 117       3       NaN        NaN          NaN
#> 118       5 0.5353062 0.40830318 0.0563905830
#> 119       2       NaN        NaN          NaN
#> 120       4 0.5679484 0.38333246 0.0487191821
#> 121       4 0.7546281 0.21900272 0.0263691545
#> 122       6 0.8002992 0.14837324 0.0513275854
#> 123       3       NaN        NaN          NaN
#> 124       2       NaN        NaN          NaN
#> 125       3       NaN        NaN          NaN
#> 126       5 0.5725745 0.33878176 0.0886437630
#> 127       7 0.8096820 0.12825144 0.0620665336
#> 128       2       NaN        NaN          NaN
#> 129       4 0.6869484 0.31112373 0.0019279094
#> 130       6 0.6027118 0.34182183 0.0554663563
#> 131       5 0.7199105 0.25727167 0.0228178311
#> 132       4 0.6269875 0.30462520 0.0683873278
#> 133       2       NaN        NaN          NaN
#> 134       5 0.7844127 0.20789079 0.0076965329
#> 135       5 0.5598287 0.31980306 0.1203682789
#> 136       6 0.6143385 0.31105576 0.0746057170
#> 137       8 0.6147812 0.24891370 0.1363050804
#> 138       2       NaN        NaN          NaN
#> 139       5 0.5225841 0.31015831 0.1672575976
#> 140       3       NaN        NaN          NaN
#> 141       3       NaN        NaN          NaN
#> 142       4 0.7894049 0.20709602 0.0034990947
#> 143       1       NaN        NaN          NaN
#> 144       3       NaN        NaN          NaN
#> 145       6 0.5175870 0.26553070 0.2168822569
#> 146       5 0.4790452 0.29758541 0.2233693432
#> 147       4 0.6727741 0.22716102 0.1000648975
#> 148       3       NaN        NaN          NaN
#> 149       4 0.8377557 0.16051855 0.0017257968
#> 150       5 0.6215036 0.20832798 0.1701684205
#> 151       7 0.4479353 0.34344762 0.2086170779
#> 152       4 0.8168084 0.17926281 0.0039287543
#> 153       4 0.6463595 0.23663803 0.1170024488
#> 154       8 0.6354005 0.30203486 0.0625646246
#> 155       4 0.7938536 0.17618048 0.0299658833
#> 156       6 0.7372641 0.20376246 0.0589734703
#> 157       7 0.6232835 0.24872212 0.1279943351
#> 158       3       NaN        NaN          NaN
#> 159       2       NaN        NaN          NaN
#> 160       8 0.5020645 0.32025095 0.1776845910
#> 161       3       NaN        NaN          NaN
#> 162       8 0.5547243 0.39113980 0.0541358625
#> 163       8 0.5429167 0.39435008 0.0627332567
#> 164       4 0.6501735 0.30494580 0.0448807040
#> 165       4 0.7349798 0.25999346 0.0050267734
#> 166       6 0.4573636 0.39011694 0.1525195002
#> 167       3       NaN        NaN          NaN
#> 168       7 0.7645191 0.19113076 0.0443501214
#> 169       3       NaN        NaN          NaN
#> 170       6 0.6505159 0.30539897 0.0440850850
#> 171       7 0.5731841 0.30230936 0.1245065413
#> 172       8 0.5339278 0.41533058 0.0507416346
#> 173       3       NaN        NaN          NaN
#> 174       9 0.5240839 0.35473059 0.1211855323
#> 175       3       NaN        NaN          NaN
#> 176       7 0.6509858 0.26164117 0.0873729917
#> 177       4 0.7464620 0.24754475 0.0059932490
#> 178       1       NaN        NaN          NaN
#> 179       3       NaN        NaN          NaN
#> 180       5 0.5193503 0.37923932 0.1014103271
#> 181       3       NaN        NaN          NaN
#> 182       8 0.5658580 0.38923152 0.0449104958
#> 183       2       NaN        NaN          NaN
#> 184       3       NaN        NaN          NaN
#> 185       4 0.8265802 0.13779136 0.0356284282
#> 186       6 0.7445241 0.23817372 0.0173022135
#> 187       4 0.9565383 0.04263668 0.0008250457
#> 188       9 0.5322367 0.30133629 0.1664269767
#> 189       5 0.8146010 0.14962096 0.0357780853
#> 190       5 0.8238702 0.15244159 0.0236882483
#> 191       6 0.7609118 0.19940324 0.0396849462
#> 192       3       NaN        NaN          NaN
#> 193       6 0.6027118 0.34182183 0.0554663563
#> 194       6 0.6027249 0.25728005 0.1399950530
#> 195       3       NaN        NaN          NaN
#> 196       3       NaN        NaN          NaN
#> 197       4 0.6254233 0.37044308 0.0041336360
#> 198       4 0.6006327 0.39019126 0.0091760390
#> 199       3       NaN        NaN          NaN
#> 200       2       NaN        NaN          NaN
#> 
#> , , radius_4
#> 
#>     npoints      eig1        eig2         eig3
#> 1         4 0.7217507 0.260584692 0.0176645869
#> 2         5 0.6965881 0.294993999 0.0084178808
#> 3         2       NaN         NaN          NaN
#> 4         7 0.6923030 0.254275788 0.0534211959
#> 5         9 0.4429404 0.380588675 0.1764709031
#> 6         5 0.6215036 0.208327977 0.1701684205
#> 7         5 0.7440442 0.229991410 0.0259644342
#> 8         8 0.6116082 0.277552042 0.1108397681
#> 9         7 0.4232000 0.312612444 0.2641875356
#> 10        8 0.5658580 0.389231522 0.0449104958
#> 11       11 0.5805315 0.245320809 0.1741477186
#> 12        2       NaN         NaN          NaN
#> 13        5 0.7606817 0.173669926 0.0656484182
#> 14        5 0.7925793 0.169120504 0.0383002411
#> 15        6 0.6068356 0.321884178 0.0712802648
#> 16       10 0.5107192 0.303183860 0.1860969410
#> 17        5 0.5960616 0.291612577 0.1123258369
#> 18        6 0.4976784 0.423253214 0.0790684149
#> 19        4 0.9028101 0.083360751 0.0138291248
#> 20       12 0.5827921 0.332398246 0.0848096879
#> 21        6 0.6358699 0.306029888 0.0581001836
#> 22        2       NaN         NaN          NaN
#> 23        5 0.5605969 0.318721250 0.1206818828
#> 24        3       NaN         NaN          NaN
#> 25        7 0.5494933 0.273744011 0.1767626695
#> 26        6 0.6903427 0.257183577 0.0524737223
#> 27        8 0.6521922 0.306156857 0.0416509662
#> 28        2       NaN         NaN          NaN
#> 29        5 0.5360791 0.381904661 0.0820162393
#> 30        8 0.5113946 0.275337040 0.2132683304
#> 31        8 0.6030960 0.253300956 0.1436030769
#> 32        4 0.9230866 0.059788339 0.0171251065
#> 33        6 0.6788467 0.252370144 0.0687831643
#> 34        5 0.6954595 0.244692371 0.0598481539
#> 35        6 0.6805797 0.211154858 0.1082654691
#> 36        2       NaN         NaN          NaN
#> 37        4 0.7345093 0.204755979 0.0607346938
#> 38        6 0.5046528 0.459107462 0.0362397544
#> 39        4 0.8764926 0.108122085 0.0153852777
#> 40        3       NaN         NaN          NaN
#> 41        5 0.8070409 0.172603895 0.0203551767
#> 42        6 0.6527159 0.263424263 0.0838598500
#> 43        8 0.4351772 0.384136108 0.1806867032
#> 44        9 0.6684813 0.296379675 0.0351390598
#> 45        8 0.6071363 0.257347171 0.1355165259
#> 46        6 0.6781170 0.315467271 0.0064157480
#> 47       10 0.7235694 0.163049572 0.1133810261
#> 48        9 0.6394962 0.249821866 0.1106819563
#> 49        8 0.5549144 0.283243442 0.1618421351
#> 50        4 0.6426875 0.213032189 0.1442803199
#> 51        5 0.7936528 0.190209108 0.0161380727
#> 52        4 0.5865691 0.408917063 0.0045138081
#> 53        8 0.5525650 0.285072548 0.1623624086
#> 54        6 0.7687876 0.211179959 0.0200324136
#> 55        7 0.6647927 0.290147285 0.0450600551
#> 56        9 0.5723496 0.303363046 0.1242873907
#> 57        4 0.5520262 0.428127349 0.0198464666
#> 58        6 0.6711210 0.239119726 0.0897592611
#> 59        5 0.5097063 0.355909519 0.1343841466
#> 60        4 0.8643564 0.096679318 0.0389643051
#> 61        3       NaN         NaN          NaN
#> 62        2       NaN         NaN          NaN
#> 63        8 0.5206348 0.329996124 0.1493690412
#> 64        6 0.6274503 0.237670936 0.1348787894
#> 65        4 0.7402555 0.257950319 0.0017941992
#> 66        7 0.4649665 0.359401154 0.1756322972
#> 67       10 0.6949872 0.170069438 0.1349434018
#> 68        9 0.5026097 0.357824210 0.1395661251
#> 69        8 0.6288422 0.330345782 0.0408119942
#> 70        9 0.5590372 0.327243915 0.1137188853
#> 71        5 0.4255586 0.408074001 0.1663674024
#> 72        7 0.5680112 0.320563234 0.1114255206
#> 73        5 0.7935190 0.144772460 0.0617085428
#> 74        6 0.6703940 0.230699305 0.0989066629
#> 75        8 0.5317385 0.259900815 0.2083606624
#> 76        6 0.6061440 0.356945660 0.0369103388
#> 77        5 0.5323414 0.368734611 0.0989239855
#> 78        3       NaN         NaN          NaN
#> 79        5 0.5578501 0.301221752 0.1409281938
#> 80        3       NaN         NaN          NaN
#> 81        7 0.5705703 0.347646725 0.0817829298
#> 82        4 0.7567168 0.227394525 0.0158886258
#> 83        9 0.5918837 0.329401998 0.0787142735
#> 84        5 0.7642671 0.205498093 0.0302347912
#> 85        8 0.5582289 0.269602534 0.1721686003
#> 86        9 0.6813769 0.239095326 0.0795277928
#> 87        5 0.7925793 0.169120504 0.0383002411
#> 88        6 0.5122628 0.337232871 0.1505043233
#> 89        4 0.8007456 0.146044401 0.0532100154
#> 90        6 0.6016369 0.275064220 0.1232988740
#> 91        6 0.6114004 0.230738204 0.1578613551
#> 92        7 0.6377563 0.233471833 0.1287718175
#> 93        9 0.5872379 0.318110187 0.0946519355
#> 94        5 0.6954595 0.244692371 0.0598481539
#> 95        9 0.5520155 0.258799574 0.1891848765
#> 96        8 0.6071988 0.350927614 0.0418736192
#> 97        4 0.6779980 0.269371488 0.0526305164
#> 98        6 0.6261087 0.304480783 0.0694104859
#> 99        9 0.5861135 0.308201117 0.1056853368
#> 100       5 0.8778416 0.087666324 0.0344920984
#> 101       5 0.5287811 0.421225242 0.0499936602
#> 102       5 0.5224557 0.294377222 0.1831670424
#> 103       3       NaN         NaN          NaN
#> 104       5 0.8160740 0.181639873 0.0022861335
#> 105       6 0.6180891 0.367812892 0.0140979765
#> 106       7 0.6857819 0.287911782 0.0263063475
#> 107      10 0.5096626 0.366017292 0.1243201094
#> 108       6 0.4599158 0.396278436 0.1438058091
#> 109       8 0.6120439 0.276773054 0.1111830662
#> 110       5 0.7642671 0.205498093 0.0302347912
#> 111      11 0.5391204 0.422803643 0.0380759532
#> 112       9 0.5180998 0.375714580 0.1061856254
#> 113       3       NaN         NaN          NaN
#> 114       9 0.5431369 0.326130519 0.1307326124
#> 115       7 0.5008311 0.341524555 0.1576443152
#> 116       3       NaN         NaN          NaN
#> 117       5 0.7247039 0.237475182 0.0378209029
#> 118      12 0.5209154 0.286569282 0.1925152876
#> 119       3       NaN         NaN          NaN
#> 120       6 0.5832691 0.280462618 0.1362682337
#> 121       7 0.5735654 0.262303575 0.1641310705
#> 122      12 0.4332699 0.297753669 0.2689764464
#> 123       8 0.4885430 0.356782371 0.1546746343
#> 124       3       NaN         NaN          NaN
#> 125       6 0.6355220 0.211487078 0.1529908886
#> 126       8 0.5913640 0.305448598 0.1031873896
#> 127      10 0.6538820 0.210572920 0.1355450323
#> 128       3       NaN         NaN          NaN
#> 129       5 0.5115316 0.467345543 0.0211228867
#> 130       6 0.6027118 0.341821827 0.0554663563
#> 131       9 0.4352402 0.341035302 0.2237244679
#> 132       5 0.6175282 0.210370126 0.1721016686
#> 133       6 0.7639005 0.212412256 0.0236872100
#> 134       7 0.5938275 0.338691527 0.0674809364
#> 135       6 0.6403523 0.263134299 0.0965134160
#> 136       8 0.5828285 0.319942883 0.0972286141
#> 137       9 0.4634558 0.410881679 0.1256625032
#> 138       3       NaN         NaN          NaN
#> 139       7 0.6514597 0.216423888 0.1321163811
#> 140       3       NaN         NaN          NaN
#> 141       4 0.7062161 0.275991577 0.0177923636
#> 142       6 0.7654408 0.162991398 0.0715677596
#> 143       2       NaN         NaN          NaN
#> 144       3       NaN         NaN          NaN
#> 145       6 0.5175870 0.265530705 0.2168822569
#> 146       6 0.5637167 0.266202990 0.1700802670
#> 147       8 0.5580856 0.308102065 0.1338122931
#> 148       3       NaN         NaN          NaN
#> 149       4 0.8377557 0.160518548 0.0017257968
#> 150       6 0.5122628 0.337232871 0.1505043233
#> 151      11 0.4404948 0.341972354 0.2175328233
#> 152       7 0.5359058 0.411896669 0.0521975310
#> 153       7 0.6381738 0.324642115 0.0371841043
#> 154      13 0.5333491 0.317985262 0.1486656653
#> 155       5 0.5323414 0.368734611 0.0989239855
#> 156       9 0.5620586 0.386351294 0.0515900680
#> 157       8 0.5301368 0.331796126 0.1380670874
#> 158       5 0.8335132 0.149534817 0.0169520042
#> 159       4 0.9879054 0.008139715 0.0039549241
#> 160      10 0.4314948 0.347697221 0.2208079544
#> 161       5 0.8377001 0.125636991 0.0366629246
#> 162      11 0.4532560 0.324167386 0.2225766156
#> 163       9 0.5230943 0.321562263 0.1553434419
#> 164       6 0.5955389 0.291631244 0.1128298431
#> 165       6 0.6060843 0.309550689 0.0843650570
#> 166       8 0.5240134 0.326913219 0.1490733863
#> 167       4 0.7545665 0.245048853 0.0003846345
#> 168      10 0.6411988 0.262282016 0.0965191410
#> 169       4 0.6463595 0.236638028 0.1170024488
#> 170       9 0.5249931 0.264882662 0.2101242126
#> 171       8 0.4593870 0.424733915 0.1158791194
#> 172       9 0.5799277 0.377577922 0.0424943344
#> 173       5 0.5475491 0.349100040 0.1033508212
#> 174      11 0.5376369 0.367730077 0.0946330528
#> 175       5 0.6091905 0.227792972 0.1630165163
#> 176       8 0.6057058 0.306619360 0.0876748697
#> 177       4 0.7464620 0.247544748 0.0059932490
#> 178       6 0.6570370 0.300377923 0.0425850701
#> 179       4 0.8733911 0.108127610 0.0184812823
#> 180       7 0.4922029 0.330890245 0.1769069024
#> 181       4 0.8914400 0.107633274 0.0009266918
#> 182       9 0.4570472 0.357416819 0.1855359994
#> 183       6 0.5850801 0.326287853 0.0886320943
#> 184       7 0.8042613 0.110313997 0.0854247409
#> 185       6 0.5956710 0.369583714 0.0347452802
#> 186       9 0.6208099 0.248926970 0.1302630934
#> 187       5 0.6948333 0.279503280 0.0256634487
#> 188      11 0.4902526 0.320434402 0.1893130032
#> 189       8 0.7434008 0.183852179 0.0727469924
#> 190       8 0.4826690 0.385065614 0.1322653849
#> 191       8 0.5803499 0.254535772 0.1651142960
#> 192       6 0.6079857 0.374359216 0.0176551234
#> 193       8 0.5968639 0.287340015 0.1157960679
#> 194       7 0.4659565 0.346420821 0.1876226654
#> 195       5 0.5930888 0.349090854 0.0578202970
#> 196       4 0.8892351 0.110584868 0.0001799963
#> 197       7 0.5322972 0.376908605 0.0907941781
#> 198       6 0.6368513 0.325576800 0.0375719059
#> 199       4 0.7894049 0.207096019 0.0034990947
#> 200       2       NaN         NaN          NaN
#>