K Nearest Neighbors

K nearest neighbors based on FLANN C++

knn(query, ref, k, same = FALSE, build = "kdtree", threads = 1L, checks = 1L)

Arguments

query
A data.table containing the set of query points where each row represent a point and each column a given coordinate.

ref
A numeric containing the set of reference points where each row represent a point and each column a given coordinate.

k
An integer describing the number of nearest neighbors to search for.

same
Logic. If TRUE, it delete neighbors with distance of 0, useful when the k search is based on the same query.

build
A character describing the search structure to be used: "kdtree", "kmeans", "linear".

threads
An integer specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware. If 0, then the maximum allowable cores are used.

checks
Number of checks during searching. Higher value gives better search precision but takes longer. See FLANN C++ manual for more details.

Value

A data.table with three columns describing the indices of the query, ref, and k neighbors and the distances.

Details

This function is based on the FLAAN C++ Library (Muja and Lowe 2009) for a fast estimation of neighbors points. It is adapted from rflann (Muja, Lowe, Yee, 2018) to simplify the workflow with TLS data. If you use this function, please consider cite both documents.

References

Muja, M., Lowe, D.G. (2009). Fast approximate nearest neighbors with automatic algorithm configuration. VISAPP (1), 2(331-340), 2.

Muja, M., Lowe, D.G., Yee, J., (2018). rflann: Basic R Interface to the 'FLANN' C++ Library. R package version 1.6.0.

See also

Author

J. Antonio Guzmán Q.

Examples

#Point cloud
data("pc_tree")

#knn search using k = 3
knn(pc_tree, pc_tree, k = 3, same = TRUE, "kdtree", checks = 10)

#>         query   ref k_index    distance
#>      1:     1     3       1 0.002586577
#>      2:     1     2       2 0.005620912
#>      3:     1    12       3 0.006598868
#>      4:     2     8       1 0.002943683
#>      5:     2     7       2 0.003253250
#>     ---                                
#> 227540: 75847 75848       2 0.002777766
#> 227541: 75847 74605       3 0.002952626
#> 227542: 75848 75847       1 0.002777766
#> 227543: 75848 75846       2 0.003168111
#> 227544: 75848 75842       3 0.004644131