Stand Counting

Applies the voxels_counting function on a grid base point cloud.

stand_counting(
  cloud,
  xy.res,
  z.res = NULL,
  points.min = NULL,
  min_size,
  edge_sizes = NULL,
  length_out = 10,
  bootstrap = FALSE,
  R = NULL,
  progress = TRUE,
  parallel = FALSE,
  threads = NULL
)

Arguments

cloud
A data.table of a point cloud with xyz coordinates in the first three columns.

xy.res
A positive numeric vector describing the grid resolution of the xy coordinates to perform.

z.res
A positive numeric vector of length 1 describing the vertical resolution. If z.res = NULL vertical profiles are not used.

points.min
A positive numeric vector of length 1 minimum number of points to retain a sub-grid.

min_size
A positive numeric vector of length 1 describing the minimum cube edge length to perform. This is required if edge_sizes = NULL.

edge_sizes
A positive numeric vector describing the edge length of the different cubes to perform within each subgrid when z.res = NULL. If edge_sizes = NULL, it uses the maximum range of values for the xyz coordinates.

length_out
A positive interger of length 1 indicating the number of different edge lengths to use for each subgrid. This is required if edge_sizes = NULL.

bootstrap
Logical. If TRUE, it computes a bootstrap on the H index calculations. FALSE as default.

R
A positive integer of length 1 indicating the number of bootstrap replicates. This need to be used if bootstrap = TRUE.

progress
Logical, if TRUE displays a graphical progress bar. TRUE as default.

parallel
Logical, if TRUE it uses a parallel processing for the voxelization. FALSE as default.

threads
An integer >= 0 describing the number of threads to use. This need to be used if parallel = TRUE.

Value

A data.table with the summary of the voxels per grid created with their features.

See also

Author

J. Antonio Guzmán Q.

Examples

data(pc_tree)

#Applying stand_counting.
stand_counting(pc_tree, xy.res = c(4, 4), min_size = 2)

#> Estimating stand_counting  |                                                                              |                                                                      |   0%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%

#>            X        Y   Edge.X   Edge.Y   Edge.Z N_voxels    Volume   Surface
#>  1:  8.63575 -1.53375 6.036600 6.036600 6.036600        1 219.97696 36.440539
#>  2:  8.63575 -1.53375 4.177068 4.177068 4.177068        2 145.76216 17.447901
#>  3:  8.63575 -1.53375 4.722571 4.722571 4.722571        2 210.65200 22.302680
#>  4:  8.63575 -1.53375 5.339314 5.339314 5.339314        2 304.42926 28.508274
#>  5:  8.63575 -1.53375 3.267817 3.267817 3.267817        8 279.16642 42.714512
#>  6:  8.63575 -1.53375 3.694577 3.694577 3.694577        8 403.44467 54.599582
#>  7:  8.63575 -1.53375 2.890352 2.890352 2.890352       10 241.46400 33.416548
#>  8:  8.63575 -1.53375 2.000000 2.000000 2.000000       12  96.00000 16.000000
#>  9:  8.63575 -1.53375 2.261189 2.261189 2.261189       12 138.73692 20.451909
#> 10:  8.63575 -1.53375 2.556489 2.556489 2.556489       12 200.49929 26.142536
#> 11:  8.63575  2.46625 3.800350 3.800350 3.800350        1  54.88718 14.442662
#> 12:  8.63575  2.46625 3.538723 3.538723 3.538723        2  88.62774 12.522560
#> 13:  8.63575  2.46625 2.147865 2.147865 2.147865        4  39.63521  9.226652
#> 14:  8.63575  2.46625 2.306663 2.306663 2.306663        4  49.09219 10.641388
#> 15:  8.63575  2.46625 2.477201 2.477201 2.477201        4  60.80560 12.273047
#> 16:  8.63575  2.46625 2.660347 2.660347 2.660347        4  75.31385 14.154892
#> 17:  8.63575  2.46625 2.857034 2.857034 2.857034        4  93.28376 16.325282
#> 18:  8.63575  2.46625 3.068262 3.068262 3.068262        4 115.54130 18.828461
#> 19:  8.63575  2.46625 3.295107 3.295107 3.295107        4 143.10950 21.715457
#> 20:  8.63575  2.46625 2.000000 2.000000 2.000000        5  40.00000 12.000000
#> 21: 12.63575 -1.53375 3.881850 3.881850 3.881850        1  58.49467 15.068760
#> 22: 12.63575 -1.53375 3.606100 3.606100 3.606100        3 140.68073 26.007917
#> 23: 12.63575 -1.53375 2.494784 2.494784 2.494784        4  62.10960 12.447893
#> 24: 12.63575 -1.53375 2.685554 2.685554 2.685554        4  77.47501 14.424401
#> 25: 12.63575 -1.53375 2.890912 2.890912 2.890912        4  96.64171 16.714744
#> 26: 12.63575 -1.53375 3.111973 3.111973 3.111973        4 120.55008 19.368754
#> 27: 12.63575 -1.53375 3.349938 3.349938 3.349938        4 150.37320 22.444174
#> 28: 12.63575 -1.53375 2.317565 2.317565 2.317565        5  62.23946 16.113324
#> 29: 12.63575 -1.53375 2.000000 2.000000 2.000000        6  48.00000 16.000000
#> 30: 12.63575 -1.53375 2.152935 2.152935 2.152935        6  59.87481 13.905391
#> 31: 12.63575  2.46625 3.272350 3.272350 3.272350        1  35.04122 10.708274
#> 32: 12.63575  2.46625 2.112462 2.112462 2.112462        2  18.85370  4.462495
#> 33: 12.63575  2.46625 2.231247 2.231247 2.231247        2  22.21637  4.978465
#> 34: 12.63575  2.46625 2.356712 2.356712 2.356712        2  26.17880  5.554093
#> 35: 12.63575  2.46625 2.489232 2.489232 2.489232        2  30.84795  6.196278
#> 36: 12.63575  2.46625 2.629204 2.629204 2.629204        2  36.34987  6.912714
#> 37: 12.63575  2.46625 2.777047 2.777047 2.777047        2  42.83310  7.711988
#> 38: 12.63575  2.46625 2.933202 2.933202 2.933202        2  50.47264  8.603676
#> 39: 12.63575  2.46625 3.098139 3.098139 3.098139        2  59.47475  9.598464
#> 40: 12.63575  2.46625 2.000000 2.000000 2.000000        3  24.00000  8.000000
#>            X        Y   Edge.X   Edge.Y   Edge.Z N_voxels    Volume   Surface
#>     Density_mean Density_sd          H      Hmax Equitavility   Negentropy
#>  1:    1129.0173         NA 0.00000000 0.0000000          NaN 0.0000000000
#>  2:    1178.9957   564.9429 0.63459338 0.6931472   0.91552473 0.0585537968
#>  3:     922.3555   871.2538 0.44954267 0.6931472   0.64855299 0.2436045074
#>  4:     721.5800   903.9411 0.21889817 0.6931472   0.31580330 0.4742490106
#>  5:     481.5928   431.1619 1.75535256 2.0794415   0.84414615 0.3240889768
#>  6:     376.7611   522.9736 1.34472626 2.0794415   0.64667663 0.7347152855
#>  7:     492.4746   478.4023 1.86229125 2.3025851   0.80878281 0.4402938479
#>  8:     857.1250   790.1135 2.02594336 2.4849066   0.81529958 0.4589632896
#>  9:     670.5487   631.0936 2.06654022 2.4849066   0.83163696 0.4183664325
#> 10:     524.5857   524.2199 2.04789081 2.4849066   0.82413189 0.4370158352
#> 11:    1123.2001         NA 0.00000000 0.0000000          NaN 0.0000000000
#> 12:     647.7110   902.3368 0.04396782 0.6931472   0.06343215 0.6491793648
#> 13:     879.0838   101.7269 1.38144787 1.3862944   0.99650400 0.0048464902
#> 14:     762.2126   186.5675 1.36273762 1.3862944   0.98300740 0.0235567421
#> 15:     660.8791   315.7204 1.29690211 1.3862944   0.93551712 0.0893922550
#> 16:     573.0175   442.7600 1.16668236 1.3862944   0.84158343 0.2196120031
#> 17:     496.8367   526.5978 0.98893289 1.3862944   0.71336429 0.3973614752
#> 18:     430.7840   615.7567 0.71144293 1.3862944   0.51319760 0.6748514264
#> 19:     373.5127   670.1128 0.33636421 1.3862944   0.24263549 1.0499301462
#> 20:     811.1000   481.3807 1.37467479 1.6094379   0.85413347 0.2347631247
#> 21:     896.4905         NA 0.00000000 0.0000000          NaN 0.0000000000
#> 22:     346.2792   577.4047 0.13296706 1.0986123   0.12103183 0.9656452289
#> 23:     542.6220   405.4637 1.16993108 1.3862944   0.84392688 0.2163632822
#> 24:     468.2690   489.4047 0.95522632 1.3862944   0.68905014 0.4310680410
#> 25:     404.1043   528.2919 0.73483550 1.3862944   0.53007176 0.6514588642
#> 26:     348.7318   527.4356 0.55245577 1.3862944   0.39851260 0.8338385937
#> 27:     300.9467   507.6134 0.41386782 1.3862944   0.29854252 0.9724265459
#> 28:     503.0247   379.1392 1.30435617 1.6094379   0.81044205 0.3050817458
#> 29:     562.8750   412.3890 1.47748360 1.7917595   0.82459930 0.3142758645
#> 30:     485.7469   387.7720 1.40174654 1.7917595   0.78232964 0.3900129327
#> 31:     464.5006         NA 0.00000000 0.0000000          NaN 0.0000000000
#> 32:     557.3116    34.5433 0.69218643 0.6931472   0.99861393 0.0009607504
#> 33:     499.5516   165.3265 0.66550966 0.6931472   0.96012749 0.0276375167
#> 34:     447.7779   281.6158 0.59071413 0.6931472   0.85222035 0.1024330465
#> 35:     401.3700   347.6034 0.49169745 0.6931472   0.70936947 0.2014497344
#> 36:     359.7719   370.4971 0.39746283 0.6931472   0.57341765 0.2956843518
#> 37:     322.4850   375.3760 0.29896025 0.6931472   0.43130847 0.3941869277
#> 38:     289.0625   372.1409 0.18301102 0.6931472   0.26402910 0.5101361571
#> 39:     259.1039   360.9768 0.04386838 0.6931472   0.06328870 0.6492787999
#> 40:     414.5000   363.5075 0.69905656 1.0986123   0.63630870 0.3995557298
#>     Density_mean Density_sd          H      Hmax Equitavility   Negentropy

#Applying stand_counting using bootstrap in the H index.
stand_counting(pc_tree, xy.res = c(4, 4), min_size = 1, bootstrap = TRUE, R = 1000)

#> Estimating stand_counting  |                                                                              |                                                                      |   0%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%

#>            X        Y   Edge.X   Edge.Y   Edge.Z N_voxels    Volume   Surface
#>  1:  8.63575 -1.53375 6.036600 6.036600 6.036600        1 219.97696 36.440539
#>  2:  8.63575 -1.53375 4.048396 4.048396 4.048396        2 132.70244 16.389509
#>  3:  8.63575 -1.53375 4.943536 4.943536 4.943536        2 241.62566 24.438547
#>  4:  8.63575 -1.53375 3.315341 3.315341 3.315341        8 291.52431 43.965956
#>  5:  8.63575 -1.53375 2.223406 2.223406 2.223406       12 131.89787 19.774143
#>  6:  8.63575 -1.53375 2.715023 2.715023 2.715023       12 240.16068 29.485405
#>  7:  8.63575 -1.53375 1.820808 1.820808 1.820808       28 169.02480 29.838073
#>  8:  8.63575 -1.53375 1.491109 1.491109 1.491109       29  96.14490 20.010657
#>  9:  8.63575 -1.53375 1.221110 1.221110 1.221110       59 107.42767 23.857746
#> 10:  8.63575 -1.53375 1.000000 1.000000 1.000000       66  66.00000 16.000000
#> 11:  8.63575  2.46625 3.800350 3.800350 3.800350        1  54.88718 14.442662
#> 12:  8.63575  2.46625 2.099540 2.099540 2.099540        4  37.01968  8.816140
#> 13:  8.63575  2.46625 2.435281 2.435281 2.435281        4  57.77065 11.861187
#> 14:  8.63575  2.46625 2.824710 2.824710 2.824710        4  90.15333 15.957978
#> 15:  8.63575  2.46625 3.276414 3.276414 3.276414        4 140.68777 21.469778
#> 16:  8.63575  2.46625 1.810087 1.810087 1.810087        6  35.58356  9.829242
#> 17:  8.63575  2.46625 1.560539 1.560539 1.560539       13  49.40455 14.611686
#> 18:  8.63575  2.46625 1.159911 1.159911 1.159911       17  26.52916  8.072368
#> 19:  8.63575  2.46625 1.345395 1.345395 1.345395       17  41.39978 10.860521
#> 20:  8.63575  2.46625 1.000000 1.000000 1.000000       27  27.00000  9.000000
#> 21: 12.63575 -1.53375 3.881850 3.881850 3.881850        1  58.49467 15.068760
#> 22: 12.63575 -1.53375 2.469975 2.469975 2.469975        4  60.27504 12.201550
#> 23: 12.63575 -1.53375 2.871714 2.871714 2.871714        4  94.72918 16.493488
#> 24: 12.63575 -1.53375 3.338797 3.338797 3.338797        4 148.87783 22.295130
#> 25: 12.63575 -1.53375 2.124437 2.124437 2.124437        7  67.11650 18.052922
#> 26: 12.63575 -1.53375 1.827238 1.827238 1.827238       10  61.00775 13.355188
#> 27: 12.63575 -1.53375 1.571615 1.571615 1.571615       14  54.34590 14.819848
#> 28: 12.63575 -1.53375 1.351753 1.351753 1.351753       15  37.04962 10.963425
#> 29: 12.63575 -1.53375 1.000000 1.000000 1.000000       26  26.00000 10.000000
#> 30: 12.63575 -1.53375 1.162649 1.162649 1.162649       26  40.86200 12.165781
#> 31: 12.63575  2.46625 3.272350 3.272350 3.272350        1  35.04122 10.708274
#> 32: 12.63575  2.46625 2.204143 2.204143 2.204143        2  21.41655  4.858248
#> 33: 12.63575  2.46625 2.514470 2.514470 2.514470        2  31.79577  6.322560
#> 34: 12.63575  2.46625 2.868488 2.868488 2.868488        2  47.20514  8.228226
#> 35: 12.63575  2.46625 1.693661 1.693661 1.693661        3  14.57474  5.736977
#> 36: 12.63575  2.46625 1.932116 1.932116 1.932116        3  21.63818  7.466144
#> 37: 12.63575  2.46625 1.484636 1.484636 1.484636        4  13.08940  4.408287
#> 38: 12.63575  2.46625 1.301407 1.301407 1.301407        6  13.22486  5.080984
#> 39: 12.63575  2.46625 1.140792 1.140792 1.140792        7  10.39245  3.904222
#> 40: 12.63575  2.46625 1.000000 1.000000 1.000000       10  10.00000  5.000000
#>            X        Y   Edge.X   Edge.Y   Edge.Z N_voxels    Volume   Surface
#>     Density_mean Density_sd         H      Hmax Equitavility  Negentropy
#>  1:    1129.0173         NA 0.0000000 0.0000000          NaN 0.000000000
#>  2:    1255.1321   450.6808 0.6605587 0.6931472    0.9529848 0.032588462
#>  3:     841.7440   884.3411 0.3836301 0.6931472    0.5534612 0.309517076
#>  4:     467.8847   437.2539 1.7292198 2.0794415    0.8315789 0.350221730
#>  5:     693.5319   639.8425 2.0752316 2.4849066    0.8351346 0.409675053
#>  6:     465.1115   498.3290 1.9735241 2.4849066    0.7942045 0.511382532
#>  7:     443.1993   546.6997 2.6926361 3.3322045    0.8080645 0.639568439
#>  8:     638.0703   425.4575 3.1135049 3.3672958    0.9246307 0.253790890
#>  9:     467.6533   440.1526 3.6677723 4.0775374    0.8995067 0.409765121
#> 10:     623.3636   471.7274 3.8523390 4.1896547    0.9194884 0.337315785
#> 11:    1123.2001         NA 0.0000000 0.0000000          NaN 0.000000000
#> 12:     920.0172   109.6527 1.3809921 1.3862944    0.9961752 0.005302293
#> 13:     683.8270   285.5811 1.3176136 1.3862944    0.9504573 0.068680807
#> 14:     508.2724   518.8509 1.0164002 1.3862944    0.7331778 0.369894120
#> 15:     377.7869   670.0431 0.3626057 1.3862944    0.2615648 1.023688623
#> 16:     825.1908   709.0685 1.3905044 1.7917595    0.7760553 0.401255058
#> 17:     512.4033   653.2202 1.9340863 2.5649494    0.7540446 0.630863017
#> 18:     709.2605   556.5018 2.4895802 2.8332133    0.8787126 0.343633110
#> 19:     527.1766   541.6750 2.3242042 2.8332133    0.8203421 0.509009095
#> 20:     600.8148   464.2011 2.9703844 3.2958369    0.9012535 0.325452464
#> 21:     896.4905         NA 0.0000000 0.0000000          NaN 0.000000000
#> 22:     553.5772   393.2946 1.1901507 1.3862944    0.8585122 0.196143700
#> 23:     409.5253   526.8182 0.7515380 1.3862944    0.5421201 0.634756371
#> 24:     302.9585   508.7174 0.4203259 1.3862944    0.3032010 0.965968459
#> 25:     427.6000   402.1570 1.4182626 1.9459101    0.7288428 0.527647587
#> 26:     404.6068   507.6943 1.5553657 2.3025851    0.6754867 0.747219418
#> 27:     390.6633   496.7502 1.9366175 2.6390573    0.7338293 0.702439805
#> 28:     492.8752   500.7680 2.2639081 2.7080502    0.8359919 0.444142147
#> 29:     519.5769   366.1633 2.9923865 3.2580965    0.9184462 0.265709991
#> 30:     384.3725   400.5737 2.7225885 3.2580965    0.8356378 0.535508050
#> 31:     464.5006         NA 0.0000000 0.0000000          NaN 0.000000000
#> 32:     511.9130   134.1950 0.6758675 0.6931472    0.9750707 0.017279647
#> 33:     393.3534   354.5286 0.4734238 0.6931472    0.6830062 0.219723362
#> 34:     302.2523   376.4029 0.2261702 0.6931472    0.3262946 0.466976987
#> 35:     578.0048   550.5213 0.7272371 1.0986123    0.6619598 0.371375158
#> 36:     444.1382   396.3025 0.7086877 1.0986123    0.6450754 0.389924558
#> 37:     564.1648   587.4070 0.9691492 1.3862944    0.6990934 0.417145116
#> 38:     489.4721   529.0995 1.2783259 1.7917595    0.7134473 0.513433600
#> 39:     546.0023   495.7409 1.5112247 1.9459101    0.7766159 0.434685436
#> 40:     497.4000   591.3253 1.6193287 2.3025851    0.7032655 0.683256399
#>     Density_mean Density_sd         H      Hmax Equitavility  Negentropy
#>     H_boot_mean   H_boot_sd Equitavility_boot Negentropy_boot
#>  1:   0.0000000 0.000000000               NaN     0.000000000
#>  2:   0.6774070 0.016292958         0.9772917     0.015740227
#>  3:   0.5430314 0.154766284         0.7834287     0.150115782
#>  4:   1.7733606 0.112475386         0.8528062     0.306080915
#>  5:   2.0886639 0.121182646         0.8405402     0.396242771
#>  6:   1.9931570 0.141081607         0.8021054     0.491749617
#>  7:   2.7126893 0.115705253         0.8140825     0.619515166
#>  8:   3.1150843 0.069157773         0.9250997     0.252211537
#>  9:   3.6718121 0.053335392         0.9004975     0.405725337
#> 10:   3.8565620 0.054249509         0.9204964     0.333092695
#> 11:   0.0000000 0.000000000               NaN     0.000000000
#> 12:   1.3822512 0.002392096         0.9970835     0.004043184
#> 13:   1.3312987 0.036308455         0.9603290     0.054995636
#> 14:   1.1141673 0.168570667         0.8037018     0.272127055
#> 15:   0.7626648 0.370262673         0.5501464     0.623629543
#> 16:   1.3816812 0.262477502         0.7711309     0.410078318
#> 17:   1.9720070 0.152882002         0.7688288     0.592942389
#> 18:   2.4857021 0.107901796         0.8773438     0.347511203
#> 19:   2.3401924 0.133435442         0.8259852     0.493020947
#> 20:   2.9723930 0.087753204         0.9018629     0.323443845
#> 21:   0.0000000 0.000000000               NaN     0.000000000
#> 22:   1.2312089 0.091849182         0.8881295     0.155085489
#> 23:   0.8468215 0.281166577         0.6108526     0.539472831
#> 24:   0.6413647 0.331515763         0.4626468     0.744929687
#> 25:   1.3648135 0.320176929         0.7013754     0.581096634
#> 26:   1.5373611 0.271198901         0.6676674     0.765223978
#> 27:   1.9559149 0.212362625         0.7411415     0.683142386
#> 28:   2.2854536 0.112644297         0.8439480     0.422596589
#> 29:   2.9971377 0.065565169         0.9199045     0.260958856
#> 30:   2.7205521 0.126582955         0.8350127     0.537544389
#> 31:   0.0000000 0.000000000               NaN     0.000000000
#> 32:   0.6846974 0.008642054         0.9878096     0.008449747
#> 33:   0.5839447 0.109914675         0.8424541     0.109202511
#> 34:   0.4615266 0.233597850         0.6658421     0.231620585
#> 35:   0.7767095 0.254866038         0.7069915     0.321902747
#> 36:   0.7115540 0.353231631         0.6476843     0.387058306
#> 37:   1.0500773 0.186178865         0.7574707     0.336217041
#> 38:   1.3012513 0.256225355         0.7262422     0.490508140
#> 39:   1.4970035 0.250523802         0.7693076     0.448906620
#> 40:   1.6173081 0.243451937         0.7023880     0.685276993
#>     H_boot_mean   H_boot_sd Equitavility_boot Negentropy_boot