Voxels Summary

Create a summary objects of class "voxels" created using the voxels.

summary_voxels(voxels, voxel.size = NULL, bootstrap = FALSE, R = NULL)

Arguments

voxels
An object of class voxels created using the voxels() function or a data.table describing the voxels coordinates and their number of points produced using voxels().

voxel.size
A positive numeric vector with the size of the voxel. This need to be used if class(voxels) != "voxels". It use the same dimensional scale of the point cloud.

bootstrap
Logical, if TRUE it computes a bootstrap on the H index calculations. FALSE as default.

R
A positive integer of length 1 indicating the number of bootstrap replicates. This need to be used if bootstrap = TRUE.

Value

A data.table with with the summary of voxels.

Details

The function estimate 10 main statistics of the voxels. Specifically, the first four columns (ei. Voxel.size, N_voxels, volume) describe the size of the voxel used, the number of voxels created, the total volume that they represent, and the surface area that they cover. Following columns represent the mean (Density_mean) and sd (Density_sd) of the density of points per voxel. Columns 7:10 provide metrics calculated using the Shannon Index. Specifically, H describe the entropy, H_max the maximum entropy, Equitavility the ratio between H and Hmax, and Negentropy describe the product of Hmax - H. If bootstrap = TRUE four more columns are created (11:14). These represent the mean and sd of the H index estimated using bootstrap (H_boot_mean and H_boot_sd), the Equtavility_boot as the ratio of the ratio between H_boot_sd and Hmax, and Negentropy_boot as the product Hmax - H_boot_mean.

See also

Examples

data("pc_tree")

#Apply a summary on a object of class "voxels" using bootstrap with 1000 replicates.
vox <- voxels(pc_tree, voxel.size = 0.5)
summary_voxels(vox, bootstrap = TRUE, R = 1000)

#>    Voxel.size N_voxels volume Surface Density_mean Density_sd        H     Hmax
#> 1:        0.5      679 84.875   31.75     893.6436   653.6787 6.221814 6.520621
#>    Equitavility Negentropy H_boot_mean  H_boot_sd Equitavility_boot
#> 1:     0.954175  0.2988072    6.222862 0.01507356         0.9543358
#>    Negentropy_boot
#> 1:       0.2977592

#Apply a summary on a product from 'voxels' using bootstrap with 1000 replicates.
vox <- voxels(pc_tree, voxel.size = 0.5, obj.voxels = FALSE)
summary_voxels(vox, voxel.size = 0.5, bootstrap = TRUE, R = 1000)

#>    Voxel.size N_voxels volume Surface Density_mean Density_sd        H     Hmax
#> 1:        0.5      679 84.875   31.75     893.6436   653.6787 6.221814 6.520621
#>    Equitavility Negentropy H_boot_mean  H_boot_sd Equitavility_boot
#> 1:     0.954175  0.2988072    6.222152 0.01535183          0.954227
#>    Negentropy_boot
#> 1:       0.2984687