Voxels Summary

Create a summary objects of class "voxels" created using the voxels.

summary_voxels(voxels, edge_length = NULL, bootstrap = FALSE, R = NULL)

Arguments

voxels
An object of class voxels created using the voxels() function or a data.table describing the voxels coordinates and their number of points produced using voxels().

edge_length
A positive numeric vector with the voxel-edge length for the x, y, and z coordinates. This need to be used if class(voxels) != "voxels". It use the same dimensional scale of the point cloud.

bootstrap
Logical, if TRUE it computes a bootstrap on the H index calculations. FALSE as default.

R
A positive integer of length 1 indicating the number of bootstrap replicates. This need to be used if bootstrap = TRUE.

Value

A data.table with with the summary of voxels.

Details

The function provides 12 main statistics of the voxels. Specifically, the first three columns represent the edge length of the voxels, the following three columns (ei. N_voxels, Volume, Surface) describe the number of voxels created, the total volume that they represent, and the surface area that they cover. Following columns represent the mean (Density_mean) and sd (Density_sd) of the density of points per voxel (e.g. points/m2). Columns 9:12 provide metrics calculated using the Shannon Index. Specifically, H describe the entropy, H_max the maximum entropy, Equitavility the ratio between H and Hmax, and Negentropy describe the product of Hmax - H. If bootstrap = TRUE four more columns are created (13:16). These represent the mean and sd of the H index estimated using bootstrap (H_boot_mean and H_boot_sd), the Equtavility_boot as the ratio of the ratio between H_boot_sd and Hmax, and Negentropy_boot as the product Hmax - H_boot_mean.

See also

Author

J. Antonio Guzmán Q.

Examples

data("pc_tree")

#Apply a summary on a object of class "voxels" using bootstrap with 1000 replicates.
vox <- voxels(pc_tree, edge_length = c(0.5, 0.5, 0.5))
summary_voxels(vox, bootstrap = TRUE, R = 1000)

#>    Edge.X Edge.Y Edge.Z N_voxels Volume Surface Density_mean Density_sd
#> 1:    0.5    0.5    0.5      679 84.875   31.75     446.8218   326.8393
#>           H     Hmax Equitavility Negentropy H_boot_mean  H_boot_sd
#> 1: 6.221814 6.520621     0.954175  0.2988072    6.222789 0.01509371
#>    Equitavility_boot Negentropy_boot
#> 1:         0.9543246       0.2978322

#Apply a summary on a product from 'voxels' using bootstrap with 1000 replicates.
vox <- voxels(pc_tree, edge_length = c(0.5, 0.5, 0.5), obj.voxels = FALSE)
summary_voxels(vox, edge_length = c(0.5, 0.5, 0.5), bootstrap = TRUE, R = 1000)

#>    Edge.X Edge.Y Edge.Z N_voxels Volume Surface Density_mean Density_sd
#> 1:    0.5    0.5    0.5      679 84.875   31.75     446.8218   326.8393
#>           H     Hmax Equitavility Negentropy H_boot_mean  H_boot_sd
#> 1: 6.221814 6.520621     0.954175  0.2988072    6.222228 0.01533672
#>    Equitavility_boot Negentropy_boot
#> 1:         0.9542386        0.298393