Voxels Counting

Creates cube like voxels of different size on a point cloud using the voxels function, and then return a summary_voxels of their features.

voxels_counting(
  cloud,
  edge_sizes = NULL,
  min_size,
  length_out = 10,
  bootstrap = FALSE,
  R = NULL,
  progress = TRUE,
  parallel = FALSE,
  threads = NULL
)

Arguments

cloud
A data.table with xyz coordinates of the point clouds in the first three columns.

edge_sizes
A positive numeric vector describing the edge length of the different cubes to perform. If NULL, it use edge sizes by default based on the largest range of XYZ and min_size.

min_size
A positive numeric vector of length 1 describing the minimum cube edge length to perform. This is required if edge_sizes = NULL.

length_out
A positive interger of length 1 indicating the number of different edge lengths to use. This is required if edge_sizes = NULL.

bootstrap
Logical. If TRUE, it computes a bootstrap on the H index calculations. FALSE as default.

R
A positive integer of length 1 indicating the number of bootstrap replicates. This need to be used if bootstrap = TRUE.

progress
Logical, if TRUE displays a graphical progress bar. TRUE as default.

parallel
Logical, if TRUE it uses a parallel processing for the voxelization. FALSE as default.

threads
An integer >= 0 describing the number of threads to use. This need to be used if parallel = TRUE.

Value

A data.table with the summary of the voxels created with their features.

See also

Author

J. Antonio Guzmán Q.

Examples

data(pc_tree)

#Applying voxels counting.
voxels_counting(pc_tree, min_size = 2)

#> [1] "Creating voxels"
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#>       Edge.X   Edge.Y   Edge.Z N_voxels   Volume  Surface Density_mean
#>  1: 2.000000 2.000000 2.000000       26 208.0000 52.00000     729.3077
#>  2: 2.273717 2.273717 2.273717       23 270.3565 46.52809     637.8868
#>  3: 2.584894 2.584894 2.584894       18 310.8856 53.45340     630.6469
#>  4: 2.938658 2.938658 2.938658       14 355.2836 69.08568     627.3617
#>  5: 3.340838 3.340838 3.340838        8 298.3020 44.64479     849.4609
#>  6: 3.798059 3.798059 3.798059        8 438.3037 57.70101     657.2501
#>  7: 4.317855 4.317855 4.317855        8 644.0124 74.57549     508.5317
#>  8: 4.908790 4.908790 4.908790        7 827.9828 96.38486     449.6735
#>  9: 5.580598 5.580598 5.580598        4 695.1880 93.42923     608.8672
#> 10: 6.344350 6.344350 6.344350        1 255.3650 40.25077    1884.3862
#>     Density_sd         H     Hmax Equitavility Negentropy
#>  1:   666.6192 2.8071337 3.258097    0.8615870  0.4509629
#>  2:   731.5321 2.5806326 3.135494    0.8230385  0.5548616
#>  3:   861.5751 2.2028562 2.890372    0.7621360  0.6875155
#>  4:   778.8986 1.9014672 2.639057    0.7205100  0.7375901
#>  5:   416.4392 1.9635014 2.079442    0.9442446  0.1159402
#>  6:   408.2369 1.9136861 2.079442    0.9202885  0.1657555
#>  7:   598.9605 1.5854415 2.079442    0.7624362  0.4940001
#>  8:   823.1561 0.9318226 1.945910    0.4788621  1.0140875
#>  9:  1143.8455 0.2341148 1.386294    0.1688781  1.1521796
#> 10:         NA 0.0000000 0.000000          NaN  0.0000000

#Voxels counting using bootstrap on the H indexes with 1000 repetitions.
voxels_counting(pc_tree, min_size = 2, bootstrap = TRUE, R = 1000)

#> [1] "Creating voxels"
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#>       Edge.X   Edge.Y   Edge.Z N_voxels   Volume  Surface Density_mean
#>  1: 2.000000 2.000000 2.000000       26 208.0000 52.00000     729.3077
#>  2: 2.273717 2.273717 2.273717       23 270.3565 46.52809     637.8868
#>  3: 2.584894 2.584894 2.584894       18 310.8856 53.45340     630.6469
#>  4: 2.938658 2.938658 2.938658       14 355.2836 69.08568     627.3617
#>  5: 3.340838 3.340838 3.340838        8 298.3020 44.64479     849.4609
#>  6: 3.798059 3.798059 3.798059        8 438.3037 57.70101     657.2501
#>  7: 4.317855 4.317855 4.317855        8 644.0124 74.57549     508.5317
#>  8: 4.908790 4.908790 4.908790        7 827.9828 96.38486     449.6735
#>  9: 5.580598 5.580598 5.580598        4 695.1880 93.42923     608.8672
#> 10: 6.344350 6.344350 6.344350        1 255.3650 40.25077    1884.3862
#>     Density_sd         H     Hmax Equitavility Negentropy H_boot_mean
#>  1:   666.6192 2.8071337 3.258097    0.8615870  0.4509629   2.8165245
#>  2:   731.5321 2.5806326 3.135494    0.8230385  0.5548616   2.6013670
#>  3:   861.5751 2.2028562 2.890372    0.7621360  0.6875155   2.2400107
#>  4:   778.8986 1.9014672 2.639057    0.7205100  0.7375901   1.9018532
#>  5:   416.4392 1.9635014 2.079442    0.9442446  0.1159402   1.9726031
#>  6:   408.2369 1.9136861 2.079442    0.9202885  0.1657555   1.9332019
#>  7:   598.9605 1.5854415 2.079442    0.7624362  0.4940001   1.6757899
#>  8:   823.1561 0.9318226 1.945910    0.4788621  1.0140875   1.1684127
#>  9:  1143.8455 0.2341148 1.386294    0.1688781  1.1521796   0.7513146
#> 10:         NA 0.0000000 0.000000          NaN  0.0000000   0.0000000
#>      H_boot_sd Equitavility_boot Negentropy_boot
#>  1: 0.11564166         0.8644693       0.4415721
#>  2: 0.10959497         0.8296513       0.5341272
#>  3: 0.13150712         0.7749905       0.6503610
#>  4: 0.23052684         0.7206563       0.7372041
#>  5: 0.04108726         0.9486216       0.1068384
#>  6: 0.05918866         0.9296736       0.1462396
#>  7: 0.17092959         0.8058846       0.4036516
#>  8: 0.32414184         0.6004453       0.7774975
#>  9: 0.49430205         0.5419589       0.6349797
#> 10: 0.00000000               NaN       0.0000000