Voxels Counting

Creates voxels of different size on a point cloud using the voxels function, and then return a summary_voxels of their features.

voxels_counting(
  cloud,
  voxel.sizes = NULL,
  min.size,
  length.out = 10,
  bootstrap = FALSE,
  R = NULL,
  progress = TRUE,
  parallel = FALSE,
  cores = NULL
)

Arguments

cloud
A data.table with xyz coordinates of the point clouds in the first three columns.

voxel.sizes
A positive numeric vector describing the different voxel edge lengths to perform. If NULL, it use voxel sizes by default based on the largest range of XYZ and min.size.

min.size
A positive numeric vector of length 1 describing the minimum voxel edge length to perform. This is required if voxel.sizes = NULL.

length.out
A positive interger of length 1 indicating the number of different voxel sizes to use. This is required if voxel.sizes = NULL.

bootstrap
Logical. If TRUE, it computes a bootstrap on the H index calculations. FALSE as default.

R
A positive integer of length 1 indicating the number of bootstrap replicates. This need to be used if bootstrap = TRUE.

progress
Logical, if TRUE displays a graphical progress bar. TRUE as default.

parallel
Logical, if TRUE it uses a parallel processing for the voxelization. FALSE as default.

cores
An integer >= 0 describing the number of cores to use. This need to be used if parallel = TRUE.

Value

A data.table with the summary of the voxels created with their features.

See also

Examples


data(pc_tree)

#Applying voxels counting.
voxels_counting(pc_tree, min.size = 0.05)

#> [1] "Creating voxels"
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#>     Voxel.size N_voxels     volume  Surface Density_mean Density_sd         H
#>  1:  0.0500000    72859   9.107375 25.25250    8328.1956  1591.7595 11.181794
#>  2:  0.0856426    38369  24.101833 26.25073    3146.9806  1698.9219 10.419240
#>  3:  0.1466931    12990  41.005141 27.39352    1849.7193  1356.6095  9.206435
#>  4:  0.2512636     3735  59.248769 28.97823    1280.1616  1019.8631  7.902519
#>  5:  0.4303774     1012  80.673112 30.93252     940.1893   742.7959  6.586281
#>  6:  0.7371728      262 104.956457 34.23569     722.6616   539.7403  5.255786
#>  7:  1.2626679       79 159.035666 41.45259     476.9245   432.0273  3.956065
#>  8:  2.1627634       23 232.677747 42.09791     325.9787   316.0107  2.701463
#>  9:  3.7044937        8 406.702247 54.89309     186.4952   103.6981  1.941538
#> 10:  6.3452497        1 255.473675 40.26219     296.8916         NA  0.000000
#>          Hmax Equitavility Negentropy
#>  1: 11.196281    0.9987060 0.01448758
#>  2: 10.555005    0.9871373 0.13576547
#>  3:  9.471935    0.9719698 0.26550005
#>  4:  8.225503    0.9607338 0.32298420
#>  5:  6.919684    0.9518181 0.33340334
#>  6:  5.568345    0.9438686 0.31255878
#>  7:  4.369448    0.9053924 0.41338314
#>  8:  3.135494    0.8615749 0.43403118
#>  9:  2.079442    0.9336824 0.13790355
#> 10:  0.000000          NaN 0.00000000


#Voxels counting using boostrap on the H indixes with 1000 repetitions.
voxels_counting(pc_tree, min.size = 0.05, bootstrap = TRUE, R = 1000)

#> [1] "Creating voxels"
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#>     Voxel.size N_voxels     volume  Surface Density_mean Density_sd         H
#>  1:  0.0500000    72859   9.107375 25.25250    8328.1956  1591.7595 11.181794
#>  2:  0.0856426    38369  24.101833 26.25073    3146.9806  1698.9219 10.419240
#>  3:  0.1466931    12990  41.005141 27.39352    1849.7193  1356.6095  9.206435
#>  4:  0.2512636     3735  59.248769 28.97823    1280.1616  1019.8631  7.902519
#>  5:  0.4303774     1012  80.673112 30.93252     940.1893   742.7959  6.586281
#>  6:  0.7371728      262 104.956457 34.23569     722.6616   539.7403  5.255786
#>  7:  1.2626679       79 159.035666 41.45259     476.9245   432.0273  3.956065
#>  8:  2.1627634       23 232.677747 42.09791     325.9787   316.0107  2.701463
#>  9:  3.7044937        8 406.702247 54.89309     186.4952   103.6981  1.941538
#> 10:  6.3452497        1 255.473675 40.26219     296.8916         NA  0.000000
#>          Hmax Equitavility Negentropy H_boot_mean    H_boot_sd
#>  1: 11.196281    0.9987060 0.01448758   11.181798 0.0002444957
#>  2: 10.555005    0.9871373 0.13576547   10.419243 0.0006875815
#>  3:  9.471935    0.9719698 0.26550005    9.206300 0.0025179284
#>  4:  8.225503    0.9607338 0.32298420    7.902664 0.0060826024
#>  5:  6.919684    0.9518181 0.33340334    6.586349 0.0128106250
#>  6:  5.568345    0.9438686 0.31255878    5.257734 0.0256478765
#>  7:  4.369448    0.9053924 0.41338314    3.959126 0.0501174616
#>  8:  3.135494    0.8615749 0.43403118    2.712053 0.0963596360
#>  9:  2.079442    0.9336824 0.13790355    1.952823 0.0496017772
#> 10:  0.000000          NaN 0.00000000    0.000000 0.0000000000
#>     Equitavility_boot Negentropy_boot
#>  1:         0.9987064      0.01448301
#>  2:         0.9871377      0.13576177
#>  3:         0.9719556      0.26563509
#>  4:         0.9607514      0.32283955
#>  5:         0.9518280      0.33333487
#>  6:         0.9442185      0.31061075
#>  7:         0.9060930      0.41032157
#>  8:         0.8649523      0.42344129
#>  9:         0.9391092      0.12661886
#> 10:               NaN      0.00000000


if (FALSE) {
#Voxels counting using bootstrap on the H indices and using parallel processing.
voxels_counting(pc_tree, min.size = 0.05, bootstrap = TRUE, R = 1000, parallel = TRUE, cores = 4)
}